Variance-Covariance Matrix for INLAvaan Models

Extract the posterior variance-covariance matrix of model parameters from a fitted INLAvaan model.

Usage

# S4 method for class 'INLAvaan'
vcov(object, type = c("lavaan", "theta"), ...)

Arguments

object

An object of class INLAvaan.

type

Character. "lavaan" (default) returns the posterior covariance matrix of the model parameters computed from posterior samples, with one row per entry of coef() (matching lavaan output). "theta" returns the Laplace approximation covariance in the internal parameterisation, with one row per distinct free parameter.

...

Currently unused.

Value

A square numeric matrix.

See also

Examples

# \donttest{
HS.model <- "
  visual  =~ x1 + x2 + x3
  textual =~ x4 + x5 + x6
  speed   =~ x7 + x8 + x9
"
utils::data("HolzingerSwineford1939", package = "lavaan")
fit <- acfa(HS.model, HolzingerSwineford1939, std.lv = TRUE, nsamp = 100,
            test = "none", verbose = FALSE)

# Default: posterior covariance of lavaan parameters
vcov(fit)
#>                    visual=~x1    visual=~x2    visual=~x3   textual=~x4
#> visual=~x1       9.113152e-03 -3.735615e-04 -1.392575e-03 -3.217921e-05
#> visual=~x2      -3.735615e-04  5.839548e-03  2.094213e-03 -8.600609e-04
#> visual=~x3      -1.392575e-03  2.094213e-03  6.413485e-03  4.652204e-04
#> textual=~x4     -3.217921e-05 -8.600609e-04  4.652204e-04  3.028697e-03
#> textual=~x5     -4.377863e-04 -5.012386e-04 -2.650123e-04  1.259857e-03
#> textual=~x6     -8.174463e-06 -7.235904e-04 -2.393892e-04  5.447130e-04
#> speed=~x7       -2.323173e-04  5.257470e-05  4.566766e-05  7.844008e-04
#> speed=~x8        9.144403e-04 -3.587648e-04 -6.297251e-04  7.191930e-04
#> speed=~x9        1.141409e-04  5.425906e-05  3.893687e-04 -3.075646e-04
#> x1~~x1          -1.037410e-02  2.356735e-03  3.628763e-03  2.054754e-04
#> x2~~x2           9.229303e-04 -1.183797e-03  6.953990e-04 -1.331077e-04
#> x3~~x3           2.632122e-03 -2.195653e-03 -2.967415e-03 -3.139396e-04
#> x4~~x4           3.651553e-04 -1.850129e-04 -5.261054e-04 -2.636080e-04
#> x5~~x5           2.981280e-05  2.189195e-04  8.900387e-04  2.668453e-04
#> x6~~x6          -4.722871e-04 -4.696807e-04  5.261139e-04  7.279208e-04
#> x7~~x7           5.726529e-04 -9.071361e-04 -3.891645e-04  4.665475e-04
#> x8~~x8          -2.283746e-04  5.054085e-04  9.494756e-04 -3.363705e-04
#> x9~~x9          -3.225818e-04  4.412639e-04  3.536285e-04  5.538392e-04
#> visual~~textual  1.176362e-03 -3.981830e-04 -8.882267e-04 -5.854405e-05
#> visual~~speed   -4.771680e-04  1.421863e-04  1.307978e-03 -1.831682e-04
#> textual~~speed  -1.429647e-04  1.728171e-04  1.044372e-03 -5.654342e-04
#>                   textual=~x5   textual=~x6     speed=~x7     speed=~x8
#> visual=~x1      -4.377863e-04 -8.174463e-06 -2.323173e-04  9.144403e-04
#> visual=~x2      -5.012386e-04 -7.235904e-04  5.257470e-05 -3.587648e-04
#> visual=~x3      -2.650123e-04 -2.393892e-04  4.566766e-05 -6.297251e-04
#> textual=~x4      1.259857e-03  5.447130e-04  7.844008e-04  7.191930e-04
#> textual=~x5      3.399931e-03  1.055007e-03 -6.806095e-05  4.806175e-04
#> textual=~x6      1.055007e-03  2.655397e-03 -5.948145e-05  7.048243e-04
#> speed=~x7       -6.806095e-05 -5.948145e-05  6.849203e-03  1.911127e-03
#> speed=~x8        4.806175e-04  7.048243e-04  1.911127e-03  5.672322e-03
#> speed=~x9        4.631106e-04  2.044414e-04 -1.234135e-03 -1.628056e-03
#> x1~~x1           9.794978e-04  4.044820e-04  5.379952e-04 -8.827055e-04
#> x2~~x2           2.234824e-04 -4.667976e-05 -1.214578e-03 -4.950358e-05
#> x3~~x3           9.539817e-05  7.610931e-04  1.785155e-05  1.707706e-03
#> x4~~x4           6.150573e-04  9.095323e-04 -6.931050e-04 -5.871635e-05
#> x5~~x5          -7.946144e-04  3.702418e-04  7.978078e-04  4.260112e-04
#> x6~~x6           4.372777e-04 -3.961237e-04  1.443303e-05 -9.196512e-05
#> x7~~x7           8.231981e-04  7.810077e-04 -3.140653e-03 -1.724306e-03
#> x8~~x8          -3.917829e-04 -1.412250e-03 -2.225626e-03 -5.088103e-03
#> x9~~x9           4.308224e-05  9.060996e-05  2.969975e-03  2.857274e-03
#> visual~~textual  5.774241e-04  7.051516e-04 -9.685160e-05  1.892653e-04
#> visual~~speed    7.687581e-04  2.668733e-04 -1.440452e-03 -1.861889e-03
#> textual~~speed  -2.304354e-04 -1.406422e-04 -1.110058e-05 -4.761736e-04
#>                     speed=~x9        x1~~x1        x2~~x2        x3~~x3
#> visual=~x1       1.141409e-04 -0.0103740966  9.229303e-04  2.632122e-03
#> visual=~x2       5.425906e-05  0.0023567347 -1.183797e-03 -2.195653e-03
#> visual=~x3       3.893687e-04  0.0036287633  6.953990e-04 -2.967415e-03
#> textual=~x4     -3.075646e-04  0.0002054754 -1.331077e-04 -3.139396e-04
#> textual=~x5      4.631106e-04  0.0009794978  2.234824e-04  9.539817e-05
#> textual=~x6      2.044414e-04  0.0004044820 -4.667976e-05  7.610931e-04
#> speed=~x7       -1.234135e-03  0.0005379952 -1.214578e-03  1.785155e-05
#> speed=~x8       -1.628056e-03 -0.0008827055 -4.950358e-05  1.707706e-03
#> speed=~x9        5.647097e-03  0.0001435735  1.279712e-04 -9.607371e-04
#> x1~~x1           1.435735e-04  0.0204286382 -2.332912e-03 -5.243299e-03
#> x2~~x2           1.279712e-04 -0.0023329124  8.908702e-03  2.403958e-03
#> x3~~x3          -9.607371e-04 -0.0052432993  2.403958e-03  9.955018e-03
#> x4~~x4           5.096346e-05  0.0001538215  8.970430e-05  2.454088e-05
#> x5~~x5           1.898354e-04  0.0002405122  9.788292e-04  7.013041e-05
#> x6~~x6           2.961475e-04  0.0010619435  5.620827e-04  1.473995e-04
#> x7~~x7           2.548057e-03 -0.0011965295  4.908713e-04 -1.259637e-04
#> x8~~x8           3.422891e-03  0.0000914443 -5.710645e-04 -3.344855e-03
#> x9~~x9          -4.040091e-03  0.0010279039 -1.219643e-04 -2.664598e-04
#> visual~~textual -3.635833e-04 -0.0017376152  2.758418e-04  2.250578e-03
#> visual~~speed    2.461592e-03  0.0015946991 -2.307834e-04 -1.834728e-03
#> textual~~speed   1.194665e-03 -0.0001329301 -2.812745e-04  1.872353e-04
#>                        x4~~x4        x5~~x5        x6~~x6        x7~~x7
#> visual=~x1       3.651553e-04  2.981280e-05 -4.722871e-04  5.726529e-04
#> visual=~x2      -1.850129e-04  2.189195e-04 -4.696807e-04 -9.071361e-04
#> visual=~x3      -5.261054e-04  8.900387e-04  5.261139e-04 -3.891645e-04
#> textual=~x4     -2.636080e-04  2.668453e-04  7.279208e-04  4.665475e-04
#> textual=~x5      6.150573e-04 -7.946144e-04  4.372777e-04  8.231981e-04
#> textual=~x6      9.095323e-04  3.702418e-04 -3.961237e-04  7.810077e-04
#> speed=~x7       -6.931050e-04  7.978078e-04  1.443303e-05 -3.140653e-03
#> speed=~x8       -5.871635e-05  4.260112e-04 -9.196512e-05 -1.724306e-03
#> speed=~x9        5.096346e-05  1.898354e-04  2.961475e-04  2.548057e-03
#> x1~~x1           1.538215e-04  2.405122e-04  1.061943e-03 -1.196530e-03
#> x2~~x2           8.970430e-05  9.788292e-04  5.620827e-04  4.908713e-04
#> x3~~x3           2.454088e-05  7.013041e-05  1.473995e-04 -1.259637e-04
#> x4~~x4           2.086329e-03 -2.432560e-04 -4.540313e-04  2.965096e-05
#> x5~~x5          -2.432560e-04  3.305524e-03 -7.408240e-04 -3.584055e-04
#> x6~~x6          -4.540313e-04 -7.408240e-04  2.170568e-03  3.851836e-04
#> x7~~x7           2.965096e-05 -3.584055e-04  3.851836e-04  8.132583e-03
#> x8~~x8          -1.177508e-04 -4.505811e-04  4.718342e-04  2.772895e-03
#> x9~~x9           3.626607e-04  8.090018e-04 -7.598028e-04 -3.374710e-03
#> visual~~textual  5.030031e-04  6.749120e-05 -3.488794e-04  4.270476e-04
#> visual~~speed   -1.209177e-04  1.831806e-04 -1.329300e-06  3.810149e-03
#> textual~~speed  -2.921729e-04  3.579827e-04 -2.629875e-04  1.402891e-03
#>                        x8~~x8        x9~~x9 visual~~textual visual~~speed
#> visual=~x1      -0.0002283746 -3.225818e-04    1.176362e-03 -4.771680e-04
#> visual=~x2       0.0005054085  4.412639e-04   -3.981830e-04  1.421863e-04
#> visual=~x3       0.0009494756  3.536285e-04   -8.882267e-04  1.307978e-03
#> textual=~x4     -0.0003363705  5.538392e-04   -5.854405e-05 -1.831682e-04
#> textual=~x5     -0.0003917829  4.308224e-05    5.774241e-04  7.687581e-04
#> textual=~x6     -0.0014122501  9.060996e-05    7.051516e-04  2.668733e-04
#> speed=~x7       -0.0022256261  2.969975e-03   -9.685160e-05 -1.440452e-03
#> speed=~x8       -0.0050881027  2.857274e-03    1.892653e-04 -1.861889e-03
#> speed=~x9        0.0034228908 -4.040091e-03   -3.635833e-04  2.461592e-03
#> x1~~x1           0.0000914443  1.027904e-03   -1.737615e-03  1.594699e-03
#> x2~~x2          -0.0005710645 -1.219643e-04    2.758418e-04 -2.307834e-04
#> x3~~x3          -0.0033448549 -2.664598e-04    2.250578e-03 -1.834728e-03
#> x4~~x4          -0.0001177508  3.626607e-04    5.030031e-04 -1.209177e-04
#> x5~~x5          -0.0004505811  8.090018e-04    6.749120e-05  1.831806e-04
#> x6~~x6           0.0004718342 -7.598028e-04   -3.488794e-04 -1.329300e-06
#> x7~~x7           0.0027728951 -3.374710e-03    4.270476e-04  3.810149e-03
#> x8~~x8           0.0094124671 -4.912353e-03   -9.894985e-04  3.088387e-03
#> x9~~x9          -0.0049123528  8.133712e-03   -2.824734e-04 -2.487850e-03
#> visual~~textual -0.0009894985 -2.824734e-04    4.076011e-03 -5.465387e-05
#> visual~~speed    0.0030883868 -2.487850e-03   -5.465387e-05  8.102847e-03
#> textual~~speed   0.0009586712 -9.353375e-04    1.330444e-03  2.186593e-03
#>                 textual~~speed
#> visual=~x1       -1.429647e-04
#> visual=~x2        1.728171e-04
#> visual=~x3        1.044372e-03
#> textual=~x4      -5.654342e-04
#> textual=~x5      -2.304354e-04
#> textual=~x6      -1.406422e-04
#> speed=~x7        -1.110058e-05
#> speed=~x8        -4.761736e-04
#> speed=~x9         1.194665e-03
#> x1~~x1           -1.329301e-04
#> x2~~x2           -2.812745e-04
#> x3~~x3            1.872353e-04
#> x4~~x4           -2.921729e-04
#> x5~~x5            3.579827e-04
#> x6~~x6           -2.629875e-04
#> x7~~x7            1.402891e-03
#> x8~~x8            9.586712e-04
#> x9~~x9           -9.353375e-04
#> visual~~textual   1.330444e-03
#> visual~~speed     2.186593e-03
#> textual~~speed    4.745312e-03

# Internal parameterisation (Laplace approximation)
vcov(fit, type = "theta")
#>                    visual=~x1    visual=~x2    visual=~x3   textual=~x4
#> visual=~x1       6.944819e-03 -9.228843e-04 -9.935751e-04  3.389857e-04
#> visual=~x2      -9.228843e-04  6.568863e-03  1.577215e-03  1.410972e-04
#> visual=~x3      -9.935751e-04  1.577215e-03  6.061736e-03  1.941726e-04
#> textual=~x4      3.389857e-04  1.410972e-04  1.941726e-04  3.223624e-03
#> textual=~x5      3.141650e-04  2.000129e-04  2.591496e-04  1.688801e-03
#> textual=~x6      2.546537e-04  1.692562e-04  2.206257e-04  1.376938e-03
#> speed=~x7        3.049172e-04  7.658901e-05  5.006757e-05  8.922238e-05
#> speed=~x8        3.435120e-04  1.427343e-04  4.258362e-05  9.357838e-05
#> speed=~x9        6.233622e-05  1.250035e-04  3.071841e-04  7.405772e-05
#> x1~~x1          -1.222508e-02  5.410147e-03  6.403818e-03 -1.279744e-04
#> x2~~x2           1.466413e-03 -2.090859e-03 -9.100595e-04  2.237214e-05
#> x3~~x3           3.072863e-03 -1.610649e-03 -3.944877e-03  3.916836e-05
#> x4~~x4          -2.114624e-04  1.400470e-04  1.444548e-04 -1.889064e-03
#> x5~~x5           9.341801e-05 -6.713443e-05 -6.283285e-05  5.955463e-04
#> x6~~x6           1.154535e-04 -7.284124e-05 -8.025778e-05  6.602594e-04
#> x7~~x7          -1.669503e-04  5.012766e-05  1.437528e-04 -1.903188e-05
#> x8~~x8          -3.329928e-04 -4.342221e-05  3.783558e-04 -8.334976e-06
#> x9~~x9           3.507098e-04 -1.901485e-05 -3.605765e-04  2.021852e-05
#> visual~~textual  9.461949e-04  2.082300e-05 -3.796066e-04  7.715873e-04
#> visual~~speed   -4.825363e-04  3.751536e-04  9.875436e-04  2.256550e-04
#> textual~~speed   4.314192e-04  2.654701e-04  4.026555e-04  4.093776e-04
#>                   textual=~x5   textual=~x6     speed=~x7     speed=~x8
#> visual=~x1       3.141650e-04  2.546537e-04  3.049172e-04  3.435120e-04
#> visual=~x2       2.000129e-04  1.692562e-04  7.658901e-05  1.427343e-04
#> visual=~x3       2.591496e-04  2.206257e-04  5.006757e-05  4.258362e-05
#> textual=~x4      1.688801e-03  1.376938e-03  8.922238e-05  9.357838e-05
#> textual=~x5      3.930062e-03  1.578490e-03  7.437476e-05  9.582845e-05
#> textual=~x6      1.578490e-03  2.898913e-03  7.457536e-05  9.654040e-05
#> speed=~x7        7.437476e-05  7.457536e-05  5.475475e-03  1.898565e-03
#> speed=~x8        9.582845e-05  9.654040e-05  1.898565e-03  5.612839e-03
#> speed=~x9        1.034695e-04  6.539474e-05 -8.867327e-04 -1.687832e-03
#> x1~~x1           6.481860e-05  7.383460e-05 -3.332406e-04 -3.390605e-04
#> x2~~x2          -1.259862e-05 -1.328243e-05  3.157452e-05 -8.647084e-06
#> x3~~x3          -2.269572e-05 -2.738332e-05  1.527783e-04  2.064197e-04
#> x4~~x4           6.459757e-04  6.832362e-04 -5.449319e-05  5.461796e-06
#> x5~~x5          -2.080084e-03  4.801831e-04  6.677323e-05  4.678771e-05
#> x6~~x6           5.039398e-04 -1.603122e-03 -7.335699e-06 -4.579406e-05
#> x7~~x7           1.757196e-05 -5.183366e-06 -3.368940e-03 -1.767978e-03
#> x8~~x8           1.574371e-05 -3.771666e-05 -3.429949e-03 -9.323437e-03
#> x9~~x9          -2.677522e-05  2.604058e-05  3.695614e-03  5.862929e-03
#> visual~~textual  5.976293e-04  7.336587e-04  1.891355e-04  2.281799e-04
#> visual~~speed    2.858364e-04  2.187081e-04 -1.926030e-03 -2.320677e-03
#> textual~~speed   4.661669e-04  4.047452e-04 -3.246937e-04 -8.927829e-04
#>                     speed=~x9        x1~~x1        x2~~x2        x3~~x3
#> visual=~x1       6.233622e-05 -0.0122250842  1.466413e-03  3.072863e-03
#> visual=~x2       1.250035e-04  0.0054101467 -2.090859e-03 -1.610649e-03
#> visual=~x3       3.071841e-04  0.0064038183 -9.100595e-04 -3.944877e-03
#> textual=~x4      7.405772e-05 -0.0001279744  2.237214e-05  3.916836e-05
#> textual=~x5      1.034695e-04  0.0000648186 -1.259862e-05 -2.269572e-05
#> textual=~x6      6.539474e-05  0.0000738346 -1.328243e-05 -2.738332e-05
#> speed=~x7       -8.867327e-04 -0.0003332406  3.157452e-05  1.527783e-04
#> speed=~x8       -1.687832e-03 -0.0003390605 -8.647084e-06  2.064197e-04
#> speed=~x9        5.894651e-03  0.0004510226 -1.161187e-05 -2.491848e-04
#> x1~~x1           4.510226e-04  0.0463660406 -4.766128e-03 -9.986706e-03
#> x2~~x2          -1.161187e-05 -0.0047661278  8.511238e-03  1.417195e-03
#> x3~~x3          -2.491848e-04 -0.0099867063  1.417195e-03  1.281872e-02
#> x4~~x4           2.680573e-05  0.0006882199 -1.230916e-04 -2.246406e-04
#> x5~~x5          -7.597103e-05 -0.0003116688  5.786411e-05  9.502759e-05
#> x6~~x6           4.038167e-05 -0.0003723394  6.453136e-05  1.260108e-04
#> x7~~x7           2.439345e-03  0.0005089865 -4.915125e-05 -2.357085e-04
#> x8~~x8           7.513275e-03  0.0010008137  2.603809e-05 -6.179704e-04
#> x9~~x9          -7.620125e-03 -0.0010703061  2.712778e-05  5.860744e-04
#> visual~~textual  1.179829e-04 -0.0009458378  3.021593e-04  1.338676e-03
#> visual~~speed    3.691166e-03  0.0036899830 -1.057782e-05 -7.935360e-04
#> textual~~speed   1.371421e-03  0.0001321966 -2.703898e-06 -8.890268e-05
#>                        x4~~x4        x5~~x5        x6~~x6        x7~~x7
#> visual=~x1      -2.114624e-04  9.341801e-05  1.154535e-04 -1.669503e-04
#> visual=~x2       1.400470e-04 -6.713443e-05 -7.284124e-05  5.012766e-05
#> visual=~x3       1.444548e-04 -6.283285e-05 -8.025778e-05  1.437528e-04
#> textual=~x4     -1.889064e-03  5.955463e-04  6.602594e-04 -1.903188e-05
#> textual=~x5      6.459757e-04 -2.080084e-03  5.039398e-04  1.757196e-05
#> textual=~x6      6.832362e-04  4.801831e-04 -1.603122e-03 -5.183366e-06
#> speed=~x7       -5.449319e-05  6.677323e-05 -7.335699e-06 -3.368940e-03
#> speed=~x8        5.461796e-06  4.678771e-05 -4.579406e-05 -1.767978e-03
#> speed=~x9        2.680573e-05 -7.597103e-05  4.038167e-05  2.439345e-03
#> x1~~x1           6.882199e-04 -3.116688e-04 -3.723394e-04  5.089865e-04
#> x2~~x2          -1.230916e-04  5.786411e-05  6.453136e-05 -4.915125e-05
#> x3~~x3          -2.246406e-04  9.502759e-05  1.260108e-04 -2.357085e-04
#> x4~~x4           1.672689e-02 -3.182478e-03 -3.506878e-03  8.429821e-05
#> x5~~x5          -3.182478e-03  1.691445e-02 -2.479085e-03 -1.040738e-04
#> x6~~x6          -3.506878e-03 -2.479085e-03  1.491448e-02  1.148294e-05
#> x7~~x7           8.429821e-05 -1.040738e-04  1.148294e-05  1.190422e-02
#> x8~~x8          -1.761092e-05 -1.415088e-04  1.382795e-04  5.305242e-03
#> x9~~x9          -6.349724e-05  1.774102e-04 -9.433640e-05 -5.724535e-03
#> visual~~textual -2.529357e-04  1.027251e-03 -3.238691e-04 -5.033374e-05
#> visual~~speed    7.175173e-05 -9.662949e-05  1.430018e-05  3.700656e-03
#> textual~~speed   1.292443e-04  8.197794e-05 -1.581051e-05  9.250200e-04
#>                        x8~~x8        x9~~x9 visual~~textual visual~~speed
#> visual=~x1      -3.329928e-04  3.507098e-04    9.461949e-04 -4.825363e-04
#> visual=~x2      -4.342221e-05 -1.901485e-05    2.082300e-05  3.751536e-04
#> visual=~x3       3.783558e-04 -3.605765e-04   -3.796066e-04  9.875436e-04
#> textual=~x4     -8.334976e-06  2.021852e-05    7.715873e-04  2.256550e-04
#> textual=~x5      1.574371e-05 -2.677522e-05    5.976293e-04  2.858364e-04
#> textual=~x6     -3.771666e-05  2.604058e-05    7.336587e-04  2.187081e-04
#> speed=~x7       -3.429949e-03  3.695614e-03    1.891355e-04 -1.926030e-03
#> speed=~x8       -9.323437e-03  5.862929e-03    2.281799e-04 -2.320677e-03
#> speed=~x9        7.513275e-03 -7.620125e-03    1.179829e-04  3.691166e-03
#> x1~~x1           1.000814e-03 -1.070306e-03   -9.458378e-04  3.689983e-03
#> x2~~x2           2.603809e-05  2.712778e-05    3.021593e-04 -1.057782e-05
#> x3~~x3          -6.179704e-04  5.860744e-04    1.338676e-03 -7.935360e-04
#> x4~~x4          -1.761092e-05 -6.349724e-05   -2.529357e-04  7.175173e-05
#> x5~~x5          -1.415088e-04  1.774102e-04    1.027251e-03 -9.662949e-05
#> x6~~x6           1.382795e-04 -9.433640e-05   -3.238691e-04  1.430018e-05
#> x7~~x7           5.305242e-03 -5.724535e-03   -5.033374e-05  3.700656e-03
#> x8~~x8           3.472513e-02 -1.762404e-02   -1.304225e-04  8.608315e-03
#> x9~~x9          -1.762404e-02  2.454052e-02    1.179607e-04 -7.497054e-03
#> visual~~textual -1.304225e-04  1.179607e-04    6.423496e-03  7.193127e-04
#> visual~~speed    8.608315e-03 -7.497054e-03    7.193127e-04  1.177453e-02
#> textual~~speed   3.652556e-03 -2.529804e-03    1.493595e-03  3.066816e-03
#>                 textual~~speed
#> visual=~x1        4.314192e-04
#> visual=~x2        2.654701e-04
#> visual=~x3        4.026555e-04
#> textual=~x4       4.093776e-04
#> textual=~x5       4.661669e-04
#> textual=~x6       4.047452e-04
#> speed=~x7        -3.246937e-04
#> speed=~x8        -8.927829e-04
#> speed=~x9         1.371421e-03
#> x1~~x1            1.321966e-04
#> x2~~x2           -2.703898e-06
#> x3~~x3           -8.890268e-05
#> x4~~x4            1.292443e-04
#> x5~~x5            8.197794e-05
#> x6~~x6           -1.581051e-05
#> x7~~x7            9.250200e-04
#> x8~~x8            3.652556e-03
#> x9~~x9           -2.529804e-03
#> visual~~textual   1.493595e-03
#> visual~~speed     3.066816e-03
#> textual~~speed    5.945061e-03
# }