Compare Bayesian Models Fitted with INLAvaan

Compare Bayesian Models Fitted with INLAvaan

Usage

compare(x, y, ...)

# S4 method for class 'INLAvaan'
compare(x, y, ...)

Arguments

x, y, ...

An object of class INLAvaan or inlavaan_internal.

Value

A data frame of class compare.inlavaan_internal containing model fit statistics.

Details

The function computes the log Bayes Factor (logBF) relative to the best fitting model (the one with the highest Marginal Log-Likelihood).

The output table sorts models by descending Marginal Log-Likelihood.

Marg.Loglik: The approximated marginal log-likelihood.

DIC: Deviance Information Criterion (if available).

pD: Effective number of parameters (if available).

logBF: The natural logarithm of the Bayes Factor relative to the best model.

References

https://lavaan.ugent.be/tutorial/groups.html

Examples

# Model comparison on multigroup analysis (measurement invariance)
HS.model <- "
  visual  =~ x1 + x2 + x3
  textual =~ x4 + x5 + x6
  speed   =~ x7 + x8 + x9
"
utils::data("HolzingerSwineford1939", package = "lavaan")

# Configural invariance
fit1 <- acfa(HS.model, data = HolzingerSwineford1939, group = "school")
#> ℹ Using MVN log-likelihood.
#> ℹ Finding posterior mode.
#> ✔ Finding posterior mode. [807ms]
#> 
#> ℹ Computing the Hessian.
#> ✔ Computing the Hessian. [2.4s]
#> 
#> ℹ Using skew normal approximation.
#> ℹ Fitting skew normal to marginals.
#> ✔ Fitting skew normal to marginals. [3.2s]
#> 
#> ℹ Sampling posterior covariances.
#> ✔ Sampling posterior covariances. [356ms]
#> 
#> ⠙ Computing ppp and DIC.
#> ⠹ Computing ppp and DIC.
#> ✔ Computing ppp and DIC. [5.1s]
#> 

# Weak invariance
fit2 <- acfa(
  HS.model,
  data = HolzingerSwineford1939,
  group = "school",
  group.equal = "loadings"
)
#> ℹ Using MVN log-likelihood.
#> ℹ Finding posterior mode.
#> ✔ Finding posterior mode. [853ms]
#> 
#> ℹ Computing the Hessian.
#> ✔ Computing the Hessian. [2s]
#> 
#> ℹ Using skew normal approximation.
#> ℹ Fitting skew normal to marginals.
#> ✔ Fitting skew normal to marginals. [2.7s]
#> 
#> ℹ Sampling posterior covariances.
#> ✔ Sampling posterior covariances. [336ms]
#> 
#> ⠙ Computing ppp and DIC.
#> ⠹ Computing ppp and DIC.
#> ⠸ Computing ppp and DIC.
#> ✔ Computing ppp and DIC. [4.5s]
#> 

# Strong invariance
fit3 <- acfa(
  HS.model,
  data = HolzingerSwineford1939,
  group = "school",
  group.equal = c("intercepts", "loadings")
)
#> ℹ Using MVN log-likelihood.
#> ℹ Finding posterior mode.
#> ✔ Finding posterior mode. [590ms]
#> 
#> ℹ Computing the Hessian.
#> ✔ Computing the Hessian. [1.7s]
#> 
#> ℹ Using skew normal approximation.
#> ℹ Fitting skew normal to marginals.
#> ✔ Fitting skew normal to marginals. [2.3s]
#> 
#> ℹ Sampling posterior covariances.
#> ✔ Sampling posterior covariances. [338ms]
#> 
#> ⠙ Computing ppp and DIC.
#> ⠹ Computing ppp and DIC.
#> ⠸ Computing ppp and DIC.
#> ✔ Computing ppp and DIC. [4.5s]
#> 

# Compare models
compare(fit1, fit2, fit3)
#> Error in compare.inlavaan_internal(x = fit1, y = fit2, fit3): could not find function "compare.inlavaan_internal"