Variance-Covariance Matrix for INLAvaan Models

Extract the posterior variance-covariance matrix of model parameters from a fitted INLAvaan model.

Usage

# S4 method for class 'INLAvaan'
vcov(object, type = c("lavaan", "theta"), ...)

Arguments

object

An object of class INLAvaan.

type

Character. "lavaan" (default) returns the posterior covariance matrix of the model parameters computed from posterior samples, with one row per entry of coef() (matching lavaan output). "theta" returns the Laplace approximation covariance in the internal parameterisation, with one row per distinct free parameter.

...

Currently unused.

Value

A square numeric matrix.

See also

Examples

# \donttest{
HS.model <- "
  visual  =~ x1 + x2 + x3
  textual =~ x4 + x5 + x6
  speed   =~ x7 + x8 + x9
"
utils::data("HolzingerSwineford1939", package = "lavaan")
fit <- acfa(HS.model, HolzingerSwineford1939, std.lv = TRUE, nsamp = 100,
            test = "none", verbose = FALSE)

# Default: posterior covariance of lavaan parameters
vcov(fit)
#>                    visual=~x1    visual=~x2    visual=~x3   textual=~x4
#> visual=~x1       5.413181e-03  4.474265e-04 -6.693032e-04 -5.015573e-05
#> visual=~x2       4.474265e-04  6.037591e-03  1.343305e-03  2.329609e-04
#> visual=~x3      -6.693032e-04  1.343305e-03  6.050402e-03  1.524591e-04
#> textual=~x4     -5.015573e-05  2.329609e-04  1.524591e-04  3.114140e-03
#> textual=~x5     -4.287202e-04 -3.720655e-04  1.749072e-04  8.311886e-04
#> textual=~x6     -3.697571e-04 -4.496549e-04 -9.442889e-05  4.851978e-04
#> speed=~x7        2.297338e-04 -3.384641e-04 -5.625356e-04 -4.599568e-04
#> speed=~x8        6.574913e-04  7.682927e-04  2.126156e-04 -2.398714e-04
#> speed=~x9        3.861679e-04  1.725802e-04  2.772709e-04  6.346486e-05
#> x1~~x1          -4.627131e-03  2.696443e-03  4.002414e-03  1.020177e-03
#> x2~~x2           1.509745e-03 -3.870671e-04 -1.344123e-03  5.698338e-06
#> x3~~x3           1.521946e-03 -7.924580e-04 -3.215621e-03  1.994392e-04
#> x4~~x4          -1.482432e-04  6.509829e-05  3.506266e-05 -8.502947e-04
#> x5~~x5           4.491172e-04 -9.010202e-04  1.490473e-04  1.795175e-04
#> x6~~x6           1.141003e-04 -1.805599e-04 -3.394637e-04  4.912329e-04
#> x7~~x7           2.954097e-04  9.295761e-04 -1.995881e-04  4.912906e-04
#> x8~~x8          -4.299219e-04 -4.430443e-04  8.277957e-04  5.336865e-04
#> x9~~x9          -3.801630e-05 -6.480917e-04 -5.526397e-04 -1.967915e-04
#> visual~~textual  9.684291e-04  5.896701e-05 -1.453104e-03  6.169455e-04
#> visual~~speed    2.180709e-04  9.474808e-05  1.143912e-03  6.694756e-04
#> textual~~speed   7.433904e-04 -8.599445e-04 -6.076821e-04  3.298792e-04
#>                   textual=~x5   textual=~x6     speed=~x7     speed=~x8
#> visual=~x1      -4.287202e-04 -3.697571e-04  2.297338e-04  0.0006574913
#> visual=~x2      -3.720655e-04 -4.496549e-04 -3.384641e-04  0.0007682927
#> visual=~x3       1.749072e-04 -9.442889e-05 -5.625356e-04  0.0002126156
#> textual=~x4      8.311886e-04  4.851978e-04 -4.599568e-04 -0.0002398714
#> textual=~x5      3.146609e-03  7.680857e-04 -3.368174e-04  0.0002601112
#> textual=~x6      7.680857e-04  2.498056e-03  2.130508e-04 -0.0003163204
#> speed=~x7       -3.368174e-04  2.130508e-04  4.879471e-03  0.0018419361
#> speed=~x8        2.601112e-04 -3.163204e-04  1.841936e-03  0.0060778418
#> speed=~x9       -9.381968e-05 -1.169154e-04 -6.380521e-04 -0.0012260774
#> x1~~x1           4.499296e-04  1.059653e-04 -1.673950e-03 -0.0013031756
#> x2~~x2           8.819019e-05 -1.762153e-04 -7.476152e-04 -0.0010637581
#> x3~~x3          -2.324088e-05  4.020758e-04 -1.754290e-04 -0.0003726484
#> x4~~x4          -4.028424e-04  1.895027e-04 -3.153563e-05 -0.0002633289
#> x5~~x5          -9.914897e-04  1.343896e-04 -6.863038e-05 -0.0011099706
#> x6~~x6           6.238540e-04 -2.021647e-04 -6.720296e-05  0.0002212149
#> x7~~x7          -4.234199e-04  4.042934e-05 -2.434874e-03 -0.0011185613
#> x8~~x8          -1.486898e-04  9.452025e-05 -1.155630e-03 -0.0047938208
#> x9~~x9           4.811932e-05  1.115673e-04  1.745520e-03  0.0032131992
#> visual~~textual  8.496692e-05  8.268373e-04  1.470228e-04 -0.0006192190
#> visual~~speed    6.628415e-04  2.070453e-04 -1.498161e-03 -0.0020100962
#> textual~~speed   1.165877e-03  5.850574e-04 -1.590911e-04 -0.0010561780
#>                     speed=~x9        x1~~x1        x2~~x2        x3~~x3
#> visual=~x1       3.861679e-04 -4.627131e-03  1.509745e-03  1.521946e-03
#> visual=~x2       1.725802e-04  2.696443e-03 -3.870671e-04 -7.924580e-04
#> visual=~x3       2.772709e-04  4.002414e-03 -1.344123e-03 -3.215621e-03
#> textual=~x4      6.346486e-05  1.020177e-03  5.698338e-06  1.994392e-04
#> textual=~x5     -9.381968e-05  4.499296e-04  8.819019e-05 -2.324088e-05
#> textual=~x6     -1.169154e-04  1.059653e-04 -1.762153e-04  4.020758e-04
#> speed=~x7       -6.380521e-04 -1.673950e-03 -7.476152e-04 -1.754290e-04
#> speed=~x8       -1.226077e-03 -1.303176e-03 -1.063758e-03 -3.726484e-04
#> speed=~x9        4.631740e-03  2.205036e-04  1.431289e-03  2.724994e-04
#> x1~~x1           2.205036e-04  1.261891e-02 -2.855042e-03 -3.932248e-03
#> x2~~x2           1.431289e-03 -2.855042e-03  1.185299e-02  6.532886e-04
#> x3~~x3           2.724994e-04 -3.932248e-03  6.532886e-04  8.239372e-03
#> x4~~x4           8.793289e-05  9.530773e-05  6.408439e-04 -2.796439e-04
#> x5~~x5          -5.348283e-04  1.694920e-04  7.652578e-04 -1.284710e-04
#> x6~~x6          -6.774589e-05 -1.562295e-04  3.522314e-04 -1.397624e-04
#> x7~~x7           1.572025e-03  1.741113e-04 -8.470831e-04  5.267880e-04
#> x8~~x8           2.883448e-03  9.033929e-04  1.043909e-03  1.384895e-04
#> x9~~x9          -3.537998e-03 -1.591788e-03 -4.422017e-04 -7.481051e-04
#> visual~~textual -1.604861e-04 -1.212355e-03 -2.244052e-04  1.843143e-03
#> visual~~speed    1.897531e-03  1.714909e-03  2.119762e-03 -1.966221e-04
#> textual~~speed   1.456440e-03 -4.451294e-04  1.895450e-04  1.143457e-03
#>                        x4~~x4        x5~~x5        x6~~x6        x7~~x7
#> visual=~x1      -1.482432e-04  4.491172e-04  1.141003e-04  2.954097e-04
#> visual=~x2       6.509829e-05 -9.010202e-04 -1.805599e-04  9.295761e-04
#> visual=~x3       3.506266e-05  1.490473e-04 -3.394637e-04 -1.995881e-04
#> textual=~x4     -8.502947e-04  1.795175e-04  4.912329e-04  4.912906e-04
#> textual=~x5     -4.028424e-04 -9.914897e-04  6.238540e-04 -4.234199e-04
#> textual=~x6      1.895027e-04  1.343896e-04 -2.021647e-04  4.042934e-05
#> speed=~x7       -3.153563e-05 -6.863038e-05 -6.720296e-05 -2.434874e-03
#> speed=~x8       -2.633289e-04 -1.109971e-03  2.212149e-04 -1.118561e-03
#> speed=~x9        8.793289e-05 -5.348283e-04 -6.774589e-05  1.572025e-03
#> x1~~x1           9.530773e-05  1.694920e-04 -1.562295e-04  1.741113e-04
#> x2~~x2           6.408439e-04  7.652578e-04  3.522314e-04 -8.470831e-04
#> x3~~x3          -2.796439e-04 -1.284710e-04 -1.397624e-04  5.267880e-04
#> x4~~x4           1.764372e-03 -2.066581e-04 -2.578032e-04  2.304409e-04
#> x5~~x5          -2.066581e-04  3.178506e-03 -2.766564e-04 -4.911884e-04
#> x6~~x6          -2.578032e-04 -2.766564e-04  1.750929e-03  2.234305e-04
#> x7~~x7           2.304409e-04 -4.911884e-04  2.234305e-04  7.021683e-03
#> x8~~x8          -2.115732e-04  9.375363e-04 -3.679690e-04  1.163424e-03
#> x9~~x9          -7.472324e-05  9.877891e-05  5.738860e-05 -1.615648e-03
#> visual~~textual -1.051931e-04  1.555510e-04  2.281494e-04  6.243665e-04
#> visual~~speed    1.547870e-04 -3.909697e-04 -8.792440e-05  2.932713e-04
#> textual~~speed   1.042183e-04  2.684125e-04  5.365461e-04  2.627113e-04
#>                        x8~~x8        x9~~x9 visual~~textual visual~~speed
#> visual=~x1      -4.299219e-04 -3.801630e-05    9.684291e-04  2.180709e-04
#> visual=~x2      -4.430443e-04 -6.480917e-04    5.896701e-05  9.474808e-05
#> visual=~x3       8.277957e-04 -5.526397e-04   -1.453104e-03  1.143912e-03
#> textual=~x4      5.336865e-04 -1.967915e-04    6.169455e-04  6.694756e-04
#> textual=~x5     -1.486898e-04  4.811932e-05    8.496692e-05  6.628415e-04
#> textual=~x6      9.452025e-05  1.115673e-04    8.268373e-04  2.070453e-04
#> speed=~x7       -1.155630e-03  1.745520e-03    1.470228e-04 -1.498161e-03
#> speed=~x8       -4.793821e-03  3.213199e-03   -6.192190e-04 -2.010096e-03
#> speed=~x9        2.883448e-03 -3.537998e-03   -1.604861e-04  1.897531e-03
#> x1~~x1           9.033929e-04 -1.591788e-03   -1.212355e-03  1.714909e-03
#> x2~~x2           1.043909e-03 -4.422017e-04   -2.244052e-04  2.119762e-03
#> x3~~x3           1.384895e-04 -7.481051e-04    1.843143e-03 -1.966221e-04
#> x4~~x4          -2.115732e-04 -7.472324e-05   -1.051931e-04  1.547870e-04
#> x5~~x5           9.375363e-04  9.877891e-05    1.555510e-04 -3.909697e-04
#> x6~~x6          -3.679690e-04  5.738860e-05    2.281494e-04 -8.792440e-05
#> x7~~x7           1.163424e-03 -1.615648e-03    6.243665e-04  2.932713e-04
#> x8~~x8           8.534511e-03 -4.527370e-03    4.016480e-04  3.095034e-03
#> x9~~x9          -4.527370e-03  7.133600e-03   -3.253176e-04 -3.031485e-03
#> visual~~textual  4.016480e-04 -3.253176e-04    4.406254e-03  8.067023e-04
#> visual~~speed    3.095034e-03 -3.031485e-03    8.067023e-04  6.959602e-03
#> textual~~speed   2.247422e-03 -1.581332e-03    1.758606e-03  2.411570e-03
#>                 textual~~speed
#> visual=~x1        0.0007433904
#> visual=~x2       -0.0008599445
#> visual=~x3       -0.0006076821
#> textual=~x4       0.0003298792
#> textual=~x5       0.0011658767
#> textual=~x6       0.0005850574
#> speed=~x7        -0.0001590911
#> speed=~x8        -0.0010561780
#> speed=~x9         0.0014564398
#> x1~~x1           -0.0004451294
#> x2~~x2            0.0001895450
#> x3~~x3            0.0011434569
#> x4~~x4            0.0001042183
#> x5~~x5            0.0002684125
#> x6~~x6            0.0005365461
#> x7~~x7            0.0002627113
#> x8~~x8            0.0022474220
#> x9~~x9           -0.0015813321
#> visual~~textual   0.0017586064
#> visual~~speed     0.0024115696
#> textual~~speed    0.0062624927

# Internal parameterisation (Laplace approximation)
vcov(fit, type = "theta")
#>                    visual=~x1    visual=~x2    visual=~x3   textual=~x4
#> visual=~x1       6.944819e-03 -9.228843e-04 -9.935751e-04  3.389857e-04
#> visual=~x2      -9.228843e-04  6.568863e-03  1.577215e-03  1.410972e-04
#> visual=~x3      -9.935751e-04  1.577215e-03  6.061736e-03  1.941726e-04
#> textual=~x4      3.389857e-04  1.410972e-04  1.941726e-04  3.223624e-03
#> textual=~x5      3.141650e-04  2.000129e-04  2.591496e-04  1.688801e-03
#> textual=~x6      2.546537e-04  1.692562e-04  2.206257e-04  1.376938e-03
#> speed=~x7        3.049172e-04  7.658901e-05  5.006757e-05  8.922238e-05
#> speed=~x8        3.435120e-04  1.427343e-04  4.258362e-05  9.357838e-05
#> speed=~x9        6.233622e-05  1.250035e-04  3.071841e-04  7.405772e-05
#> x1~~x1          -1.222508e-02  5.410147e-03  6.403818e-03 -1.279744e-04
#> x2~~x2           1.466413e-03 -2.090859e-03 -9.100595e-04  2.237214e-05
#> x3~~x3           3.072863e-03 -1.610649e-03 -3.944877e-03  3.916836e-05
#> x4~~x4          -2.114624e-04  1.400470e-04  1.444548e-04 -1.889064e-03
#> x5~~x5           9.341801e-05 -6.713443e-05 -6.283285e-05  5.955463e-04
#> x6~~x6           1.154535e-04 -7.284124e-05 -8.025778e-05  6.602594e-04
#> x7~~x7          -1.669503e-04  5.012766e-05  1.437528e-04 -1.903188e-05
#> x8~~x8          -3.329928e-04 -4.342221e-05  3.783558e-04 -8.334976e-06
#> x9~~x9           3.507098e-04 -1.901485e-05 -3.605765e-04  2.021852e-05
#> visual~~textual  9.461949e-04  2.082300e-05 -3.796066e-04  7.715873e-04
#> visual~~speed   -4.825363e-04  3.751536e-04  9.875436e-04  2.256550e-04
#> textual~~speed   4.314192e-04  2.654701e-04  4.026555e-04  4.093776e-04
#>                   textual=~x5   textual=~x6     speed=~x7     speed=~x8
#> visual=~x1       3.141650e-04  2.546537e-04  3.049172e-04  3.435120e-04
#> visual=~x2       2.000129e-04  1.692562e-04  7.658901e-05  1.427343e-04
#> visual=~x3       2.591496e-04  2.206257e-04  5.006757e-05  4.258362e-05
#> textual=~x4      1.688801e-03  1.376938e-03  8.922238e-05  9.357838e-05
#> textual=~x5      3.930062e-03  1.578490e-03  7.437476e-05  9.582845e-05
#> textual=~x6      1.578490e-03  2.898913e-03  7.457536e-05  9.654040e-05
#> speed=~x7        7.437476e-05  7.457536e-05  5.475475e-03  1.898565e-03
#> speed=~x8        9.582845e-05  9.654040e-05  1.898565e-03  5.612839e-03
#> speed=~x9        1.034695e-04  6.539474e-05 -8.867327e-04 -1.687832e-03
#> x1~~x1           6.481860e-05  7.383460e-05 -3.332406e-04 -3.390605e-04
#> x2~~x2          -1.259862e-05 -1.328243e-05  3.157452e-05 -8.647084e-06
#> x3~~x3          -2.269572e-05 -2.738332e-05  1.527783e-04  2.064197e-04
#> x4~~x4           6.459757e-04  6.832362e-04 -5.449319e-05  5.461796e-06
#> x5~~x5          -2.080084e-03  4.801831e-04  6.677323e-05  4.678771e-05
#> x6~~x6           5.039398e-04 -1.603122e-03 -7.335699e-06 -4.579406e-05
#> x7~~x7           1.757196e-05 -5.183366e-06 -3.368940e-03 -1.767978e-03
#> x8~~x8           1.574371e-05 -3.771666e-05 -3.429949e-03 -9.323437e-03
#> x9~~x9          -2.677522e-05  2.604058e-05  3.695614e-03  5.862929e-03
#> visual~~textual  5.976293e-04  7.336587e-04  1.891355e-04  2.281799e-04
#> visual~~speed    2.858364e-04  2.187081e-04 -1.926030e-03 -2.320677e-03
#> textual~~speed   4.661669e-04  4.047452e-04 -3.246937e-04 -8.927829e-04
#>                     speed=~x9        x1~~x1        x2~~x2        x3~~x3
#> visual=~x1       6.233622e-05 -0.0122250842  1.466413e-03  3.072863e-03
#> visual=~x2       1.250035e-04  0.0054101467 -2.090859e-03 -1.610649e-03
#> visual=~x3       3.071841e-04  0.0064038183 -9.100595e-04 -3.944877e-03
#> textual=~x4      7.405772e-05 -0.0001279744  2.237214e-05  3.916836e-05
#> textual=~x5      1.034695e-04  0.0000648186 -1.259862e-05 -2.269572e-05
#> textual=~x6      6.539474e-05  0.0000738346 -1.328243e-05 -2.738332e-05
#> speed=~x7       -8.867327e-04 -0.0003332406  3.157452e-05  1.527783e-04
#> speed=~x8       -1.687832e-03 -0.0003390605 -8.647084e-06  2.064197e-04
#> speed=~x9        5.894651e-03  0.0004510226 -1.161187e-05 -2.491848e-04
#> x1~~x1           4.510226e-04  0.0463660406 -4.766128e-03 -9.986706e-03
#> x2~~x2          -1.161187e-05 -0.0047661278  8.511238e-03  1.417195e-03
#> x3~~x3          -2.491848e-04 -0.0099867063  1.417195e-03  1.281872e-02
#> x4~~x4           2.680573e-05  0.0006882199 -1.230916e-04 -2.246406e-04
#> x5~~x5          -7.597103e-05 -0.0003116688  5.786411e-05  9.502759e-05
#> x6~~x6           4.038167e-05 -0.0003723394  6.453136e-05  1.260108e-04
#> x7~~x7           2.439345e-03  0.0005089865 -4.915125e-05 -2.357085e-04
#> x8~~x8           7.513275e-03  0.0010008137  2.603809e-05 -6.179704e-04
#> x9~~x9          -7.620125e-03 -0.0010703061  2.712778e-05  5.860744e-04
#> visual~~textual  1.179829e-04 -0.0009458378  3.021593e-04  1.338676e-03
#> visual~~speed    3.691166e-03  0.0036899830 -1.057782e-05 -7.935360e-04
#> textual~~speed   1.371421e-03  0.0001321966 -2.703898e-06 -8.890268e-05
#>                        x4~~x4        x5~~x5        x6~~x6        x7~~x7
#> visual=~x1      -2.114624e-04  9.341801e-05  1.154535e-04 -1.669503e-04
#> visual=~x2       1.400470e-04 -6.713443e-05 -7.284124e-05  5.012766e-05
#> visual=~x3       1.444548e-04 -6.283285e-05 -8.025778e-05  1.437528e-04
#> textual=~x4     -1.889064e-03  5.955463e-04  6.602594e-04 -1.903188e-05
#> textual=~x5      6.459757e-04 -2.080084e-03  5.039398e-04  1.757196e-05
#> textual=~x6      6.832362e-04  4.801831e-04 -1.603122e-03 -5.183366e-06
#> speed=~x7       -5.449319e-05  6.677323e-05 -7.335699e-06 -3.368940e-03
#> speed=~x8        5.461796e-06  4.678771e-05 -4.579406e-05 -1.767978e-03
#> speed=~x9        2.680573e-05 -7.597103e-05  4.038167e-05  2.439345e-03
#> x1~~x1           6.882199e-04 -3.116688e-04 -3.723394e-04  5.089865e-04
#> x2~~x2          -1.230916e-04  5.786411e-05  6.453136e-05 -4.915125e-05
#> x3~~x3          -2.246406e-04  9.502759e-05  1.260108e-04 -2.357085e-04
#> x4~~x4           1.672689e-02 -3.182478e-03 -3.506878e-03  8.429821e-05
#> x5~~x5          -3.182478e-03  1.691445e-02 -2.479085e-03 -1.040738e-04
#> x6~~x6          -3.506878e-03 -2.479085e-03  1.491448e-02  1.148294e-05
#> x7~~x7           8.429821e-05 -1.040738e-04  1.148294e-05  1.190422e-02
#> x8~~x8          -1.761092e-05 -1.415088e-04  1.382795e-04  5.305242e-03
#> x9~~x9          -6.349724e-05  1.774102e-04 -9.433640e-05 -5.724535e-03
#> visual~~textual -2.529357e-04  1.027251e-03 -3.238691e-04 -5.033374e-05
#> visual~~speed    7.175173e-05 -9.662949e-05  1.430018e-05  3.700656e-03
#> textual~~speed   1.292443e-04  8.197794e-05 -1.581051e-05  9.250200e-04
#>                        x8~~x8        x9~~x9 visual~~textual visual~~speed
#> visual=~x1      -3.329928e-04  3.507098e-04    9.461949e-04 -4.825363e-04
#> visual=~x2      -4.342221e-05 -1.901485e-05    2.082300e-05  3.751536e-04
#> visual=~x3       3.783558e-04 -3.605765e-04   -3.796066e-04  9.875436e-04
#> textual=~x4     -8.334976e-06  2.021852e-05    7.715873e-04  2.256550e-04
#> textual=~x5      1.574371e-05 -2.677522e-05    5.976293e-04  2.858364e-04
#> textual=~x6     -3.771666e-05  2.604058e-05    7.336587e-04  2.187081e-04
#> speed=~x7       -3.429949e-03  3.695614e-03    1.891355e-04 -1.926030e-03
#> speed=~x8       -9.323437e-03  5.862929e-03    2.281799e-04 -2.320677e-03
#> speed=~x9        7.513275e-03 -7.620125e-03    1.179829e-04  3.691166e-03
#> x1~~x1           1.000814e-03 -1.070306e-03   -9.458378e-04  3.689983e-03
#> x2~~x2           2.603809e-05  2.712778e-05    3.021593e-04 -1.057782e-05
#> x3~~x3          -6.179704e-04  5.860744e-04    1.338676e-03 -7.935360e-04
#> x4~~x4          -1.761092e-05 -6.349724e-05   -2.529357e-04  7.175173e-05
#> x5~~x5          -1.415088e-04  1.774102e-04    1.027251e-03 -9.662949e-05
#> x6~~x6           1.382795e-04 -9.433640e-05   -3.238691e-04  1.430018e-05
#> x7~~x7           5.305242e-03 -5.724535e-03   -5.033374e-05  3.700656e-03
#> x8~~x8           3.472513e-02 -1.762404e-02   -1.304225e-04  8.608315e-03
#> x9~~x9          -1.762404e-02  2.454052e-02    1.179607e-04 -7.497054e-03
#> visual~~textual -1.304225e-04  1.179607e-04    6.423496e-03  7.193127e-04
#> visual~~speed    8.608315e-03 -7.497054e-03    7.193127e-04  1.177453e-02
#> textual~~speed   3.652556e-03 -2.529804e-03    1.493595e-03  3.066816e-03
#>                 textual~~speed
#> visual=~x1        4.314192e-04
#> visual=~x2        2.654701e-04
#> visual=~x3        4.026555e-04
#> textual=~x4       4.093776e-04
#> textual=~x5       4.661669e-04
#> textual=~x6       4.047452e-04
#> speed=~x7        -3.246937e-04
#> speed=~x8        -8.927829e-04
#> speed=~x9         1.371421e-03
#> x1~~x1            1.321966e-04
#> x2~~x2           -2.703898e-06
#> x3~~x3           -8.890268e-05
#> x4~~x4            1.292443e-04
#> x5~~x5            8.197794e-05
#> x6~~x6           -1.581051e-05
#> x7~~x7            9.250200e-04
#> x8~~x8            3.652556e-03
#> x9~~x9           -2.529804e-03
#> visual~~textual   1.493595e-03
#> visual~~speed     3.066816e-03
#> textual~~speed    5.945061e-03
# }