Variance-Covariance Matrix for INLAvaan Models

Extract the posterior variance-covariance matrix of model parameters from a fitted INLAvaan model.

Usage

# S4 method for class 'INLAvaan'
vcov(object, type = c("lavaan", "theta"), ...)

Arguments

object

An object of class INLAvaan.

type

Character. "lavaan" (default) returns the posterior covariance matrix of the model parameters computed from posterior samples (matching lavaan output). "theta" returns the Laplace approximation covariance in the internal parameterisation.

...

Currently unused.

Value

A square numeric matrix.

See also

Examples

# \donttest{
HS.model <- "
  visual  =~ x1 + x2 + x3
  textual =~ x4 + x5 + x6
  speed   =~ x7 + x8 + x9
"
utils::data("HolzingerSwineford1939", package = "lavaan")
fit <- acfa(HS.model, HolzingerSwineford1939, std.lv = TRUE, nsamp = 100,
            test = "none", verbose = FALSE)

# Default: posterior covariance of lavaan parameters
vcov(fit)
#>                    visual=~x1    visual=~x2    visual=~x3   textual=~x4
#> visual=~x1       6.847538e-03 -1.412008e-03 -8.196304e-04  1.002393e-03
#> visual=~x2      -1.412008e-03  5.643680e-03  2.042673e-03  1.655345e-04
#> visual=~x3      -8.196304e-04  2.042673e-03  6.692347e-03  5.405046e-04
#> textual=~x4      1.002393e-03  1.655345e-04  5.405046e-04  2.415708e-03
#> textual=~x5      1.391023e-03  1.750327e-05  1.016579e-05  1.433172e-03
#> textual=~x6      5.932711e-04  1.854464e-04  6.562130e-04  6.946984e-04
#> speed=~x7        6.754114e-04 -3.365962e-04 -5.215430e-04 -2.757445e-04
#> speed=~x8        9.373531e-04  2.861740e-04 -8.567036e-04  1.103813e-05
#> speed=~x9       -5.104127e-04  1.574667e-04  5.230883e-04  5.137150e-04
#> x1~~x1          -7.771143e-03  4.725993e-03  4.147244e-03 -5.453412e-04
#> x2~~x2           2.348616e-03 -1.898994e-03 -1.905215e-03 -4.037545e-04
#> x3~~x3           2.843093e-03 -2.290764e-03 -4.580720e-03 -1.802005e-04
#> x4~~x4          -2.888516e-04  7.570994e-05  1.516207e-05 -6.765642e-04
#> x5~~x5          -2.292643e-04  1.427267e-05  3.682219e-04  4.287643e-04
#> x6~~x6           5.006054e-05 -2.100117e-04 -4.616133e-04  2.755004e-04
#> x7~~x7          -1.468657e-03  1.456483e-04  5.752518e-05 -7.403259e-04
#> x8~~x8          -4.870778e-04 -5.153287e-04  1.585256e-04  2.765141e-04
#> x9~~x9          -7.569126e-05  7.671686e-05 -1.441393e-03 -3.911255e-04
#> visual~~textual  1.909355e-04  3.276829e-04 -8.897202e-05  2.769292e-04
#> visual~~speed   -2.004819e-03  5.214882e-04  1.362945e-03 -2.192979e-04
#> textual~~speed  -6.628713e-04  1.527536e-04  3.072322e-06  7.795972e-05
#>                   textual=~x5   textual=~x6     speed=~x7     speed=~x8
#> visual=~x1       1.391023e-03  5.932711e-04  6.754114e-04  9.373531e-04
#> visual=~x2       1.750327e-05  1.854464e-04 -3.365962e-04  2.861740e-04
#> visual=~x3       1.016579e-05  6.562130e-04 -5.215430e-04 -8.567036e-04
#> textual=~x4      1.433172e-03  6.946984e-04 -2.757445e-04  1.103813e-05
#> textual=~x5      4.148521e-03  1.397525e-03  1.840193e-04 -8.896656e-04
#> textual=~x6      1.397525e-03  1.963308e-03 -1.747663e-05 -3.354776e-04
#> speed=~x7        1.840193e-04 -1.747663e-05  5.815542e-03  2.342792e-03
#> speed=~x8       -8.896656e-04 -3.354776e-04  2.342792e-03  5.374343e-03
#> speed=~x9        3.445170e-04 -1.288758e-04 -1.144978e-03 -1.426619e-03
#> x1~~x1          -5.654062e-04  5.100267e-04 -1.648550e-03 -1.257961e-03
#> x2~~x2           3.262175e-04 -3.854760e-04  5.785550e-04  1.628736e-04
#> x3~~x3          -1.824719e-04 -6.337393e-04  1.429897e-03  1.707489e-03
#> x4~~x4           1.798594e-04  1.288724e-04 -4.328785e-05 -3.635727e-04
#> x5~~x5          -1.099462e-03 -3.765781e-04 -5.225957e-04  2.985409e-04
#> x6~~x6          -8.469294e-05 -3.712681e-04  1.927254e-04  3.583642e-04
#> x7~~x7          -1.397395e-03 -6.007579e-04 -3.458024e-03 -8.395279e-04
#> x8~~x8           9.964741e-04 -2.154562e-05 -1.874030e-03 -4.148270e-03
#> x9~~x9          -7.841455e-04 -4.153911e-04  2.720352e-03  3.445408e-03
#> visual~~textual  1.312136e-04  1.655927e-04  8.503478e-05  5.640515e-04
#> visual~~speed   -9.433983e-04 -5.137909e-04 -1.632422e-03 -1.958550e-03
#> textual~~speed  -2.308592e-04 -9.830260e-05 -4.029309e-04 -3.220829e-04
#>                     speed=~x9        x1~~x1        x2~~x2        x3~~x3
#> visual=~x1      -0.0005104127 -0.0077711427  0.0023486159  2.843093e-03
#> visual=~x2       0.0001574667  0.0047259930 -0.0018989940 -2.290764e-03
#> visual=~x3       0.0005230883  0.0041472441 -0.0019052149 -4.580720e-03
#> textual=~x4      0.0005137150 -0.0005453412 -0.0004037545 -1.802005e-04
#> textual=~x5      0.0003445170 -0.0005654062  0.0003262175 -1.824719e-04
#> textual=~x6     -0.0001288758  0.0005100267 -0.0003854760 -6.337393e-04
#> speed=~x7       -0.0011449778 -0.0016485504  0.0005785550  1.429897e-03
#> speed=~x8       -0.0014266192 -0.0012579611  0.0001628736  1.707489e-03
#> speed=~x9        0.0062696115  0.0014307919 -0.0014654627 -1.007833e-03
#> x1~~x1           0.0014307919  0.0158039209 -0.0053972809 -7.300628e-03
#> x2~~x2          -0.0014654627 -0.0053972809  0.0131775767  1.472780e-03
#> x3~~x3          -0.0010078329 -0.0073006285  0.0014727797  1.119177e-02
#> x4~~x4           0.0001210293  0.0004937304 -0.0003920801 -5.381854e-05
#> x5~~x5           0.0001021271 -0.0002629080  0.0003244541 -5.489556e-04
#> x6~~x6          -0.0002183344 -0.0005737319  0.0004101031  8.765432e-04
#> x7~~x7           0.0017884841  0.0016865813 -0.0007331177 -1.250641e-04
#> x8~~x8           0.0032358009  0.0003561320  0.0002111940 -7.898368e-04
#> x9~~x9          -0.0051372468 -0.0010934576  0.0004460766  2.327830e-03
#> visual~~textual -0.0003392786 -0.0005132122 -0.0002475150  6.028219e-04
#> visual~~speed    0.0031865990  0.0033720807 -0.0001251571 -1.814244e-03
#> textual~~speed   0.0008226112  0.0009387047 -0.0004037773 -2.256635e-04
#>                        x4~~x4        x5~~x5        x6~~x6        x7~~x7
#> visual=~x1      -2.888516e-04 -2.292643e-04  5.006054e-05 -1.468657e-03
#> visual=~x2       7.570994e-05  1.427267e-05 -2.100117e-04  1.456483e-04
#> visual=~x3       1.516207e-05  3.682219e-04 -4.616133e-04  5.752518e-05
#> textual=~x4     -6.765642e-04  4.287643e-04  2.755004e-04 -7.403259e-04
#> textual=~x5      1.798594e-04 -1.099462e-03 -8.469294e-05 -1.397395e-03
#> textual=~x6      1.288724e-04 -3.765781e-04 -3.712681e-04 -6.007579e-04
#> speed=~x7       -4.328785e-05 -5.225957e-04  1.927254e-04 -3.458024e-03
#> speed=~x8       -3.635727e-04  2.985409e-04  3.583642e-04 -8.395279e-04
#> speed=~x9        1.210293e-04  1.021271e-04 -2.183344e-04  1.788484e-03
#> x1~~x1           4.937304e-04 -2.629080e-04 -5.737319e-04  1.686581e-03
#> x2~~x2          -3.920801e-04  3.244541e-04  4.101031e-04 -7.331177e-04
#> x3~~x3          -5.381854e-05 -5.489556e-04  8.765432e-04 -1.250641e-04
#> x4~~x4           2.563896e-03 -8.977679e-04 -5.251396e-04 -1.910518e-04
#> x5~~x5          -8.977679e-04  3.401453e-03 -1.727292e-04  8.589562e-04
#> x6~~x6          -5.251396e-04 -1.727292e-04  1.838956e-03 -7.537420e-05
#> x7~~x7          -1.910518e-04  8.589562e-04 -7.537420e-05  1.064637e-02
#> x8~~x8           6.998711e-04 -3.479767e-04 -4.080486e-04  1.439718e-03
#> x9~~x9           1.553826e-04 -3.213926e-04  8.617011e-04 -2.539044e-03
#> visual~~textual -7.533719e-04  6.034468e-04 -3.270652e-06  2.335296e-04
#> visual~~speed    2.042577e-04  7.184816e-05 -9.254184e-05  2.444751e-03
#> textual~~speed   1.706279e-04 -2.175334e-04 -1.546909e-04  2.432527e-03
#>                        x8~~x8        x9~~x9 visual~~textual visual~~speed
#> visual=~x1      -4.870778e-04 -7.569126e-05    1.909355e-04 -2.004819e-03
#> visual=~x2      -5.153287e-04  7.671686e-05    3.276829e-04  5.214882e-04
#> visual=~x3       1.585256e-04 -1.441393e-03   -8.897202e-05  1.362945e-03
#> textual=~x4      2.765141e-04 -3.911255e-04    2.769292e-04 -2.192979e-04
#> textual=~x5      9.964741e-04 -7.841455e-04    1.312136e-04 -9.433983e-04
#> textual=~x6     -2.154562e-05 -4.153911e-04    1.655927e-04 -5.137909e-04
#> speed=~x7       -1.874030e-03  2.720352e-03    8.503478e-05 -1.632422e-03
#> speed=~x8       -4.148270e-03  3.445408e-03    5.640515e-04 -1.958550e-03
#> speed=~x9        3.235801e-03 -5.137247e-03   -3.392786e-04  3.186599e-03
#> x1~~x1           3.561320e-04 -1.093458e-03   -5.132122e-04  3.372081e-03
#> x2~~x2           2.111940e-04  4.460766e-04   -2.475150e-04 -1.251571e-04
#> x3~~x3          -7.898368e-04  2.327830e-03    6.028219e-04 -1.814244e-03
#> x4~~x4           6.998711e-04  1.553826e-04   -7.533719e-04  2.042577e-04
#> x5~~x5          -3.479767e-04 -3.213926e-04    6.034468e-04  7.184816e-05
#> x6~~x6          -4.080486e-04  8.617011e-04   -3.270652e-06 -9.254184e-05
#> x7~~x7           1.439718e-03 -2.539044e-03    2.335296e-04  2.444751e-03
#> x8~~x8           7.352235e-03 -4.914877e-03   -2.148178e-04  3.576341e-03
#> x9~~x9          -4.914877e-03  8.593392e-03    6.942857e-04 -3.322114e-03
#> visual~~textual -2.148178e-04  6.942857e-04    3.189345e-03  3.498811e-04
#> visual~~speed    3.576341e-03 -3.322114e-03    3.498811e-04  7.504892e-03
#> textual~~speed   1.739016e-03 -7.896997e-04    3.547672e-04  2.451169e-03
#>                 textual~~speed
#> visual=~x1       -6.628713e-04
#> visual=~x2        1.527536e-04
#> visual=~x3        3.072322e-06
#> textual=~x4       7.795972e-05
#> textual=~x5      -2.308592e-04
#> textual=~x6      -9.830260e-05
#> speed=~x7        -4.029309e-04
#> speed=~x8        -3.220829e-04
#> speed=~x9         8.226112e-04
#> x1~~x1            9.387047e-04
#> x2~~x2           -4.037773e-04
#> x3~~x3           -2.256635e-04
#> x4~~x4            1.706279e-04
#> x5~~x5           -2.175334e-04
#> x6~~x6           -1.546909e-04
#> x7~~x7            2.432527e-03
#> x8~~x8            1.739016e-03
#> x9~~x9           -7.896997e-04
#> visual~~textual   3.547672e-04
#> visual~~speed     2.451169e-03
#> textual~~speed    4.181388e-03

# Internal parameterisation (Laplace approximation)
vcov(fit, type = "theta")
#>                    visual=~x1    visual=~x2    visual=~x3   textual=~x4
#> visual=~x1       6.944819e-03 -9.228843e-04 -9.935751e-04  3.389857e-04
#> visual=~x2      -9.228843e-04  6.568863e-03  1.577215e-03  1.410972e-04
#> visual=~x3      -9.935751e-04  1.577215e-03  6.061736e-03  1.941726e-04
#> textual=~x4      3.389857e-04  1.410972e-04  1.941726e-04  3.223624e-03
#> textual=~x5      3.141650e-04  2.000129e-04  2.591496e-04  1.688801e-03
#> textual=~x6      2.546537e-04  1.692562e-04  2.206257e-04  1.376938e-03
#> speed=~x7        3.049172e-04  7.658901e-05  5.006757e-05  8.922238e-05
#> speed=~x8        3.435120e-04  1.427343e-04  4.258362e-05  9.357838e-05
#> speed=~x9        6.233622e-05  1.250035e-04  3.071841e-04  7.405772e-05
#> x1~~x1          -1.222508e-02  5.410147e-03  6.403818e-03 -1.279744e-04
#> x2~~x2           1.466413e-03 -2.090859e-03 -9.100595e-04  2.237214e-05
#> x3~~x3           3.072863e-03 -1.610649e-03 -3.944877e-03  3.916836e-05
#> x4~~x4          -2.114624e-04  1.400470e-04  1.444548e-04 -1.889064e-03
#> x5~~x5           9.341801e-05 -6.713443e-05 -6.283285e-05  5.955463e-04
#> x6~~x6           1.154535e-04 -7.284124e-05 -8.025778e-05  6.602594e-04
#> x7~~x7          -1.669503e-04  5.012766e-05  1.437528e-04 -1.903188e-05
#> x8~~x8          -3.329928e-04 -4.342221e-05  3.783558e-04 -8.334977e-06
#> x9~~x9           3.507098e-04 -1.901485e-05 -3.605765e-04  2.021852e-05
#> visual~~textual  9.461949e-04  2.082300e-05 -3.796066e-04  7.715873e-04
#> visual~~speed   -4.825363e-04  3.751536e-04  9.875436e-04  2.256550e-04
#> textual~~speed   4.314192e-04  2.654701e-04  4.026555e-04  4.093776e-04
#>                   textual=~x5   textual=~x6     speed=~x7     speed=~x8
#> visual=~x1       3.141650e-04  2.546537e-04  3.049172e-04  3.435120e-04
#> visual=~x2       2.000129e-04  1.692562e-04  7.658901e-05  1.427343e-04
#> visual=~x3       2.591496e-04  2.206257e-04  5.006757e-05  4.258362e-05
#> textual=~x4      1.688801e-03  1.376938e-03  8.922238e-05  9.357838e-05
#> textual=~x5      3.930062e-03  1.578490e-03  7.437476e-05  9.582845e-05
#> textual=~x6      1.578490e-03  2.898913e-03  7.457536e-05  9.654040e-05
#> speed=~x7        7.437476e-05  7.457536e-05  5.475475e-03  1.898565e-03
#> speed=~x8        9.582845e-05  9.654040e-05  1.898565e-03  5.612839e-03
#> speed=~x9        1.034695e-04  6.539474e-05 -8.867327e-04 -1.687832e-03
#> x1~~x1           6.481860e-05  7.383460e-05 -3.332406e-04 -3.390605e-04
#> x2~~x2          -1.259862e-05 -1.328243e-05  3.157452e-05 -8.647084e-06
#> x3~~x3          -2.269572e-05 -2.738332e-05  1.527783e-04  2.064197e-04
#> x4~~x4           6.459757e-04  6.832362e-04 -5.449319e-05  5.461796e-06
#> x5~~x5          -2.080084e-03  4.801831e-04  6.677323e-05  4.678771e-05
#> x6~~x6           5.039398e-04 -1.603122e-03 -7.335699e-06 -4.579406e-05
#> x7~~x7           1.757196e-05 -5.183366e-06 -3.368940e-03 -1.767978e-03
#> x8~~x8           1.574371e-05 -3.771666e-05 -3.429949e-03 -9.323437e-03
#> x9~~x9          -2.677522e-05  2.604058e-05  3.695614e-03  5.862929e-03
#> visual~~textual  5.976293e-04  7.336587e-04  1.891355e-04  2.281799e-04
#> visual~~speed    2.858364e-04  2.187081e-04 -1.926030e-03 -2.320677e-03
#> textual~~speed   4.661669e-04  4.047452e-04 -3.246937e-04 -8.927829e-04
#>                     speed=~x9        x1~~x1        x2~~x2        x3~~x3
#> visual=~x1       6.233622e-05 -0.0122250842  1.466413e-03  3.072863e-03
#> visual=~x2       1.250035e-04  0.0054101467 -2.090859e-03 -1.610649e-03
#> visual=~x3       3.071841e-04  0.0064038183 -9.100595e-04 -3.944877e-03
#> textual=~x4      7.405772e-05 -0.0001279744  2.237214e-05  3.916836e-05
#> textual=~x5      1.034695e-04  0.0000648186 -1.259862e-05 -2.269572e-05
#> textual=~x6      6.539474e-05  0.0000738346 -1.328243e-05 -2.738332e-05
#> speed=~x7       -8.867327e-04 -0.0003332406  3.157452e-05  1.527783e-04
#> speed=~x8       -1.687832e-03 -0.0003390605 -8.647084e-06  2.064197e-04
#> speed=~x9        5.894651e-03  0.0004510226 -1.161187e-05 -2.491848e-04
#> x1~~x1           4.510226e-04  0.0463660406 -4.766128e-03 -9.986706e-03
#> x2~~x2          -1.161187e-05 -0.0047661278  8.511238e-03  1.417195e-03
#> x3~~x3          -2.491848e-04 -0.0099867063  1.417195e-03  1.281872e-02
#> x4~~x4           2.680573e-05  0.0006882199 -1.230916e-04 -2.246406e-04
#> x5~~x5          -7.597103e-05 -0.0003116688  5.786411e-05  9.502759e-05
#> x6~~x6           4.038167e-05 -0.0003723394  6.453136e-05  1.260108e-04
#> x7~~x7           2.439345e-03  0.0005089865 -4.915125e-05 -2.357085e-04
#> x8~~x8           7.513275e-03  0.0010008137  2.603809e-05 -6.179704e-04
#> x9~~x9          -7.620125e-03 -0.0010703061  2.712778e-05  5.860744e-04
#> visual~~textual  1.179829e-04 -0.0009458378  3.021593e-04  1.338676e-03
#> visual~~speed    3.691166e-03  0.0036899830 -1.057782e-05 -7.935360e-04
#> textual~~speed   1.371421e-03  0.0001321966 -2.703898e-06 -8.890268e-05
#>                        x4~~x4        x5~~x5        x6~~x6        x7~~x7
#> visual=~x1      -2.114624e-04  9.341801e-05  1.154535e-04 -1.669503e-04
#> visual=~x2       1.400470e-04 -6.713443e-05 -7.284124e-05  5.012766e-05
#> visual=~x3       1.444548e-04 -6.283285e-05 -8.025778e-05  1.437528e-04
#> textual=~x4     -1.889064e-03  5.955463e-04  6.602594e-04 -1.903188e-05
#> textual=~x5      6.459757e-04 -2.080084e-03  5.039398e-04  1.757196e-05
#> textual=~x6      6.832362e-04  4.801831e-04 -1.603122e-03 -5.183366e-06
#> speed=~x7       -5.449319e-05  6.677323e-05 -7.335699e-06 -3.368940e-03
#> speed=~x8        5.461796e-06  4.678771e-05 -4.579406e-05 -1.767978e-03
#> speed=~x9        2.680573e-05 -7.597103e-05  4.038167e-05  2.439345e-03
#> x1~~x1           6.882199e-04 -3.116688e-04 -3.723394e-04  5.089865e-04
#> x2~~x2          -1.230916e-04  5.786411e-05  6.453136e-05 -4.915125e-05
#> x3~~x3          -2.246406e-04  9.502759e-05  1.260108e-04 -2.357085e-04
#> x4~~x4           1.672689e-02 -3.182478e-03 -3.506878e-03  8.429821e-05
#> x5~~x5          -3.182478e-03  1.691445e-02 -2.479085e-03 -1.040738e-04
#> x6~~x6          -3.506878e-03 -2.479085e-03  1.491448e-02  1.148294e-05
#> x7~~x7           8.429821e-05 -1.040738e-04  1.148294e-05  1.190422e-02
#> x8~~x8          -1.761092e-05 -1.415088e-04  1.382795e-04  5.305242e-03
#> x9~~x9          -6.349724e-05  1.774102e-04 -9.433640e-05 -5.724535e-03
#> visual~~textual -2.529357e-04  1.027251e-03 -3.238691e-04 -5.033374e-05
#> visual~~speed    7.175173e-05 -9.662949e-05  1.430018e-05  3.700656e-03
#> textual~~speed   1.292443e-04  8.197794e-05 -1.581051e-05  9.250200e-04
#>                        x8~~x8        x9~~x9 visual~~textual visual~~speed
#> visual=~x1      -3.329928e-04  3.507098e-04    9.461949e-04 -4.825363e-04
#> visual=~x2      -4.342221e-05 -1.901485e-05    2.082300e-05  3.751536e-04
#> visual=~x3       3.783558e-04 -3.605765e-04   -3.796066e-04  9.875436e-04
#> textual=~x4     -8.334977e-06  2.021852e-05    7.715873e-04  2.256550e-04
#> textual=~x5      1.574371e-05 -2.677522e-05    5.976293e-04  2.858364e-04
#> textual=~x6     -3.771666e-05  2.604058e-05    7.336587e-04  2.187081e-04
#> speed=~x7       -3.429949e-03  3.695614e-03    1.891355e-04 -1.926030e-03
#> speed=~x8       -9.323437e-03  5.862929e-03    2.281799e-04 -2.320677e-03
#> speed=~x9        7.513275e-03 -7.620125e-03    1.179829e-04  3.691166e-03
#> x1~~x1           1.000814e-03 -1.070306e-03   -9.458378e-04  3.689983e-03
#> x2~~x2           2.603809e-05  2.712778e-05    3.021593e-04 -1.057782e-05
#> x3~~x3          -6.179704e-04  5.860744e-04    1.338676e-03 -7.935360e-04
#> x4~~x4          -1.761092e-05 -6.349724e-05   -2.529357e-04  7.175173e-05
#> x5~~x5          -1.415088e-04  1.774102e-04    1.027251e-03 -9.662949e-05
#> x6~~x6           1.382795e-04 -9.433640e-05   -3.238691e-04  1.430018e-05
#> x7~~x7           5.305242e-03 -5.724535e-03   -5.033374e-05  3.700656e-03
#> x8~~x8           3.472513e-02 -1.762404e-02   -1.304225e-04  8.608315e-03
#> x9~~x9          -1.762404e-02  2.454052e-02    1.179607e-04 -7.497054e-03
#> visual~~textual -1.304225e-04  1.179607e-04    6.423496e-03  7.193127e-04
#> visual~~speed    8.608315e-03 -7.497054e-03    7.193127e-04  1.177453e-02
#> textual~~speed   3.652556e-03 -2.529804e-03    1.493595e-03  3.066816e-03
#>                 textual~~speed
#> visual=~x1        4.314192e-04
#> visual=~x2        2.654701e-04
#> visual=~x3        4.026555e-04
#> textual=~x4       4.093776e-04
#> textual=~x5       4.661669e-04
#> textual=~x6       4.047452e-04
#> speed=~x7        -3.246937e-04
#> speed=~x8        -8.927829e-04
#> speed=~x9         1.371421e-03
#> x1~~x1            1.321966e-04
#> x2~~x2           -2.703898e-06
#> x3~~x3           -8.890268e-05
#> x4~~x4            1.292443e-04
#> x5~~x5            8.197794e-05
#> x6~~x6           -1.581051e-05
#> x7~~x7            9.250200e-04
#> x8~~x8            3.652556e-03
#> x9~~x9           -2.529804e-03
#> visual~~textual   1.493595e-03
#> visual~~speed     3.066816e-03
#> textual~~speed    5.945061e-03
# }