Variance-Covariance Matrix for INLAvaan Models

Extract the posterior variance-covariance matrix of model parameters from a fitted INLAvaan model.

Usage

# S4 method for class 'INLAvaan'
vcov(object, type = c("lavaan", "theta"), ...)

Arguments

object

An object of class INLAvaan.

type

Character. "lavaan" (default) returns the posterior covariance matrix of the model parameters computed from posterior samples (matching lavaan output). "theta" returns the Laplace approximation covariance in the internal parameterisation.

...

Currently unused.

Value

A square numeric matrix.

See also

Examples

# \donttest{
HS.model <- "
  visual  =~ x1 + x2 + x3
  textual =~ x4 + x5 + x6
  speed   =~ x7 + x8 + x9
"
utils::data("HolzingerSwineford1939", package = "lavaan")
fit <- acfa(HS.model, HolzingerSwineford1939, std.lv = TRUE, nsamp = 100,
            test = "none", verbose = FALSE)

# Default: posterior covariance of lavaan parameters
vcov(fit)
#>                    visual=~x1    visual=~x2    visual=~x3   textual=~x4
#> visual=~x1       6.847538e-03 -1.408150e-03 -8.143327e-04  1.001825e-03
#> visual=~x2      -1.408150e-03  5.642952e-03  2.035197e-03  1.658913e-04
#> visual=~x3      -8.143327e-04  2.035197e-03  6.691937e-03  5.407233e-04
#> textual=~x4      1.001825e-03  1.658913e-04  5.407233e-04  2.415512e-03
#> textual=~x5      1.390069e-03  1.769288e-05  9.261806e-06  1.427198e-03
#> textual=~x6      5.910155e-04  1.857367e-04  6.565568e-04  6.890950e-04
#> speed=~x7        6.752504e-04 -3.362636e-04 -5.208086e-04 -2.756888e-04
#> speed=~x8        9.373058e-04  2.876865e-04 -8.560006e-04  1.192257e-05
#> speed=~x9       -5.100733e-04  1.568328e-04  5.200306e-04  5.130170e-04
#> x1~~x1          -7.781854e-03  4.731633e-03  4.162375e-03 -5.439575e-04
#> x2~~x2           2.340740e-03 -1.890162e-03 -1.903154e-03 -4.052933e-04
#> x3~~x3           2.840434e-03 -2.292827e-03 -4.572583e-03 -1.807012e-04
#> x4~~x4          -2.891389e-04  7.584997e-05  1.596226e-05 -6.723007e-04
#> x5~~x5          -2.267611e-04  1.285443e-05  3.660103e-04  4.299329e-04
#> x6~~x6           5.005674e-05 -2.103507e-04 -4.606521e-04  2.741039e-04
#> x7~~x7          -1.468546e-03  1.441765e-04  5.519827e-05 -7.424792e-04
#> x8~~x8          -4.805938e-04 -5.174982e-04  1.551834e-04  2.780815e-04
#> x9~~x9          -7.251153e-05  8.122117e-05 -1.430670e-03 -3.895174e-04
#> visual~~textual  1.872738e-04  3.274285e-04 -9.112356e-05  2.737969e-04
#> visual~~speed   -2.005437e-03  5.195984e-04  1.352243e-03 -2.214314e-04
#> textual~~speed  -6.618395e-04  1.519328e-04 -1.316690e-07  7.674689e-05
#>                   textual=~x5   textual=~x6     speed=~x7     speed=~x8
#> visual=~x1       1.390069e-03  5.910155e-04  6.752504e-04  9.373058e-04
#> visual=~x2       1.769288e-05  1.857367e-04 -3.362636e-04  2.876865e-04
#> visual=~x3       9.261806e-06  6.565568e-04 -5.208086e-04 -8.560006e-04
#> textual=~x4      1.427198e-03  6.890950e-04 -2.756888e-04  1.192257e-05
#> textual=~x5      4.150867e-03  1.392171e-03  1.855625e-04 -8.913121e-04
#> textual=~x6      1.392171e-03  1.963736e-03 -1.687588e-05 -3.348509e-04
#> speed=~x7        1.855625e-04 -1.687588e-05  5.815512e-03  2.332412e-03
#> speed=~x8       -8.913121e-04 -3.348509e-04  2.332412e-03  5.372127e-03
#> speed=~x9        3.411031e-04 -1.317800e-04 -1.137785e-03 -1.415352e-03
#> x1~~x1          -5.676453e-04  5.112146e-04 -1.652006e-03 -1.262937e-03
#> x2~~x2           3.255623e-04 -3.875073e-04  5.794225e-04  1.625384e-04
#> x3~~x3          -1.876977e-04 -6.387372e-04  1.433669e-03  1.705248e-03
#> x4~~x4           1.781187e-04  1.271650e-04 -4.478255e-05 -3.623527e-04
#> x5~~x5          -1.091583e-03 -3.766323e-04 -5.220760e-04  2.967084e-04
#> x6~~x6          -8.605878e-05 -3.684909e-04  1.928070e-04  3.575923e-04
#> x7~~x7          -1.398350e-03 -5.990964e-04 -3.447550e-03 -8.282952e-04
#> x8~~x8           9.964738e-04 -2.477096e-05 -1.866017e-03 -4.132572e-03
#> x9~~x9          -7.845758e-04 -4.095310e-04  2.715056e-03  3.441206e-03
#> visual~~textual  1.293621e-04  1.631539e-04  8.591030e-05  5.643504e-04
#> visual~~speed   -9.450728e-04 -5.149098e-04 -1.619270e-03 -1.945521e-03
#> textual~~speed  -2.324567e-04 -9.841289e-05 -4.020656e-04 -3.171088e-04
#>                     speed=~x9        x1~~x1        x2~~x2        x3~~x3
#> visual=~x1      -0.0005100733 -0.0077818536  0.0023407403  2.840434e-03
#> visual=~x2       0.0001568328  0.0047316332 -0.0018901621 -2.292827e-03
#> visual=~x3       0.0005200306  0.0041623755 -0.0019031536 -4.572583e-03
#> textual=~x4      0.0005130170 -0.0005439575 -0.0004052933 -1.807012e-04
#> textual=~x5      0.0003411031 -0.0005676453  0.0003255623 -1.876977e-04
#> textual=~x6     -0.0001317800  0.0005112146 -0.0003875073 -6.387372e-04
#> speed=~x7       -0.0011377846 -0.0016520062  0.0005794225  1.433669e-03
#> speed=~x8       -0.0014153520 -0.0012629370  0.0001625384  1.705248e-03
#> speed=~x9        0.0062705407  0.0014321794 -0.0014666698 -1.006720e-03
#> x1~~x1           0.0014321794  0.0158123090 -0.0053961174 -7.339164e-03
#> x2~~x2          -0.0014666698 -0.0053961174  0.0131854041  1.479064e-03
#> x3~~x3          -0.0010067202 -0.0073391640  0.0014790638  1.121468e-02
#> x4~~x4           0.0001223883  0.0004922678 -0.0003942086 -5.382257e-05
#> x5~~x5           0.0001031556 -0.0002631744  0.0003294796 -5.441343e-04
#> x6~~x6          -0.0002188917 -0.0005743062  0.0004108519  8.775253e-04
#> x7~~x7           0.0017765524  0.0016869353 -0.0007293796 -1.217698e-04
#> x8~~x8           0.0032259731  0.0003453078  0.0002181123 -7.824121e-04
#> x9~~x9          -0.0051261061 -0.0010949478  0.0004404264  2.316625e-03
#> visual~~textual -0.0003398706 -0.0005108709 -0.0002441325  6.056836e-04
#> visual~~speed    0.0031726878  0.0033784557 -0.0001196823 -1.804271e-03
#> textual~~speed   0.0008153564  0.0009373225 -0.0004022549 -2.231427e-04
#>                        x4~~x4        x5~~x5        x6~~x6        x7~~x7
#> visual=~x1      -2.891389e-04 -2.267611e-04  5.005674e-05 -1.468546e-03
#> visual=~x2       7.584997e-05  1.285443e-05 -2.103507e-04  1.441765e-04
#> visual=~x3       1.596226e-05  3.660103e-04 -4.606521e-04  5.519827e-05
#> textual=~x4     -6.723007e-04  4.299329e-04  2.741039e-04 -7.424792e-04
#> textual=~x5      1.781187e-04 -1.091583e-03 -8.605878e-05 -1.398350e-03
#> textual=~x6      1.271650e-04 -3.766323e-04 -3.684909e-04 -5.990964e-04
#> speed=~x7       -4.478255e-05 -5.220760e-04  1.928070e-04 -3.447550e-03
#> speed=~x8       -3.623527e-04  2.967084e-04  3.575923e-04 -8.282952e-04
#> speed=~x9        1.223883e-04  1.031556e-04 -2.188917e-04  1.776552e-03
#> x1~~x1           4.922678e-04 -2.631744e-04 -5.743062e-04  1.686935e-03
#> x2~~x2          -3.942086e-04  3.294796e-04  4.108519e-04 -7.293796e-04
#> x3~~x3          -5.382257e-05 -5.441343e-04  8.775253e-04 -1.217698e-04
#> x4~~x4           2.565543e-03 -8.947624e-04 -5.231026e-04 -1.911204e-04
#> x5~~x5          -8.947624e-04  3.401601e-03 -1.737181e-04  8.560079e-04
#> x6~~x6          -5.231026e-04 -1.737181e-04  1.840363e-03 -7.513063e-05
#> x7~~x7          -1.911204e-04  8.560079e-04 -7.513063e-05  1.064793e-02
#> x8~~x8           6.991976e-04 -3.442621e-04 -4.098808e-04  1.430151e-03
#> x9~~x9           1.487600e-04 -3.200010e-04  8.604165e-04 -2.536225e-03
#> visual~~textual -7.541519e-04  6.033241e-04 -3.666610e-06  2.354149e-04
#> visual~~speed    2.094132e-04  6.595386e-05 -9.143986e-05  2.430653e-03
#> textual~~speed   1.721549e-04 -2.196303e-04 -1.553165e-04  2.435339e-03
#>                        x8~~x8        x9~~x9 visual~~textual visual~~speed
#> visual=~x1      -4.805938e-04 -7.251153e-05    1.872738e-04 -2.005437e-03
#> visual=~x2      -5.174982e-04  8.122117e-05    3.274285e-04  5.195984e-04
#> visual=~x3       1.551834e-04 -1.430670e-03   -9.112356e-05  1.352243e-03
#> textual=~x4      2.780815e-04 -3.895174e-04    2.737969e-04 -2.214314e-04
#> textual=~x5      9.964738e-04 -7.845758e-04    1.293621e-04 -9.450728e-04
#> textual=~x6     -2.477096e-05 -4.095310e-04    1.631539e-04 -5.149098e-04
#> speed=~x7       -1.866017e-03  2.715056e-03    8.591030e-05 -1.619270e-03
#> speed=~x8       -4.132572e-03  3.441206e-03    5.643504e-04 -1.945521e-03
#> speed=~x9        3.225973e-03 -5.126106e-03   -3.398706e-04  3.172688e-03
#> x1~~x1           3.453078e-04 -1.094948e-03   -5.108709e-04  3.378456e-03
#> x2~~x2           2.181123e-04  4.404264e-04   -2.441325e-04 -1.196823e-04
#> x3~~x3          -7.824121e-04  2.316625e-03    6.056836e-04 -1.804271e-03
#> x4~~x4           6.991976e-04  1.487600e-04   -7.541519e-04  2.094132e-04
#> x5~~x5          -3.442621e-04 -3.200010e-04    6.033241e-04  6.595386e-05
#> x6~~x6          -4.098808e-04  8.604165e-04   -3.666610e-06 -9.143986e-05
#> x7~~x7           1.430151e-03 -2.536225e-03    2.354149e-04  2.430653e-03
#> x8~~x8           7.354096e-03 -4.951660e-03   -2.108042e-04  3.576674e-03
#> x9~~x9          -4.951660e-03  8.582390e-03    6.919117e-04 -3.321837e-03
#> visual~~textual -2.108042e-04  6.919117e-04    3.191776e-03  3.510506e-04
#> visual~~speed    3.576674e-03 -3.321837e-03    3.510506e-04  7.506007e-03
#> textual~~speed   1.735450e-03 -8.016008e-04    3.510756e-04  2.446221e-03
#>                 textual~~speed
#> visual=~x1       -6.618395e-04
#> visual=~x2        1.519328e-04
#> visual=~x3       -1.316690e-07
#> textual=~x4       7.674689e-05
#> textual=~x5      -2.324567e-04
#> textual=~x6      -9.841289e-05
#> speed=~x7        -4.020656e-04
#> speed=~x8        -3.171088e-04
#> speed=~x9         8.153564e-04
#> x1~~x1            9.373225e-04
#> x2~~x2           -4.022549e-04
#> x3~~x3           -2.231427e-04
#> x4~~x4            1.721549e-04
#> x5~~x5           -2.196303e-04
#> x6~~x6           -1.553165e-04
#> x7~~x7            2.435339e-03
#> x8~~x8            1.735450e-03
#> x9~~x9           -8.016008e-04
#> visual~~textual   3.510756e-04
#> visual~~speed     2.446221e-03
#> textual~~speed    4.183431e-03

# Internal parameterisation (Laplace approximation)
vcov(fit, type = "theta")
#>                    visual=~x1    visual=~x2    visual=~x3   textual=~x4
#> visual=~x1       6.944819e-03 -9.228843e-04 -9.935751e-04  3.389857e-04
#> visual=~x2      -9.228843e-04  6.568863e-03  1.577215e-03  1.410972e-04
#> visual=~x3      -9.935751e-04  1.577215e-03  6.061736e-03  1.941726e-04
#> textual=~x4      3.389857e-04  1.410972e-04  1.941726e-04  3.223624e-03
#> textual=~x5      3.141650e-04  2.000129e-04  2.591496e-04  1.688801e-03
#> textual=~x6      2.546537e-04  1.692562e-04  2.206257e-04  1.376938e-03
#> speed=~x7        3.049172e-04  7.658901e-05  5.006757e-05  8.922238e-05
#> speed=~x8        3.435120e-04  1.427343e-04  4.258362e-05  9.357838e-05
#> speed=~x9        6.233622e-05  1.250035e-04  3.071841e-04  7.405772e-05
#> x1~~x1          -1.222508e-02  5.410147e-03  6.403818e-03 -1.279744e-04
#> x2~~x2           1.466413e-03 -2.090859e-03 -9.100595e-04  2.237214e-05
#> x3~~x3           3.072863e-03 -1.610649e-03 -3.944877e-03  3.916836e-05
#> x4~~x4          -2.114624e-04  1.400470e-04  1.444548e-04 -1.889064e-03
#> x5~~x5           9.341801e-05 -6.713443e-05 -6.283285e-05  5.955463e-04
#> x6~~x6           1.154535e-04 -7.284124e-05 -8.025778e-05  6.602594e-04
#> x7~~x7          -1.669503e-04  5.012766e-05  1.437528e-04 -1.903188e-05
#> x8~~x8          -3.329928e-04 -4.342221e-05  3.783558e-04 -8.334976e-06
#> x9~~x9           3.507098e-04 -1.901485e-05 -3.605765e-04  2.021852e-05
#> visual~~textual  9.461949e-04  2.082300e-05 -3.796066e-04  7.715873e-04
#> visual~~speed   -4.825363e-04  3.751536e-04  9.875436e-04  2.256550e-04
#> textual~~speed   4.314192e-04  2.654701e-04  4.026555e-04  4.093776e-04
#>                   textual=~x5   textual=~x6     speed=~x7     speed=~x8
#> visual=~x1       3.141650e-04  2.546537e-04  3.049172e-04  3.435120e-04
#> visual=~x2       2.000129e-04  1.692562e-04  7.658901e-05  1.427343e-04
#> visual=~x3       2.591496e-04  2.206257e-04  5.006757e-05  4.258362e-05
#> textual=~x4      1.688801e-03  1.376938e-03  8.922238e-05  9.357838e-05
#> textual=~x5      3.930062e-03  1.578490e-03  7.437476e-05  9.582845e-05
#> textual=~x6      1.578490e-03  2.898913e-03  7.457536e-05  9.654040e-05
#> speed=~x7        7.437476e-05  7.457536e-05  5.475475e-03  1.898565e-03
#> speed=~x8        9.582845e-05  9.654040e-05  1.898565e-03  5.612839e-03
#> speed=~x9        1.034695e-04  6.539474e-05 -8.867327e-04 -1.687832e-03
#> x1~~x1           6.481860e-05  7.383460e-05 -3.332406e-04 -3.390605e-04
#> x2~~x2          -1.259862e-05 -1.328243e-05  3.157452e-05 -8.647084e-06
#> x3~~x3          -2.269572e-05 -2.738332e-05  1.527783e-04  2.064197e-04
#> x4~~x4           6.459757e-04  6.832362e-04 -5.449319e-05  5.461796e-06
#> x5~~x5          -2.080084e-03  4.801831e-04  6.677323e-05  4.678771e-05
#> x6~~x6           5.039398e-04 -1.603122e-03 -7.335699e-06 -4.579406e-05
#> x7~~x7           1.757196e-05 -5.183366e-06 -3.368940e-03 -1.767978e-03
#> x8~~x8           1.574371e-05 -3.771666e-05 -3.429949e-03 -9.323437e-03
#> x9~~x9          -2.677522e-05  2.604058e-05  3.695614e-03  5.862929e-03
#> visual~~textual  5.976293e-04  7.336587e-04  1.891355e-04  2.281799e-04
#> visual~~speed    2.858364e-04  2.187081e-04 -1.926030e-03 -2.320677e-03
#> textual~~speed   4.661669e-04  4.047452e-04 -3.246937e-04 -8.927829e-04
#>                     speed=~x9        x1~~x1        x2~~x2        x3~~x3
#> visual=~x1       6.233622e-05 -0.0122250842  1.466413e-03  3.072863e-03
#> visual=~x2       1.250035e-04  0.0054101467 -2.090859e-03 -1.610649e-03
#> visual=~x3       3.071841e-04  0.0064038183 -9.100595e-04 -3.944877e-03
#> textual=~x4      7.405772e-05 -0.0001279744  2.237214e-05  3.916836e-05
#> textual=~x5      1.034695e-04  0.0000648186 -1.259862e-05 -2.269572e-05
#> textual=~x6      6.539474e-05  0.0000738346 -1.328243e-05 -2.738332e-05
#> speed=~x7       -8.867327e-04 -0.0003332406  3.157452e-05  1.527783e-04
#> speed=~x8       -1.687832e-03 -0.0003390605 -8.647084e-06  2.064197e-04
#> speed=~x9        5.894651e-03  0.0004510226 -1.161187e-05 -2.491848e-04
#> x1~~x1           4.510226e-04  0.0463660406 -4.766128e-03 -9.986706e-03
#> x2~~x2          -1.161187e-05 -0.0047661278  8.511238e-03  1.417195e-03
#> x3~~x3          -2.491848e-04 -0.0099867063  1.417195e-03  1.281872e-02
#> x4~~x4           2.680573e-05  0.0006882199 -1.230916e-04 -2.246406e-04
#> x5~~x5          -7.597103e-05 -0.0003116688  5.786411e-05  9.502759e-05
#> x6~~x6           4.038167e-05 -0.0003723394  6.453136e-05  1.260108e-04
#> x7~~x7           2.439345e-03  0.0005089865 -4.915125e-05 -2.357085e-04
#> x8~~x8           7.513275e-03  0.0010008137  2.603809e-05 -6.179704e-04
#> x9~~x9          -7.620125e-03 -0.0010703061  2.712778e-05  5.860744e-04
#> visual~~textual  1.179829e-04 -0.0009458378  3.021593e-04  1.338676e-03
#> visual~~speed    3.691166e-03  0.0036899830 -1.057782e-05 -7.935360e-04
#> textual~~speed   1.371421e-03  0.0001321966 -2.703898e-06 -8.890268e-05
#>                        x4~~x4        x5~~x5        x6~~x6        x7~~x7
#> visual=~x1      -2.114624e-04  9.341801e-05  1.154535e-04 -1.669503e-04
#> visual=~x2       1.400470e-04 -6.713443e-05 -7.284124e-05  5.012766e-05
#> visual=~x3       1.444548e-04 -6.283285e-05 -8.025778e-05  1.437528e-04
#> textual=~x4     -1.889064e-03  5.955463e-04  6.602594e-04 -1.903188e-05
#> textual=~x5      6.459757e-04 -2.080084e-03  5.039398e-04  1.757196e-05
#> textual=~x6      6.832362e-04  4.801831e-04 -1.603122e-03 -5.183366e-06
#> speed=~x7       -5.449319e-05  6.677323e-05 -7.335699e-06 -3.368940e-03
#> speed=~x8        5.461796e-06  4.678771e-05 -4.579406e-05 -1.767978e-03
#> speed=~x9        2.680573e-05 -7.597103e-05  4.038167e-05  2.439345e-03
#> x1~~x1           6.882199e-04 -3.116688e-04 -3.723394e-04  5.089865e-04
#> x2~~x2          -1.230916e-04  5.786411e-05  6.453136e-05 -4.915125e-05
#> x3~~x3          -2.246406e-04  9.502759e-05  1.260108e-04 -2.357085e-04
#> x4~~x4           1.672689e-02 -3.182478e-03 -3.506878e-03  8.429821e-05
#> x5~~x5          -3.182478e-03  1.691445e-02 -2.479085e-03 -1.040738e-04
#> x6~~x6          -3.506878e-03 -2.479085e-03  1.491448e-02  1.148294e-05
#> x7~~x7           8.429821e-05 -1.040738e-04  1.148294e-05  1.190422e-02
#> x8~~x8          -1.761092e-05 -1.415088e-04  1.382795e-04  5.305242e-03
#> x9~~x9          -6.349724e-05  1.774102e-04 -9.433640e-05 -5.724535e-03
#> visual~~textual -2.529357e-04  1.027251e-03 -3.238691e-04 -5.033374e-05
#> visual~~speed    7.175173e-05 -9.662949e-05  1.430018e-05  3.700656e-03
#> textual~~speed   1.292443e-04  8.197794e-05 -1.581051e-05  9.250200e-04
#>                        x8~~x8        x9~~x9 visual~~textual visual~~speed
#> visual=~x1      -3.329928e-04  3.507098e-04    9.461949e-04 -4.825363e-04
#> visual=~x2      -4.342221e-05 -1.901485e-05    2.082300e-05  3.751536e-04
#> visual=~x3       3.783558e-04 -3.605765e-04   -3.796066e-04  9.875436e-04
#> textual=~x4     -8.334976e-06  2.021852e-05    7.715873e-04  2.256550e-04
#> textual=~x5      1.574371e-05 -2.677522e-05    5.976293e-04  2.858364e-04
#> textual=~x6     -3.771666e-05  2.604058e-05    7.336587e-04  2.187081e-04
#> speed=~x7       -3.429949e-03  3.695614e-03    1.891355e-04 -1.926030e-03
#> speed=~x8       -9.323437e-03  5.862929e-03    2.281799e-04 -2.320677e-03
#> speed=~x9        7.513275e-03 -7.620125e-03    1.179829e-04  3.691166e-03
#> x1~~x1           1.000814e-03 -1.070306e-03   -9.458378e-04  3.689983e-03
#> x2~~x2           2.603809e-05  2.712778e-05    3.021593e-04 -1.057782e-05
#> x3~~x3          -6.179704e-04  5.860744e-04    1.338676e-03 -7.935360e-04
#> x4~~x4          -1.761092e-05 -6.349724e-05   -2.529357e-04  7.175173e-05
#> x5~~x5          -1.415088e-04  1.774102e-04    1.027251e-03 -9.662949e-05
#> x6~~x6           1.382795e-04 -9.433640e-05   -3.238691e-04  1.430018e-05
#> x7~~x7           5.305242e-03 -5.724535e-03   -5.033374e-05  3.700656e-03
#> x8~~x8           3.472513e-02 -1.762404e-02   -1.304225e-04  8.608315e-03
#> x9~~x9          -1.762404e-02  2.454052e-02    1.179607e-04 -7.497054e-03
#> visual~~textual -1.304225e-04  1.179607e-04    6.423496e-03  7.193127e-04
#> visual~~speed    8.608315e-03 -7.497054e-03    7.193127e-04  1.177453e-02
#> textual~~speed   3.652556e-03 -2.529804e-03    1.493595e-03  3.066816e-03
#>                 textual~~speed
#> visual=~x1        4.314192e-04
#> visual=~x2        2.654701e-04
#> visual=~x3        4.026555e-04
#> textual=~x4       4.093776e-04
#> textual=~x5       4.661669e-04
#> textual=~x6       4.047452e-04
#> speed=~x7        -3.246937e-04
#> speed=~x8        -8.927829e-04
#> speed=~x9         1.371421e-03
#> x1~~x1            1.321966e-04
#> x2~~x2           -2.703898e-06
#> x3~~x3           -8.890268e-05
#> x4~~x4            1.292443e-04
#> x5~~x5            8.197794e-05
#> x6~~x6           -1.581051e-05
#> x7~~x7            9.250200e-04
#> x8~~x8            3.652556e-03
#> x9~~x9           -2.529804e-03
#> visual~~textual   1.493595e-03
#> visual~~speed     3.066816e-03
#> textual~~speed    5.945061e-03
# }