Variance-Covariance Matrix for INLAvaan Models

Extract the posterior variance-covariance matrix of model parameters from a fitted INLAvaan model.

Usage

# S4 method for class 'INLAvaan'
vcov(object, type = c("lavaan", "theta"), ...)

Arguments

object

An object of class INLAvaan.

type

Character. "lavaan" (default) returns the posterior covariance matrix of the model parameters computed from posterior samples (matching lavaan output). "theta" returns the Laplace approximation covariance in the internal parameterisation.

...

Currently unused.

Value

A square numeric matrix.

See also

Examples

# \donttest{
HS.model <- "
  visual  =~ x1 + x2 + x3
  textual =~ x4 + x5 + x6
  speed   =~ x7 + x8 + x9
"
utils::data("HolzingerSwineford1939", package = "lavaan")
fit <- acfa(HS.model, HolzingerSwineford1939, std.lv = TRUE, nsamp = 100,
            test = "none", verbose = FALSE)

# Default: posterior covariance of lavaan parameters
vcov(fit)
#>                    visual=~x1    visual=~x2    visual=~x3   textual=~x4
#> visual=~x1       7.730979e-03 -1.210827e-03 -1.572938e-03  3.959178e-04
#> visual=~x2      -1.210827e-03  5.685530e-03  1.432615e-03 -4.066016e-04
#> visual=~x3      -1.572938e-03  1.432615e-03  6.577306e-03  3.097529e-04
#> textual=~x4      3.959178e-04 -4.066016e-04  3.097529e-04  3.578224e-03
#> textual=~x5      7.698808e-04 -8.135618e-04 -3.414534e-04  1.537880e-03
#> textual=~x6      5.107557e-04 -3.855952e-04 -5.746978e-05  1.523994e-03
#> speed=~x7        5.451646e-04 -6.482155e-04  6.449880e-04 -9.814538e-05
#> speed=~x8       -2.306061e-04 -4.326022e-04 -5.090725e-04 -1.537802e-04
#> speed=~x9       -1.471399e-03  1.159452e-03  2.166409e-04 -1.751338e-04
#> x1~~x1          -7.370027e-03  2.513062e-03  4.199281e-03 -6.748830e-04
#> x2~~x2           1.160490e-03 -1.182084e-03 -4.867997e-05  9.116510e-04
#> x3~~x3           3.932905e-03 -1.434450e-03 -3.660000e-03  6.343628e-05
#> x4~~x4          -4.515145e-04 -1.035322e-04 -8.399454e-04 -1.075814e-03
#> x5~~x5           2.602009e-04 -8.180289e-05 -4.742417e-04  7.445831e-04
#> x6~~x6           1.366273e-06  1.706727e-04  1.871899e-04  2.096394e-04
#> x7~~x7          -7.657446e-04 -1.481102e-04 -8.779639e-06  2.058062e-04
#> x8~~x8           3.005712e-04  6.583909e-04 -4.784320e-04 -1.540472e-05
#> x9~~x9           4.142348e-04 -1.157744e-03  1.365538e-04  4.223790e-04
#> visual~~textual  1.850099e-03 -3.340731e-04 -1.053904e-03  1.166586e-03
#> visual~~speed   -1.801976e-05  8.041758e-04  2.300234e-04 -5.810271e-04
#> textual~~speed   7.742825e-04  5.670392e-04 -8.021047e-04  3.158271e-04
#>                   textual=~x5   textual=~x6     speed=~x7     speed=~x8
#> visual=~x1       7.698808e-04  5.107557e-04  5.451646e-04 -2.306061e-04
#> visual=~x2      -8.135618e-04 -3.855952e-04 -6.482155e-04 -4.326022e-04
#> visual=~x3      -3.414534e-04 -5.746978e-05  6.449880e-04 -5.090725e-04
#> textual=~x4      1.537880e-03  1.523994e-03 -9.814538e-05 -1.537802e-04
#> textual=~x5      4.287014e-03  1.694529e-03 -8.633042e-05  6.577335e-05
#> textual=~x6      1.694529e-03  3.121853e-03 -2.442385e-04 -1.409531e-04
#> speed=~x7       -8.633042e-05 -2.442385e-04  4.907756e-03  1.971206e-03
#> speed=~x8        6.577335e-05 -1.409531e-04  1.971206e-03  6.014213e-03
#> speed=~x9       -7.195878e-04 -3.860703e-04 -9.854936e-04 -1.111165e-03
#> x1~~x1          -1.601803e-03 -3.030788e-04 -1.610723e-04  4.993087e-04
#> x2~~x2           1.020810e-03 -2.564694e-04  4.464656e-04  4.465019e-04
#> x3~~x3          -3.504914e-04  6.515711e-05 -5.070424e-05 -4.383159e-04
#> x4~~x4          -3.907439e-04 -1.242597e-04  1.611807e-04  8.875546e-05
#> x5~~x5          -1.123352e-03  4.778455e-04  2.293539e-04  7.388625e-04
#> x6~~x6          -9.990167e-05 -4.063602e-04 -7.758219e-05 -6.101155e-05
#> x7~~x7           4.164787e-04  1.098502e-04 -1.747439e-03  7.272975e-04
#> x8~~x8          -8.348405e-05  4.569719e-04 -1.211242e-03 -3.899513e-03
#> x9~~x9           1.102490e-03  1.666266e-04  1.876103e-03  2.830815e-03
#> visual~~textual  2.800385e-04  1.378883e-03  2.080899e-04  5.754162e-04
#> visual~~speed    4.561184e-05  1.424819e-04 -1.156153e-03 -1.686712e-03
#> textual~~speed  -6.868731e-05  2.420013e-04  4.888518e-04 -7.022909e-04
#>                     speed=~x9        x1~~x1        x2~~x2        x3~~x3
#> visual=~x1      -0.0014713994 -7.370027e-03  1.160490e-03  3.932905e-03
#> visual=~x2       0.0011594524  2.513062e-03 -1.182084e-03 -1.434450e-03
#> visual=~x3       0.0002166409  4.199281e-03 -4.867997e-05 -3.660000e-03
#> textual=~x4     -0.0001751338 -6.748830e-04  9.116510e-04  6.343628e-05
#> textual=~x5     -0.0007195878 -1.601803e-03  1.020810e-03 -3.504914e-04
#> textual=~x6     -0.0003860703 -3.030788e-04 -2.564694e-04  6.515711e-05
#> speed=~x7       -0.0009854936 -1.610723e-04  4.464656e-04 -5.070424e-05
#> speed=~x8       -0.0011111649  4.993087e-04  4.465019e-04 -4.383159e-04
#> speed=~x9        0.0049364541  7.996857e-04 -4.058674e-04  1.747169e-04
#> x1~~x1           0.0007996857  1.450365e-02 -1.407847e-03 -6.524996e-03
#> x2~~x2          -0.0004058674 -1.407847e-03  6.706572e-03  1.338051e-03
#> x3~~x3           0.0001747169 -6.524996e-03  1.338051e-03  9.208474e-03
#> x4~~x4           0.0006160871 -1.168718e-04 -2.263345e-04  6.119274e-05
#> x5~~x5          -0.0001851892 -5.294567e-05  2.332161e-04  9.538343e-04
#> x6~~x6          -0.0003936430  6.445200e-05  3.951656e-04  6.516478e-05
#> x7~~x7           0.0012874969  9.484787e-05  1.339936e-04 -6.009878e-04
#> x8~~x8           0.0017855851 -8.531421e-04 -2.405966e-04  1.546080e-03
#> x9~~x9          -0.0036153434  1.766920e-04  6.679858e-04 -6.814682e-04
#> visual~~textual -0.0003517045 -1.053838e-03  4.020985e-04  1.726047e-03
#> visual~~speed    0.0014664435  1.312743e-03 -6.197156e-04  5.945135e-04
#> textual~~speed   0.0005355817 -8.544698e-04 -1.440603e-03  4.377813e-04
#>                        x4~~x4        x5~~x5        x6~~x6        x7~~x7
#> visual=~x1      -4.515145e-04  2.602009e-04  1.366273e-06 -7.657446e-04
#> visual=~x2      -1.035322e-04 -8.180289e-05  1.706727e-04 -1.481102e-04
#> visual=~x3      -8.399454e-04 -4.742417e-04  1.871899e-04 -8.779639e-06
#> textual=~x4     -1.075814e-03  7.445831e-04  2.096394e-04  2.058062e-04
#> textual=~x5     -3.907439e-04 -1.123352e-03 -9.990167e-05  4.164787e-04
#> textual=~x6     -1.242597e-04  4.778455e-04 -4.063602e-04  1.098502e-04
#> speed=~x7        1.611807e-04  2.293539e-04 -7.758219e-05 -1.747439e-03
#> speed=~x8        8.875546e-05  7.388625e-04 -6.101155e-05  7.272975e-04
#> speed=~x9        6.160871e-04 -1.851892e-04 -3.936430e-04  1.287497e-03
#> x1~~x1          -1.168718e-04 -5.294567e-05  6.445200e-05  9.484787e-05
#> x2~~x2          -2.263345e-04  2.332161e-04  3.951656e-04  1.339936e-04
#> x3~~x3           6.119274e-05  9.538343e-04  6.516478e-05 -6.009878e-04
#> x4~~x4           2.544069e-03 -7.699226e-04 -2.520784e-04  1.699750e-04
#> x5~~x5          -7.699226e-04  3.818919e-03 -2.329782e-04 -3.785871e-04
#> x6~~x6          -2.520784e-04 -2.329782e-04  1.358291e-03 -1.772477e-04
#> x7~~x7           1.699750e-04 -3.785871e-04 -1.772477e-04  6.564953e-03
#> x8~~x8           3.455788e-04 -5.801890e-04 -7.807258e-05 -4.381961e-04
#> x9~~x9          -3.013384e-04  3.628175e-04  5.523400e-04 -1.195774e-03
#> visual~~textual -3.061372e-04  1.268480e-03  1.457872e-04 -1.315854e-04
#> visual~~speed    7.296036e-04 -6.349725e-04 -3.310994e-04  1.140398e-03
#> textual~~speed   1.353946e-04  2.154841e-04 -9.314654e-05 -1.727597e-05
#>                        x8~~x8        x9~~x9 visual~~textual visual~~speed
#> visual=~x1       3.005712e-04  0.0004142348    0.0018500988 -1.801976e-05
#> visual=~x2       6.583909e-04 -0.0011577443   -0.0003340731  8.041758e-04
#> visual=~x3      -4.784320e-04  0.0001365538   -0.0010539044  2.300234e-04
#> textual=~x4     -1.540472e-05  0.0004223790    0.0011665856 -5.810271e-04
#> textual=~x5     -8.348405e-05  0.0011024897    0.0002800385  4.561184e-05
#> textual=~x6      4.569719e-04  0.0001666266    0.0013788831  1.424819e-04
#> speed=~x7       -1.211242e-03  0.0018761030    0.0002080899 -1.156153e-03
#> speed=~x8       -3.899513e-03  0.0028308153    0.0005754162 -1.686712e-03
#> speed=~x9        1.785585e-03 -0.0036153434   -0.0003517045  1.466443e-03
#> x1~~x1          -8.531421e-04  0.0001766920   -0.0010538383  1.312743e-03
#> x2~~x2          -2.405966e-04  0.0006679858    0.0004020985 -6.197156e-04
#> x3~~x3           1.546080e-03 -0.0006814682    0.0017260473  5.945135e-04
#> x4~~x4           3.455788e-04 -0.0003013384   -0.0003061372  7.296036e-04
#> x5~~x5          -5.801890e-04  0.0003628175    0.0012684800 -6.349725e-04
#> x6~~x6          -7.807258e-05  0.0005523400    0.0001457872 -3.310994e-04
#> x7~~x7          -4.381961e-04 -0.0011957740   -0.0001315854  1.140398e-03
#> x8~~x8           6.062015e-03 -0.0025598763   -0.0001442953  2.074722e-03
#> x9~~x9          -2.559876e-03  0.0062097174    0.0002551743 -1.974808e-03
#> visual~~textual -1.442953e-04  0.0002551743    0.0051609361  3.036312e-04
#> visual~~speed    2.074722e-03 -0.0019748084    0.0003036312  6.726224e-03
#> textual~~speed   1.424981e-03 -0.0008703119    0.0012011138  2.095166e-03
#>                 textual~~speed
#> visual=~x1        7.742825e-04
#> visual=~x2        5.670392e-04
#> visual=~x3       -8.021047e-04
#> textual=~x4       3.158271e-04
#> textual=~x5      -6.868731e-05
#> textual=~x6       2.420013e-04
#> speed=~x7         4.888518e-04
#> speed=~x8        -7.022909e-04
#> speed=~x9         5.355817e-04
#> x1~~x1           -8.544698e-04
#> x2~~x2           -1.440603e-03
#> x3~~x3            4.377813e-04
#> x4~~x4            1.353946e-04
#> x5~~x5            2.154841e-04
#> x6~~x6           -9.314654e-05
#> x7~~x7           -1.727597e-05
#> x8~~x8            1.424981e-03
#> x9~~x9           -8.703119e-04
#> visual~~textual   1.201114e-03
#> visual~~speed     2.095166e-03
#> textual~~speed    5.015880e-03

# Internal parameterisation (Laplace approximation)
vcov(fit, type = "theta")
#>                    visual=~x1    visual=~x2    visual=~x3   textual=~x4
#> visual=~x1       6.898587e-03 -9.165857e-04 -9.867557e-04  3.367678e-04
#> visual=~x2      -9.165857e-04  6.525261e-03  1.566651e-03  1.401828e-04
#> visual=~x3      -9.867557e-04  1.566651e-03  6.021448e-03  1.929055e-04
#> textual=~x4      3.367678e-04  1.401828e-04  1.929055e-04  3.202281e-03
#> textual=~x5      3.121193e-04  1.987118e-04  2.574554e-04  1.677661e-03
#> textual=~x6      2.529953e-04  1.681537e-04  2.191811e-04  1.367829e-03
#> speed=~x7        3.029206e-04  7.609836e-05  4.973399e-05  8.864062e-05
#> speed=~x8        3.412604e-04  1.418067e-04  4.229667e-05  9.296407e-05
#> speed=~x9        6.195091e-05  1.241973e-04  3.052028e-04  7.359271e-05
#> x1~~x1          -1.216302e-02  5.382504e-03  6.370959e-03 -1.273305e-04
#> x2~~x2           1.458995e-03 -2.080412e-03 -9.054328e-04  2.226225e-05
#> x3~~x3           3.057214e-03 -1.602436e-03 -3.925080e-03  3.897942e-05
#> x4~~x4          -2.104044e-04  1.393482e-04  1.437300e-04 -1.879696e-03
#> x5~~x5           9.295670e-05 -6.679914e-05 -6.251812e-05  5.925565e-04
#> x6~~x6           1.148743e-04 -7.247747e-05 -7.985547e-05  6.570178e-04
#> x7~~x7          -1.661245e-04  4.988021e-05  1.430742e-04 -1.892993e-05
#> x8~~x8          -3.313343e-04 -4.319711e-05  3.765540e-04 -8.262366e-06
#> x9~~x9           3.489718e-04 -1.892216e-05 -3.588609e-04  2.009704e-05
#> visual~~textual  9.415274e-04  2.075824e-05 -3.777095e-04  7.678324e-04
#> visual~~speed   -4.800760e-04  3.732852e-04  9.826828e-04  2.245913e-04
#> textual~~speed   4.293306e-04  2.641969e-04  4.007179e-04  4.074328e-04
#>                   textual=~x5   textual=~x6     speed=~x7     speed=~x8
#> visual=~x1       3.121193e-04  2.529953e-04  3.029206e-04  3.412604e-04
#> visual=~x2       1.987118e-04  1.681537e-04  7.609836e-05  1.418067e-04
#> visual=~x3       2.574554e-04  2.191811e-04  4.973399e-05  4.229667e-05
#> textual=~x4      1.677661e-03  1.367829e-03  8.864062e-05  9.296407e-05
#> textual=~x5      3.904073e-03  1.568076e-03  7.389059e-05  9.519861e-05
#> textual=~x6      1.568076e-03  2.879717e-03  7.409026e-05  9.590733e-05
#> speed=~x7        7.389059e-05  7.409026e-05  5.439411e-03  1.886296e-03
#> speed=~x8        9.519861e-05  9.590733e-05  1.886296e-03  5.575902e-03
#> speed=~x9        1.028171e-04  6.498526e-05 -8.811822e-04 -1.676962e-03
#> x1~~x1           6.449054e-05  7.346268e-05 -3.316229e-04 -3.374218e-04
#> x2~~x2          -1.253501e-05 -1.321498e-05  3.141020e-05 -8.613488e-06
#> x3~~x3          -2.257528e-05 -2.723894e-05  1.520326e-04  2.054104e-04
#> x4~~x4           6.427250e-04  6.798769e-04 -5.422684e-05  5.431733e-06
#> x5~~x5          -2.069644e-03  4.778096e-04  6.645149e-05  4.656548e-05
#> x6~~x6           5.014459e-04 -1.595198e-03 -7.304155e-06 -4.557413e-05
#> x7~~x7           1.749729e-05 -5.150650e-06 -3.352527e-03 -1.759687e-03
#> x8~~x8           1.570784e-05 -3.750155e-05 -3.413783e-03 -9.277591e-03
#> x9~~x9          -2.667457e-05  2.589312e-05  3.678129e-03  5.834693e-03
#> visual~~textual  5.947602e-04  7.300952e-04  1.882035e-04  2.270434e-04
#> visual~~speed    2.844870e-04  2.176762e-04 -1.916930e-03 -2.309643e-03
#> textual~~speed   4.639551e-04  4.028189e-04 -3.232058e-04 -8.884590e-04
#>                     speed=~x9        x1~~x1        x2~~x2        x3~~x3
#> visual=~x1       6.195091e-05 -1.216302e-02  1.458995e-03  3.057214e-03
#> visual=~x2       1.241973e-04  5.382504e-03 -2.080412e-03 -1.602436e-03
#> visual=~x3       3.052028e-04  6.370959e-03 -9.054328e-04 -3.925080e-03
#> textual=~x4      7.359271e-05 -1.273305e-04  2.226225e-05  3.897942e-05
#> textual=~x5      1.028171e-04  6.449054e-05 -1.253501e-05 -2.257528e-05
#> textual=~x6      6.498526e-05  7.346268e-05 -1.321498e-05 -2.723894e-05
#> speed=~x7       -8.811822e-04 -3.316229e-04  3.141020e-05  1.520326e-04
#> speed=~x8       -1.676962e-03 -3.374218e-04 -8.613488e-06  2.054104e-04
#> speed=~x9        5.855960e-03  4.488534e-04 -1.154329e-05 -2.479650e-04
#> x1~~x1           4.488534e-04  4.620665e-02 -4.749730e-03 -9.951987e-03
#> x2~~x2          -1.154329e-05 -4.749730e-03  8.482971e-03  1.412339e-03
#> x3~~x3          -2.479650e-04 -9.951987e-03  1.412339e-03  1.277592e-02
#> x4~~x4           2.667909e-05  6.858895e-04 -1.226832e-04 -2.238859e-04
#> x5~~x5          -7.560511e-05 -3.106132e-04  5.767489e-05  9.471622e-05
#> x6~~x6           4.018847e-05 -3.710817e-04  6.431582e-05  1.255854e-04
#> x7~~x7           2.427752e-03  5.073830e-04 -4.897727e-05 -2.349523e-04
#> x8~~x8           7.476746e-03  9.976715e-04  2.597746e-05 -6.159681e-04
#> x9~~x9          -7.583367e-03 -1.066922e-03  2.701198e-05  5.841872e-04
#> visual~~textual  1.174709e-04 -9.425100e-04  3.011801e-04  1.334329e-03
#> visual~~speed    3.673533e-03  3.677837e-03 -1.046462e-05 -7.908254e-04
#> textual~~speed   1.364862e-03  1.318368e-04 -2.688307e-06 -8.861014e-05
#>                        x4~~x4        x5~~x5        x6~~x6        x7~~x7
#> visual=~x1      -2.104044e-04  9.295670e-05  1.148743e-04 -1.661245e-04
#> visual=~x2       1.393482e-04 -6.679914e-05 -7.247747e-05  4.988021e-05
#> visual=~x3       1.437300e-04 -6.251812e-05 -7.985547e-05  1.430742e-04
#> textual=~x4     -1.879696e-03  5.925565e-04  6.570178e-04 -1.892993e-05
#> textual=~x5      6.427250e-04 -2.069644e-03  5.014459e-04  1.749729e-05
#> textual=~x6      6.798769e-04  4.778096e-04 -1.595198e-03 -5.150650e-06
#> speed=~x7       -5.422684e-05  6.645149e-05 -7.304155e-06 -3.352527e-03
#> speed=~x8        5.431733e-06  4.656548e-05 -4.557413e-05 -1.759687e-03
#> speed=~x9        2.667909e-05 -7.560511e-05  4.018847e-05  2.427752e-03
#> x1~~x1           6.858895e-04 -3.106132e-04 -3.710817e-04  5.073830e-04
#> x2~~x2          -1.226832e-04  5.767489e-05  6.431582e-05 -4.897727e-05
#> x3~~x3          -2.238859e-04  9.471622e-05  1.255854e-04 -2.349523e-04
#> x4~~x4           1.667134e-02 -3.171764e-03 -3.495430e-03  8.402710e-05
#> x5~~x5          -3.171764e-03  1.685817e-02 -2.470971e-03 -1.037440e-04
#> x6~~x6          -3.495430e-03 -2.470971e-03  1.486517e-02  1.145233e-05
#> x7~~x7           8.402710e-05 -1.037440e-04  1.145233e-05  1.186519e-02
#> x8~~x8          -1.753871e-05 -1.410673e-04  1.378402e-04  5.289076e-03
#> x9~~x9          -6.330128e-05  1.768555e-04 -9.404482e-05 -5.706960e-03
#> visual~~textual -2.520747e-04  1.023875e-03 -3.228116e-04 -5.013071e-05
#> visual~~speed    7.152447e-05 -9.631759e-05  1.425938e-05  3.689237e-03
#> textual~~speed   1.288246e-04  8.172103e-05 -1.575283e-05  9.222448e-04
#>                        x8~~x8        x9~~x9 visual~~textual visual~~speed
#> visual=~x1      -3.313343e-04  3.489718e-04    9.415274e-04 -4.800760e-04
#> visual=~x2      -4.319711e-05 -1.892216e-05    2.075824e-05  3.732852e-04
#> visual=~x3       3.765540e-04 -3.588609e-04   -3.777095e-04  9.826828e-04
#> textual=~x4     -8.262366e-06  2.009704e-05    7.678324e-04  2.245913e-04
#> textual=~x5      1.570784e-05 -2.667457e-05    5.947602e-04  2.844870e-04
#> textual=~x6     -3.750155e-05  2.589312e-05    7.300952e-04  2.176762e-04
#> speed=~x7       -3.413783e-03  3.678129e-03    1.882035e-04 -1.916930e-03
#> speed=~x8       -9.277591e-03  5.834693e-03    2.270434e-04 -2.309643e-03
#> speed=~x9        7.476746e-03 -7.583367e-03    1.174709e-04  3.673533e-03
#> x1~~x1           9.976715e-04 -1.066922e-03   -9.425100e-04  3.677837e-03
#> x2~~x2           2.597746e-05  2.701198e-05    3.011801e-04 -1.046462e-05
#> x3~~x3          -6.159681e-04  5.841872e-04    1.334329e-03 -7.908254e-04
#> x4~~x4          -1.753871e-05 -6.330128e-05   -2.520747e-04  7.152447e-05
#> x5~~x5          -1.410673e-04  1.768555e-04    1.023875e-03 -9.631759e-05
#> x6~~x6           1.378402e-04 -9.404482e-05   -3.228116e-04  1.425938e-05
#> x7~~x7           5.289076e-03 -5.706960e-03   -5.013071e-05  3.689237e-03
#> x8~~x8           3.461087e-02 -1.756801e-02   -1.298721e-04  8.581362e-03
#> x9~~x9          -1.756801e-02  2.446224e-02    1.174843e-04 -7.473895e-03
#> visual~~textual -1.298721e-04  1.174843e-04    6.402515e-03  7.171685e-04
#> visual~~speed    8.581362e-03 -7.473895e-03    7.171685e-04  1.173705e-02
#> textual~~speed   3.640917e-03 -2.521933e-03    1.488903e-03  3.057236e-03
#>                 textual~~speed
#> visual=~x1        4.293306e-04
#> visual=~x2        2.641969e-04
#> visual=~x3        4.007179e-04
#> textual=~x4       4.074328e-04
#> textual=~x5       4.639551e-04
#> textual=~x6       4.028189e-04
#> speed=~x7        -3.232058e-04
#> speed=~x8        -8.884590e-04
#> speed=~x9         1.364862e-03
#> x1~~x1            1.318368e-04
#> x2~~x2           -2.688307e-06
#> x3~~x3           -8.861014e-05
#> x4~~x4            1.288246e-04
#> x5~~x5            8.172103e-05
#> x6~~x6           -1.575283e-05
#> x7~~x7            9.222448e-04
#> x8~~x8            3.640917e-03
#> x9~~x9           -2.521933e-03
#> visual~~textual   1.488903e-03
#> visual~~speed     3.057236e-03
#> textual~~speed    5.925664e-03
# }