Variance-Covariance Matrix for INLAvaan Models

Extract the posterior variance-covariance matrix of model parameters from a fitted INLAvaan model.

Usage

# S4 method for class 'INLAvaan'
vcov(object, type = c("lavaan", "theta"), ...)

Arguments

object

An object of class INLAvaan.

type

Character. "lavaan" (default) returns the posterior covariance matrix of the model parameters computed from posterior samples (matching lavaan output). "theta" returns the Laplace approximation covariance in the internal parameterisation.

...

Currently unused.

Value

A square numeric matrix.

See also

Examples

# \donttest{
HS.model <- "
  visual  =~ x1 + x2 + x3
  textual =~ x4 + x5 + x6
  speed   =~ x7 + x8 + x9
"
utils::data("HolzingerSwineford1939", package = "lavaan")
fit <- acfa(HS.model, HolzingerSwineford1939, std.lv = TRUE, nsamp = 100,
            test = "none", verbose = FALSE)

# Default: posterior covariance of lavaan parameters
vcov(fit)
#>                    visual=~x1    visual=~x2    visual=~x3   textual=~x4
#> visual=~x1       7.781833e-03 -0.0012190458 -1.583619e-03  3.985165e-04
#> visual=~x2      -1.219046e-03  0.0057236108  1.442308e-03 -4.093758e-04
#> visual=~x3      -1.583619e-03  0.0014423078  6.621301e-03  3.117909e-04
#> textual=~x4      3.985165e-04 -0.0004093758  3.117909e-04  3.602287e-03
#> textual=~x5      7.749926e-04 -0.0008190596 -3.437699e-04  1.548133e-03
#> textual=~x6      5.140835e-04 -0.0003881927 -5.790088e-05  1.534206e-03
#> speed=~x7        5.487502e-04 -0.0006525638  6.492678e-04 -9.881894e-05
#> speed=~x8       -2.321742e-04 -0.0004354281 -5.124352e-04 -1.547929e-04
#> speed=~x9       -1.481067e-03  0.0011670926  2.180839e-04 -1.762960e-04
#> x1~~x1          -7.430748e-03  0.0025340213  4.234365e-03 -6.803656e-04
#> x2~~x2           1.170339e-03 -0.0011920324 -4.913649e-05  9.192299e-04
#> x3~~x3           3.966022e-03 -0.0014467321 -3.690973e-03  6.394508e-05
#> x4~~x4          -4.553126e-04 -0.0001044055 -8.470450e-04 -1.084856e-03
#> x5~~x5           2.624353e-04 -0.0000824760 -4.783137e-04  7.509146e-04
#> x6~~x6           1.363764e-06  0.0001721388  1.887792e-04  2.113588e-04
#> x7~~x7          -7.720919e-04 -0.0001493206 -8.988923e-06  2.074975e-04
#> x8~~x8           3.031503e-04  0.0006637434 -4.823313e-04 -1.541852e-05
#> x9~~x9           4.174976e-04 -0.0011672809  1.377676e-04  4.259554e-04
#> visual~~textual  1.859212e-03 -0.0003357834 -1.059155e-03  1.172306e-03
#> visual~~speed   -1.810588e-05  0.0008080911  2.311530e-04 -5.839758e-04
#> textual~~speed   7.779842e-04  0.0005697667 -8.060865e-04  3.173399e-04
#>                   textual=~x5   textual=~x6     speed=~x7     speed=~x8
#> visual=~x1       7.749926e-04  5.140835e-04  5.487502e-04 -2.321742e-04
#> visual=~x2      -8.190596e-04 -3.881927e-04 -6.525638e-04 -4.354281e-04
#> visual=~x3      -3.437699e-04 -5.790088e-05  6.492678e-04 -5.124352e-04
#> textual=~x4      1.548133e-03  1.534206e-03 -9.881894e-05 -1.547929e-04
#> textual=~x5      4.315878e-03  1.705860e-03 -8.687288e-05  6.623548e-05
#> textual=~x6      1.705860e-03  3.142877e-03 -2.458882e-04 -1.418621e-04
#> speed=~x7       -8.687288e-05 -2.458882e-04  4.940466e-03  1.983978e-03
#> speed=~x8        6.623548e-05 -1.418621e-04  1.983978e-03  6.053239e-03
#> speed=~x9       -7.243867e-04 -3.886670e-04 -9.917298e-04 -1.118106e-03
#> x1~~x1          -1.615078e-03 -3.055945e-04 -1.623791e-04  5.033320e-04
#> x2~~x2           1.029358e-03 -2.586598e-04  4.502692e-04  4.503039e-04
#> x3~~x3          -3.534635e-04  6.567817e-05 -5.107925e-05 -4.419945e-04
#> x4~~x4          -3.940191e-04 -1.252844e-04  1.625627e-04  8.949528e-05
#> x5~~x5          -1.132893e-03  4.819026e-04  2.313069e-04  7.449729e-04
#> x6~~x6          -1.007896e-04 -4.097875e-04 -7.824058e-05 -6.147410e-05
#> x7~~x7           4.201290e-04  1.108317e-04 -1.761941e-03  7.338067e-04
#> x8~~x8          -8.412920e-05  4.608147e-04 -1.220918e-03 -3.931161e-03
#> x9~~x9           1.111706e-03  1.680382e-04  1.891139e-03  2.853474e-03
#> visual~~textual  2.813474e-04  1.385672e-03  2.091023e-04  5.782056e-04
#> visual~~speed    4.581880e-05  1.432067e-04 -1.161539e-03 -1.694610e-03
#> textual~~speed  -6.911801e-05  2.431292e-04  4.913987e-04 -7.056785e-04
#>                     speed=~x9        x1~~x1        x2~~x2        x3~~x3
#> visual=~x1      -0.0014810674 -7.430748e-03  1.170339e-03  3.966022e-03
#> visual=~x2       0.0011670926  2.534021e-03 -1.192032e-03 -1.446732e-03
#> visual=~x3       0.0002180839  4.234365e-03 -4.913649e-05 -3.690973e-03
#> textual=~x4     -0.0001762960 -6.803656e-04  9.192299e-04  6.394508e-05
#> textual=~x5     -0.0007243867 -1.615078e-03  1.029358e-03 -3.534635e-04
#> textual=~x6     -0.0003886670 -3.055945e-04 -2.586598e-04  6.567817e-05
#> speed=~x7       -0.0009917298 -1.623791e-04  4.502692e-04 -5.107925e-05
#> speed=~x8       -0.0011181057  5.033320e-04  4.503039e-04 -4.419945e-04
#> speed=~x9        0.0049685296  8.062709e-04 -4.092676e-04  1.760427e-04
#> x1~~x1           0.0008062709  1.464658e-02 -1.421970e-03 -6.590798e-03
#> x2~~x2          -0.0004092676 -1.421970e-03  6.773897e-03  1.351619e-03
#> x3~~x3           0.0001760427 -6.590798e-03  1.351619e-03  9.302123e-03
#> x4~~x4           0.0006212335 -1.179478e-04 -2.286106e-04  6.187063e-05
#> x5~~x5          -0.0001866213 -5.363526e-05  2.355809e-04  9.635811e-04
#> x6~~x6          -0.0003969551  6.517110e-05  3.991473e-04  6.579847e-05
#> x7~~x7           0.0012979629  9.557717e-05  1.353373e-04 -6.073210e-04
#> x8~~x8           0.0017998741 -8.616074e-04 -2.430287e-04  1.561604e-03
#> x9~~x9          -0.0036444658  1.786399e-04  6.746541e-04 -6.882409e-04
#> visual~~textual -0.0003534812 -1.060825e-03  4.047887e-04  1.737646e-03
#> visual~~speed    0.0014732247  1.321052e-03 -6.238864e-04  5.983634e-04
#> textual~~speed   0.0005381056 -8.601195e-04 -1.450190e-03  4.406074e-04
#>                        x4~~x4        x5~~x5        x6~~x6        x7~~x7
#> visual=~x1      -4.553126e-04  2.624353e-04  1.363764e-06 -7.720919e-04
#> visual=~x2      -1.044055e-04 -8.247600e-05  1.721388e-04 -1.493206e-04
#> visual=~x3      -8.470450e-04 -4.783137e-04  1.887792e-04 -8.988923e-06
#> textual=~x4     -1.084856e-03  7.509146e-04  2.113588e-04  2.074975e-04
#> textual=~x5     -3.940191e-04 -1.132893e-03 -1.007896e-04  4.201290e-04
#> textual=~x6     -1.252844e-04  4.819026e-04 -4.097875e-04  1.108317e-04
#> speed=~x7        1.625627e-04  2.313069e-04 -7.824058e-05 -1.761941e-03
#> speed=~x8        8.949528e-05  7.449729e-04 -6.147410e-05  7.338067e-04
#> speed=~x9        6.212335e-04 -1.866213e-04 -3.969551e-04  1.297963e-03
#> x1~~x1          -1.179478e-04 -5.363526e-05  6.517110e-05  9.557717e-05
#> x2~~x2          -2.286106e-04  2.355809e-04  3.991473e-04  1.353373e-04
#> x3~~x3           6.187063e-05  9.635811e-04  6.579847e-05 -6.073210e-04
#> x4~~x4           2.569942e-03 -7.777697e-04 -2.546055e-04  1.717307e-04
#> x5~~x5          -7.777697e-04  3.857894e-03 -2.353696e-04 -3.823669e-04
#> x6~~x6          -2.546055e-04 -2.353696e-04  1.372173e-03 -1.790541e-04
#> x7~~x7           1.717307e-04 -3.823669e-04 -1.790541e-04  6.631861e-03
#> x8~~x8           3.489794e-04 -5.858720e-04 -7.884117e-05 -4.429488e-04
#> x9~~x9          -3.043731e-04  3.662394e-04  5.578531e-04 -1.207234e-03
#> visual~~textual -3.081167e-04  1.276894e-03  1.468275e-04 -1.325055e-04
#> visual~~speed    7.345627e-04 -6.391509e-04 -3.333134e-04  1.147778e-03
#> textual~~speed   1.363393e-04  2.169732e-04 -9.377230e-05 -1.743051e-05
#>                        x8~~x8        x9~~x9 visual~~textual visual~~speed
#> visual=~x1       3.031503e-04  0.0004174976    0.0018592116 -1.810588e-05
#> visual=~x2       6.637434e-04 -0.0011672809   -0.0003357834  8.080911e-04
#> visual=~x3      -4.823313e-04  0.0001377676   -0.0010591549  2.311530e-04
#> textual=~x4     -1.541852e-05  0.0004259554    0.0011723062 -5.839758e-04
#> textual=~x5     -8.412920e-05  0.0011117065    0.0002813474  4.581880e-05
#> textual=~x6      4.608147e-04  0.0001680382    0.0013856724  1.432067e-04
#> speed=~x7       -1.220918e-03  0.0018911386    0.0002091023 -1.161539e-03
#> speed=~x8       -3.931161e-03  0.0028534737    0.0005782056 -1.694610e-03
#> speed=~x9        1.799874e-03 -0.0036444658   -0.0003534812  1.473225e-03
#> x1~~x1          -8.616074e-04  0.0001786399   -0.0010608247  1.321052e-03
#> x2~~x2          -2.430287e-04  0.0006746541    0.0004047887 -6.238864e-04
#> x3~~x3           1.561604e-03 -0.0006882409    0.0017376463  5.983634e-04
#> x4~~x4           3.489794e-04 -0.0003043731   -0.0003081167  7.345627e-04
#> x5~~x5          -5.858720e-04  0.0003662394    0.0012768945 -6.391509e-04
#> x6~~x6          -7.884117e-05  0.0005578531    0.0001468275 -3.333134e-04
#> x7~~x7          -4.429488e-04 -0.0012072336   -0.0001325055  1.147778e-03
#> x8~~x8           6.121337e-03 -0.0025846420   -0.0001451754  2.087996e-03
#> x9~~x9          -2.584642e-03  0.0062700503    0.0002568864 -1.987168e-03
#> visual~~textual -1.451754e-04  0.0002568864    0.0051775818  3.045503e-04
#> visual~~speed    2.087996e-03 -0.0019871683    0.0003045503  6.747516e-03
#> textual~~speed   1.434294e-03 -0.0008758963    0.0012048339  2.101753e-03
#>                 textual~~speed
#> visual=~x1        7.779842e-04
#> visual=~x2        5.697667e-04
#> visual=~x3       -8.060865e-04
#> textual=~x4       3.173399e-04
#> textual=~x5      -6.911801e-05
#> textual=~x6       2.431292e-04
#> speed=~x7         4.913987e-04
#> speed=~x8        -7.056785e-04
#> speed=~x9         5.381056e-04
#> x1~~x1           -8.601195e-04
#> x2~~x2           -1.450190e-03
#> x3~~x3            4.406074e-04
#> x4~~x4            1.363393e-04
#> x5~~x5            2.169732e-04
#> x6~~x6           -9.377230e-05
#> x7~~x7           -1.743051e-05
#> x8~~x8            1.434294e-03
#> x9~~x9           -8.758963e-04
#> visual~~textual   1.204834e-03
#> visual~~speed     2.101753e-03
#> textual~~speed    5.032031e-03

# Internal parameterisation (Laplace approximation)
vcov(fit, type = "theta")
#>                    visual=~x1    visual=~x2    visual=~x3   textual=~x4
#> visual=~x1       6.944819e-03 -9.228843e-04 -9.935751e-04  3.389857e-04
#> visual=~x2      -9.228843e-04  6.568863e-03  1.577215e-03  1.410972e-04
#> visual=~x3      -9.935751e-04  1.577215e-03  6.061736e-03  1.941726e-04
#> textual=~x4      3.389857e-04  1.410972e-04  1.941726e-04  3.223624e-03
#> textual=~x5      3.141650e-04  2.000129e-04  2.591496e-04  1.688801e-03
#> textual=~x6      2.546537e-04  1.692562e-04  2.206257e-04  1.376938e-03
#> speed=~x7        3.049172e-04  7.658901e-05  5.006757e-05  8.922238e-05
#> speed=~x8        3.435120e-04  1.427343e-04  4.258362e-05  9.357838e-05
#> speed=~x9        6.233622e-05  1.250035e-04  3.071841e-04  7.405772e-05
#> x1~~x1          -1.222508e-02  5.410147e-03  6.403818e-03 -1.279744e-04
#> x2~~x2           1.466413e-03 -2.090859e-03 -9.100595e-04  2.237214e-05
#> x3~~x3           3.072863e-03 -1.610649e-03 -3.944877e-03  3.916836e-05
#> x4~~x4          -2.114624e-04  1.400470e-04  1.444548e-04 -1.889064e-03
#> x5~~x5           9.341801e-05 -6.713443e-05 -6.283285e-05  5.955463e-04
#> x6~~x6           1.154535e-04 -7.284124e-05 -8.025778e-05  6.602594e-04
#> x7~~x7          -1.669503e-04  5.012766e-05  1.437528e-04 -1.903188e-05
#> x8~~x8          -3.329928e-04 -4.342221e-05  3.783558e-04 -8.334976e-06
#> x9~~x9           3.507098e-04 -1.901485e-05 -3.605765e-04  2.021852e-05
#> visual~~textual  9.461949e-04  2.082300e-05 -3.796066e-04  7.715873e-04
#> visual~~speed   -4.825363e-04  3.751536e-04  9.875436e-04  2.256550e-04
#> textual~~speed   4.314192e-04  2.654701e-04  4.026555e-04  4.093776e-04
#>                   textual=~x5   textual=~x6     speed=~x7     speed=~x8
#> visual=~x1       3.141650e-04  2.546537e-04  3.049172e-04  3.435120e-04
#> visual=~x2       2.000129e-04  1.692562e-04  7.658901e-05  1.427343e-04
#> visual=~x3       2.591496e-04  2.206257e-04  5.006757e-05  4.258362e-05
#> textual=~x4      1.688801e-03  1.376938e-03  8.922238e-05  9.357838e-05
#> textual=~x5      3.930062e-03  1.578490e-03  7.437476e-05  9.582845e-05
#> textual=~x6      1.578490e-03  2.898913e-03  7.457536e-05  9.654040e-05
#> speed=~x7        7.437476e-05  7.457536e-05  5.475475e-03  1.898565e-03
#> speed=~x8        9.582845e-05  9.654040e-05  1.898565e-03  5.612839e-03
#> speed=~x9        1.034695e-04  6.539474e-05 -8.867327e-04 -1.687832e-03
#> x1~~x1           6.481860e-05  7.383460e-05 -3.332406e-04 -3.390605e-04
#> x2~~x2          -1.259862e-05 -1.328243e-05  3.157452e-05 -8.647084e-06
#> x3~~x3          -2.269572e-05 -2.738332e-05  1.527783e-04  2.064197e-04
#> x4~~x4           6.459757e-04  6.832362e-04 -5.449319e-05  5.461796e-06
#> x5~~x5          -2.080084e-03  4.801831e-04  6.677323e-05  4.678771e-05
#> x6~~x6           5.039398e-04 -1.603122e-03 -7.335699e-06 -4.579406e-05
#> x7~~x7           1.757196e-05 -5.183366e-06 -3.368940e-03 -1.767978e-03
#> x8~~x8           1.574371e-05 -3.771666e-05 -3.429949e-03 -9.323437e-03
#> x9~~x9          -2.677522e-05  2.604058e-05  3.695614e-03  5.862929e-03
#> visual~~textual  5.976293e-04  7.336587e-04  1.891355e-04  2.281799e-04
#> visual~~speed    2.858364e-04  2.187081e-04 -1.926030e-03 -2.320677e-03
#> textual~~speed   4.661669e-04  4.047452e-04 -3.246937e-04 -8.927829e-04
#>                     speed=~x9        x1~~x1        x2~~x2        x3~~x3
#> visual=~x1       6.233622e-05 -0.0122250842  1.466413e-03  3.072863e-03
#> visual=~x2       1.250035e-04  0.0054101467 -2.090859e-03 -1.610649e-03
#> visual=~x3       3.071841e-04  0.0064038183 -9.100595e-04 -3.944877e-03
#> textual=~x4      7.405772e-05 -0.0001279744  2.237214e-05  3.916836e-05
#> textual=~x5      1.034695e-04  0.0000648186 -1.259862e-05 -2.269572e-05
#> textual=~x6      6.539474e-05  0.0000738346 -1.328243e-05 -2.738332e-05
#> speed=~x7       -8.867327e-04 -0.0003332406  3.157452e-05  1.527783e-04
#> speed=~x8       -1.687832e-03 -0.0003390605 -8.647084e-06  2.064197e-04
#> speed=~x9        5.894651e-03  0.0004510226 -1.161187e-05 -2.491848e-04
#> x1~~x1           4.510226e-04  0.0463660406 -4.766128e-03 -9.986706e-03
#> x2~~x2          -1.161187e-05 -0.0047661278  8.511238e-03  1.417195e-03
#> x3~~x3          -2.491848e-04 -0.0099867063  1.417195e-03  1.281872e-02
#> x4~~x4           2.680573e-05  0.0006882199 -1.230916e-04 -2.246406e-04
#> x5~~x5          -7.597103e-05 -0.0003116688  5.786411e-05  9.502759e-05
#> x6~~x6           4.038167e-05 -0.0003723394  6.453136e-05  1.260108e-04
#> x7~~x7           2.439345e-03  0.0005089865 -4.915125e-05 -2.357085e-04
#> x8~~x8           7.513275e-03  0.0010008137  2.603809e-05 -6.179704e-04
#> x9~~x9          -7.620125e-03 -0.0010703061  2.712778e-05  5.860744e-04
#> visual~~textual  1.179829e-04 -0.0009458378  3.021593e-04  1.338676e-03
#> visual~~speed    3.691166e-03  0.0036899830 -1.057782e-05 -7.935360e-04
#> textual~~speed   1.371421e-03  0.0001321966 -2.703898e-06 -8.890268e-05
#>                        x4~~x4        x5~~x5        x6~~x6        x7~~x7
#> visual=~x1      -2.114624e-04  9.341801e-05  1.154535e-04 -1.669503e-04
#> visual=~x2       1.400470e-04 -6.713443e-05 -7.284124e-05  5.012766e-05
#> visual=~x3       1.444548e-04 -6.283285e-05 -8.025778e-05  1.437528e-04
#> textual=~x4     -1.889064e-03  5.955463e-04  6.602594e-04 -1.903188e-05
#> textual=~x5      6.459757e-04 -2.080084e-03  5.039398e-04  1.757196e-05
#> textual=~x6      6.832362e-04  4.801831e-04 -1.603122e-03 -5.183366e-06
#> speed=~x7       -5.449319e-05  6.677323e-05 -7.335699e-06 -3.368940e-03
#> speed=~x8        5.461796e-06  4.678771e-05 -4.579406e-05 -1.767978e-03
#> speed=~x9        2.680573e-05 -7.597103e-05  4.038167e-05  2.439345e-03
#> x1~~x1           6.882199e-04 -3.116688e-04 -3.723394e-04  5.089865e-04
#> x2~~x2          -1.230916e-04  5.786411e-05  6.453136e-05 -4.915125e-05
#> x3~~x3          -2.246406e-04  9.502759e-05  1.260108e-04 -2.357085e-04
#> x4~~x4           1.672689e-02 -3.182478e-03 -3.506878e-03  8.429821e-05
#> x5~~x5          -3.182478e-03  1.691445e-02 -2.479085e-03 -1.040738e-04
#> x6~~x6          -3.506878e-03 -2.479085e-03  1.491448e-02  1.148294e-05
#> x7~~x7           8.429821e-05 -1.040738e-04  1.148294e-05  1.190422e-02
#> x8~~x8          -1.761092e-05 -1.415088e-04  1.382795e-04  5.305242e-03
#> x9~~x9          -6.349724e-05  1.774102e-04 -9.433640e-05 -5.724535e-03
#> visual~~textual -2.529357e-04  1.027251e-03 -3.238691e-04 -5.033374e-05
#> visual~~speed    7.175173e-05 -9.662949e-05  1.430018e-05  3.700656e-03
#> textual~~speed   1.292443e-04  8.197794e-05 -1.581051e-05  9.250200e-04
#>                        x8~~x8        x9~~x9 visual~~textual visual~~speed
#> visual=~x1      -3.329928e-04  3.507098e-04    9.461949e-04 -4.825363e-04
#> visual=~x2      -4.342221e-05 -1.901485e-05    2.082300e-05  3.751536e-04
#> visual=~x3       3.783558e-04 -3.605765e-04   -3.796066e-04  9.875436e-04
#> textual=~x4     -8.334976e-06  2.021852e-05    7.715873e-04  2.256550e-04
#> textual=~x5      1.574371e-05 -2.677522e-05    5.976293e-04  2.858364e-04
#> textual=~x6     -3.771666e-05  2.604058e-05    7.336587e-04  2.187081e-04
#> speed=~x7       -3.429949e-03  3.695614e-03    1.891355e-04 -1.926030e-03
#> speed=~x8       -9.323437e-03  5.862929e-03    2.281799e-04 -2.320677e-03
#> speed=~x9        7.513275e-03 -7.620125e-03    1.179829e-04  3.691166e-03
#> x1~~x1           1.000814e-03 -1.070306e-03   -9.458378e-04  3.689983e-03
#> x2~~x2           2.603809e-05  2.712778e-05    3.021593e-04 -1.057782e-05
#> x3~~x3          -6.179704e-04  5.860744e-04    1.338676e-03 -7.935360e-04
#> x4~~x4          -1.761092e-05 -6.349724e-05   -2.529357e-04  7.175173e-05
#> x5~~x5          -1.415088e-04  1.774102e-04    1.027251e-03 -9.662949e-05
#> x6~~x6           1.382795e-04 -9.433640e-05   -3.238691e-04  1.430018e-05
#> x7~~x7           5.305242e-03 -5.724535e-03   -5.033374e-05  3.700656e-03
#> x8~~x8           3.472513e-02 -1.762404e-02   -1.304225e-04  8.608315e-03
#> x9~~x9          -1.762404e-02  2.454052e-02    1.179607e-04 -7.497054e-03
#> visual~~textual -1.304225e-04  1.179607e-04    6.423496e-03  7.193127e-04
#> visual~~speed    8.608315e-03 -7.497054e-03    7.193127e-04  1.177453e-02
#> textual~~speed   3.652556e-03 -2.529804e-03    1.493595e-03  3.066816e-03
#>                 textual~~speed
#> visual=~x1        4.314192e-04
#> visual=~x2        2.654701e-04
#> visual=~x3        4.026555e-04
#> textual=~x4       4.093776e-04
#> textual=~x5       4.661669e-04
#> textual=~x6       4.047452e-04
#> speed=~x7        -3.246937e-04
#> speed=~x8        -8.927829e-04
#> speed=~x9         1.371421e-03
#> x1~~x1            1.321966e-04
#> x2~~x2           -2.703898e-06
#> x3~~x3           -8.890268e-05
#> x4~~x4            1.292443e-04
#> x5~~x5            8.197794e-05
#> x6~~x6           -1.581051e-05
#> x7~~x7            9.250200e-04
#> x8~~x8            3.652556e-03
#> x9~~x9           -2.529804e-03
#> visual~~textual   1.493595e-03
#> visual~~speed     3.066816e-03
#> textual~~speed    5.945061e-03
# }