Variance-Covariance Matrix for INLAvaan Models

Extract the posterior variance-covariance matrix of model parameters from a fitted INLAvaan model.

Usage

# S4 method for class 'INLAvaan'
vcov(object, type = c("lavaan", "theta"), ...)

Arguments

object

An object of class INLAvaan.

type

Character. "lavaan" (default) returns the posterior covariance matrix of the model parameters computed from posterior samples (matching lavaan output). "theta" returns the Laplace approximation covariance in the internal parameterisation.

...

Currently unused.

Value

A square numeric matrix.

See also

Examples

# \donttest{
HS.model <- "
  visual  =~ x1 + x2 + x3
  textual =~ x4 + x5 + x6
  speed   =~ x7 + x8 + x9
"
utils::data("HolzingerSwineford1939", package = "lavaan")
fit <- acfa(HS.model, HolzingerSwineford1939, std.lv = TRUE, nsamp = 100,
            test = "none", verbose = FALSE)

# Default: posterior covariance of lavaan parameters
vcov(fit)
#>                    visual=~x1    visual=~x2    visual=~x3   textual=~x4
#> visual=~x1       6.965480e-03 -1.053211e-03 -9.739920e-04 -7.548480e-04
#> visual=~x2      -1.053211e-03  6.643275e-03  5.801053e-04  6.251310e-04
#> visual=~x3      -9.739920e-04  5.801053e-04  6.218651e-03  4.276893e-04
#> textual=~x4     -7.548480e-04  6.251310e-04  4.276893e-04  3.064786e-03
#> textual=~x5      3.848361e-04 -3.425974e-05  2.312434e-04  1.599415e-03
#> textual=~x6      2.816021e-04  6.074249e-05 -4.429022e-04  1.174946e-03
#> speed=~x7        7.615021e-05  4.520427e-04  5.434709e-04 -3.447962e-05
#> speed=~x8       -4.135667e-04  8.751267e-04  1.793287e-04  3.047613e-04
#> speed=~x9        1.231131e-04 -2.765372e-04  3.889592e-04 -3.703309e-04
#> x1~~x1          -6.128565e-03  2.439584e-03  3.071845e-03  6.043191e-04
#> x2~~x2          -9.276332e-06 -2.491017e-03 -1.138774e-03 -1.077178e-04
#> x3~~x3           1.723105e-03 -7.539655e-04 -2.730014e-03  6.727045e-04
#> x4~~x4           7.344802e-04 -2.966369e-04  1.017349e-04 -8.816756e-04
#> x5~~x5          -6.664794e-04 -1.794921e-04 -8.232605e-04  2.304067e-04
#> x6~~x6          -1.196548e-04 -1.936701e-04  3.985415e-04  2.249395e-04
#> x7~~x7           1.274253e-03 -1.394397e-03 -1.177083e-03  8.006498e-04
#> x8~~x8          -1.714120e-04 -6.563845e-04 -3.208312e-05 -4.827837e-04
#> x9~~x9          -1.963221e-05 -1.328356e-04 -1.337082e-04  3.922018e-04
#> visual~~textual  8.379894e-05  8.805578e-04  4.885620e-04  7.099473e-04
#> visual~~speed   -8.503821e-05 -3.743196e-04  4.449085e-04  1.442988e-04
#> textual~~speed   1.586194e-03 -5.656384e-04  5.237844e-04 -2.087731e-04
#>                   textual=~x5   textual=~x6     speed=~x7     speed=~x8
#> visual=~x1       3.848361e-04  2.816021e-04  7.615021e-05 -4.135667e-04
#> visual=~x2      -3.425974e-05  6.074249e-05  4.520427e-04  8.751267e-04
#> visual=~x3       2.312434e-04 -4.429022e-04  5.434709e-04  1.793287e-04
#> textual=~x4      1.599415e-03  1.174946e-03 -3.447962e-05  3.047613e-04
#> textual=~x5      4.021477e-03  1.787017e-03 -1.518208e-04 -3.329526e-04
#> textual=~x6      1.787017e-03  2.862064e-03 -3.444078e-04  2.273374e-04
#> speed=~x7       -1.518208e-04 -3.444078e-04  3.872567e-03  8.560095e-04
#> speed=~x8       -3.329526e-04  2.273374e-04  8.560095e-04  5.235306e-03
#> speed=~x9        2.946106e-04 -1.695491e-04 -1.339716e-03 -1.204094e-03
#> x1~~x1          -1.158473e-04 -6.374732e-04  4.228389e-04  1.170883e-03
#> x2~~x2          -6.528451e-04 -2.744712e-04  1.061047e-03  1.083807e-03
#> x3~~x3           1.612620e-04  8.646768e-04  1.446280e-04 -8.088096e-06
#> x4~~x4           2.819811e-04  5.068531e-05 -3.346237e-04 -3.005470e-04
#> x5~~x5          -1.447992e-03 -5.724272e-05 -1.385024e-04 -2.074298e-04
#> x6~~x6           2.640471e-04 -3.698803e-04  3.581982e-05 -4.687669e-05
#> x7~~x7           1.011631e-03  1.266519e-03 -2.488080e-03 -6.533148e-04
#> x8~~x8          -7.102622e-06 -8.556083e-05 -9.936834e-04 -4.624762e-03
#> x9~~x9          -1.322671e-04 -5.083200e-04  1.573160e-03  3.212899e-03
#> visual~~textual  6.664573e-04  7.034400e-04 -2.746237e-04  5.781443e-05
#> visual~~speed    9.571418e-04 -1.259194e-04 -1.339906e-03 -1.486093e-03
#> textual~~speed   8.503295e-04 -2.115341e-04  2.793326e-04 -1.178085e-03
#>                     speed=~x9        x1~~x1        x2~~x2        x3~~x3
#> visual=~x1       0.0001231131 -6.128565e-03 -9.276332e-06  1.723105e-03
#> visual=~x2      -0.0002765372  2.439584e-03 -2.491017e-03 -7.539655e-04
#> visual=~x3       0.0003889592  3.071845e-03 -1.138774e-03 -2.730014e-03
#> textual=~x4     -0.0003703309  6.043191e-04 -1.077178e-04  6.727045e-04
#> textual=~x5      0.0002946106 -1.158473e-04 -6.528451e-04  1.612620e-04
#> textual=~x6     -0.0001695491 -6.374732e-04 -2.744712e-04  8.646768e-04
#> speed=~x7       -0.0013397160  4.228389e-04  1.061047e-03  1.446280e-04
#> speed=~x8       -0.0012040943  1.170883e-03  1.083807e-03 -8.088096e-06
#> speed=~x9        0.0064148186  6.830871e-04 -3.319254e-04 -4.560229e-04
#> x1~~x1           0.0006830871  1.135865e-02 -6.889841e-04 -3.295326e-03
#> x2~~x2          -0.0003319254 -6.889841e-04  1.051956e-02  6.123437e-04
#> x3~~x3          -0.0004560229 -3.295326e-03  6.123437e-04  7.854527e-03
#> x4~~x4           0.0003670583 -3.060223e-04 -3.697022e-04 -4.327578e-04
#> x5~~x5          -0.0004492831 -4.824343e-04  7.788900e-04  4.759178e-04
#> x6~~x6           0.0001721900  2.712865e-04  3.768415e-04 -8.373952e-05
#> x7~~x7           0.0015917092 -1.910929e-03 -8.861542e-04  5.449710e-04
#> x8~~x8           0.0029944044  3.158231e-05 -8.754393e-04 -1.031633e-03
#> x9~~x9          -0.0049677084  6.685631e-04  1.419557e-03  1.326757e-04
#> visual~~textual -0.0004935794  2.877661e-04 -8.116299e-04  4.425181e-04
#> visual~~speed    0.0036123700  1.360359e-03 -3.189309e-04 -3.059583e-04
#> textual~~speed   0.0021857030 -1.021095e-03 -1.330345e-03  4.217379e-04
#>                        x4~~x4        x5~~x5        x6~~x6        x7~~x7
#> visual=~x1       7.344802e-04 -6.664794e-04 -1.196548e-04  1.274253e-03
#> visual=~x2      -2.966369e-04 -1.794921e-04 -1.936701e-04 -1.394397e-03
#> visual=~x3       1.017349e-04 -8.232605e-04  3.985415e-04 -1.177083e-03
#> textual=~x4     -8.816756e-04  2.304067e-04  2.249395e-04  8.006498e-04
#> textual=~x5      2.819811e-04 -1.447992e-03  2.640471e-04  1.011631e-03
#> textual=~x6      5.068531e-05 -5.724272e-05 -3.698803e-04  1.266519e-03
#> speed=~x7       -3.346237e-04 -1.385024e-04  3.581982e-05 -2.488080e-03
#> speed=~x8       -3.005470e-04 -2.074298e-04 -4.687669e-05 -6.533148e-04
#> speed=~x9        3.670583e-04 -4.492831e-04  1.721900e-04  1.591709e-03
#> x1~~x1          -3.060223e-04 -4.824343e-04  2.712865e-04 -1.910929e-03
#> x2~~x2          -3.697022e-04  7.788900e-04  3.768415e-04 -8.861542e-04
#> x3~~x3          -4.327578e-04  4.759178e-04 -8.373952e-05  5.449710e-04
#> x4~~x4           2.375102e-03 -1.077325e-03 -2.167743e-04  1.455982e-04
#> x5~~x5          -1.077325e-03  2.953508e-03 -1.621703e-04  3.043455e-04
#> x6~~x6          -2.167743e-04 -1.621703e-04  1.687120e-03 -2.849917e-05
#> x7~~x7           1.455982e-04  3.043455e-04 -2.849917e-05  6.484860e-03
#> x8~~x8           3.129631e-04  1.131800e-04 -8.795691e-05  1.278021e-03
#> x9~~x9          -3.646645e-04  7.556379e-05 -4.587928e-05 -1.937355e-03
#> visual~~textual -1.994089e-04 -2.659316e-05 -2.748185e-04 -5.982523e-04
#> visual~~speed    2.288388e-04 -1.328725e-03  2.240824e-04  2.133337e-03
#> textual~~speed   7.612651e-04 -1.000046e-03  9.025408e-05  1.018369e-03
#>                        x8~~x8        x9~~x9 visual~~textual visual~~speed
#> visual=~x1      -1.714120e-04 -1.963221e-05    8.379894e-05 -8.503821e-05
#> visual=~x2      -6.563845e-04 -1.328356e-04    8.805578e-04 -3.743196e-04
#> visual=~x3      -3.208312e-05 -1.337082e-04    4.885620e-04  4.449085e-04
#> textual=~x4     -4.827837e-04  3.922018e-04    7.099473e-04  1.442988e-04
#> textual=~x5     -7.102622e-06 -1.322671e-04    6.664573e-04  9.571418e-04
#> textual=~x6     -8.556083e-05 -5.083200e-04    7.034400e-04 -1.259194e-04
#> speed=~x7       -9.936834e-04  1.573160e-03   -2.746237e-04 -1.339906e-03
#> speed=~x8       -4.624762e-03  3.212899e-03    5.781443e-05 -1.486093e-03
#> speed=~x9        2.994404e-03 -4.967708e-03   -4.935794e-04  3.612370e-03
#> x1~~x1           3.158231e-05  6.685631e-04    2.877661e-04  1.360359e-03
#> x2~~x2          -8.754393e-04  1.419557e-03   -8.116299e-04 -3.189309e-04
#> x3~~x3          -1.031633e-03  1.326757e-04    4.425181e-04 -3.059583e-04
#> x4~~x4           3.129631e-04 -3.646645e-04   -1.994089e-04  2.288388e-04
#> x5~~x5           1.131800e-04  7.556379e-05   -2.659316e-05 -1.328725e-03
#> x6~~x6          -8.795691e-05 -4.587928e-05   -2.748185e-04  2.240824e-04
#> x7~~x7           1.278021e-03 -1.937355e-03   -5.982523e-04  2.133337e-03
#> x8~~x8           8.364568e-03 -4.568105e-03   -1.020938e-03  3.128770e-03
#> x9~~x9          -4.568105e-03  8.119290e-03    1.905771e-04 -3.667700e-03
#> visual~~textual -1.020938e-03  1.905771e-04    4.854447e-03  1.344096e-04
#> visual~~speed    3.128770e-03 -3.667700e-03    1.344096e-04  7.137050e-03
#> textual~~speed   2.686100e-03 -2.457090e-03    8.856628e-05  2.983955e-03
#>                 textual~~speed
#> visual=~x1        1.586194e-03
#> visual=~x2       -5.656384e-04
#> visual=~x3        5.237844e-04
#> textual=~x4      -2.087731e-04
#> textual=~x5       8.503295e-04
#> textual=~x6      -2.115341e-04
#> speed=~x7         2.793326e-04
#> speed=~x8        -1.178085e-03
#> speed=~x9         2.185703e-03
#> x1~~x1           -1.021095e-03
#> x2~~x2           -1.330345e-03
#> x3~~x3            4.217379e-04
#> x4~~x4            7.612651e-04
#> x5~~x5           -1.000046e-03
#> x6~~x6            9.025408e-05
#> x7~~x7            1.018369e-03
#> x8~~x8            2.686100e-03
#> x9~~x9           -2.457090e-03
#> visual~~textual   8.856628e-05
#> visual~~speed     2.983955e-03
#> textual~~speed    5.338562e-03

# Internal parameterisation (Laplace approximation)
vcov(fit, type = "theta")
#>                    visual=~x1    visual=~x2    visual=~x3   textual=~x4
#> visual=~x1       6.898587e-03 -9.165857e-04 -9.867557e-04  3.367678e-04
#> visual=~x2      -9.165857e-04  6.525261e-03  1.566651e-03  1.401828e-04
#> visual=~x3      -9.867557e-04  1.566651e-03  6.021448e-03  1.929055e-04
#> textual=~x4      3.367678e-04  1.401828e-04  1.929055e-04  3.202281e-03
#> textual=~x5      3.121193e-04  1.987118e-04  2.574554e-04  1.677661e-03
#> textual=~x6      2.529953e-04  1.681537e-04  2.191811e-04  1.367829e-03
#> speed=~x7        3.029206e-04  7.609836e-05  4.973399e-05  8.864062e-05
#> speed=~x8        3.412604e-04  1.418067e-04  4.229667e-05  9.296407e-05
#> speed=~x9        6.195091e-05  1.241973e-04  3.052028e-04  7.359271e-05
#> x1~~x1          -1.216302e-02  5.382504e-03  6.370959e-03 -1.273305e-04
#> x2~~x2           1.458995e-03 -2.080412e-03 -9.054328e-04  2.226225e-05
#> x3~~x3           3.057214e-03 -1.602436e-03 -3.925080e-03  3.897942e-05
#> x4~~x4          -2.104044e-04  1.393482e-04  1.437300e-04 -1.879696e-03
#> x5~~x5           9.295670e-05 -6.679914e-05 -6.251812e-05  5.925565e-04
#> x6~~x6           1.148743e-04 -7.247747e-05 -7.985547e-05  6.570178e-04
#> x7~~x7          -1.661245e-04  4.988021e-05  1.430742e-04 -1.892993e-05
#> x8~~x8          -3.313343e-04 -4.319711e-05  3.765540e-04 -8.262366e-06
#> x9~~x9           3.489718e-04 -1.892216e-05 -3.588609e-04  2.009704e-05
#> visual~~textual  9.415274e-04  2.075824e-05 -3.777095e-04  7.678324e-04
#> visual~~speed   -4.800760e-04  3.732852e-04  9.826828e-04  2.245913e-04
#> textual~~speed   4.293306e-04  2.641969e-04  4.007179e-04  4.074328e-04
#>                   textual=~x5   textual=~x6     speed=~x7     speed=~x8
#> visual=~x1       3.121193e-04  2.529953e-04  3.029206e-04  3.412604e-04
#> visual=~x2       1.987118e-04  1.681537e-04  7.609836e-05  1.418067e-04
#> visual=~x3       2.574554e-04  2.191811e-04  4.973399e-05  4.229667e-05
#> textual=~x4      1.677661e-03  1.367829e-03  8.864062e-05  9.296407e-05
#> textual=~x5      3.904073e-03  1.568076e-03  7.389059e-05  9.519861e-05
#> textual=~x6      1.568076e-03  2.879717e-03  7.409026e-05  9.590733e-05
#> speed=~x7        7.389059e-05  7.409026e-05  5.439411e-03  1.886296e-03
#> speed=~x8        9.519861e-05  9.590733e-05  1.886296e-03  5.575902e-03
#> speed=~x9        1.028171e-04  6.498526e-05 -8.811822e-04 -1.676962e-03
#> x1~~x1           6.449054e-05  7.346268e-05 -3.316229e-04 -3.374218e-04
#> x2~~x2          -1.253501e-05 -1.321498e-05  3.141020e-05 -8.613488e-06
#> x3~~x3          -2.257528e-05 -2.723894e-05  1.520326e-04  2.054104e-04
#> x4~~x4           6.427250e-04  6.798769e-04 -5.422684e-05  5.431733e-06
#> x5~~x5          -2.069644e-03  4.778096e-04  6.645149e-05  4.656548e-05
#> x6~~x6           5.014459e-04 -1.595198e-03 -7.304155e-06 -4.557413e-05
#> x7~~x7           1.749729e-05 -5.150650e-06 -3.352527e-03 -1.759687e-03
#> x8~~x8           1.570784e-05 -3.750155e-05 -3.413783e-03 -9.277591e-03
#> x9~~x9          -2.667457e-05  2.589312e-05  3.678129e-03  5.834693e-03
#> visual~~textual  5.947602e-04  7.300952e-04  1.882035e-04  2.270434e-04
#> visual~~speed    2.844870e-04  2.176762e-04 -1.916930e-03 -2.309643e-03
#> textual~~speed   4.639551e-04  4.028189e-04 -3.232058e-04 -8.884590e-04
#>                     speed=~x9        x1~~x1        x2~~x2        x3~~x3
#> visual=~x1       6.195091e-05 -1.216302e-02  1.458995e-03  3.057214e-03
#> visual=~x2       1.241973e-04  5.382504e-03 -2.080412e-03 -1.602436e-03
#> visual=~x3       3.052028e-04  6.370959e-03 -9.054328e-04 -3.925080e-03
#> textual=~x4      7.359271e-05 -1.273305e-04  2.226225e-05  3.897942e-05
#> textual=~x5      1.028171e-04  6.449054e-05 -1.253501e-05 -2.257528e-05
#> textual=~x6      6.498526e-05  7.346268e-05 -1.321498e-05 -2.723894e-05
#> speed=~x7       -8.811822e-04 -3.316229e-04  3.141020e-05  1.520326e-04
#> speed=~x8       -1.676962e-03 -3.374218e-04 -8.613488e-06  2.054104e-04
#> speed=~x9        5.855960e-03  4.488534e-04 -1.154329e-05 -2.479650e-04
#> x1~~x1           4.488534e-04  4.620665e-02 -4.749730e-03 -9.951987e-03
#> x2~~x2          -1.154329e-05 -4.749730e-03  8.482971e-03  1.412339e-03
#> x3~~x3          -2.479650e-04 -9.951987e-03  1.412339e-03  1.277592e-02
#> x4~~x4           2.667909e-05  6.858895e-04 -1.226832e-04 -2.238859e-04
#> x5~~x5          -7.560511e-05 -3.106132e-04  5.767489e-05  9.471622e-05
#> x6~~x6           4.018847e-05 -3.710817e-04  6.431582e-05  1.255854e-04
#> x7~~x7           2.427752e-03  5.073830e-04 -4.897727e-05 -2.349523e-04
#> x8~~x8           7.476746e-03  9.976715e-04  2.597746e-05 -6.159681e-04
#> x9~~x9          -7.583367e-03 -1.066922e-03  2.701198e-05  5.841872e-04
#> visual~~textual  1.174709e-04 -9.425100e-04  3.011801e-04  1.334329e-03
#> visual~~speed    3.673533e-03  3.677837e-03 -1.046462e-05 -7.908254e-04
#> textual~~speed   1.364862e-03  1.318368e-04 -2.688307e-06 -8.861014e-05
#>                        x4~~x4        x5~~x5        x6~~x6        x7~~x7
#> visual=~x1      -2.104044e-04  9.295670e-05  1.148743e-04 -1.661245e-04
#> visual=~x2       1.393482e-04 -6.679914e-05 -7.247747e-05  4.988021e-05
#> visual=~x3       1.437300e-04 -6.251812e-05 -7.985547e-05  1.430742e-04
#> textual=~x4     -1.879696e-03  5.925565e-04  6.570178e-04 -1.892993e-05
#> textual=~x5      6.427250e-04 -2.069644e-03  5.014459e-04  1.749729e-05
#> textual=~x6      6.798769e-04  4.778096e-04 -1.595198e-03 -5.150650e-06
#> speed=~x7       -5.422684e-05  6.645149e-05 -7.304155e-06 -3.352527e-03
#> speed=~x8        5.431733e-06  4.656548e-05 -4.557413e-05 -1.759687e-03
#> speed=~x9        2.667909e-05 -7.560511e-05  4.018847e-05  2.427752e-03
#> x1~~x1           6.858895e-04 -3.106132e-04 -3.710817e-04  5.073830e-04
#> x2~~x2          -1.226832e-04  5.767489e-05  6.431582e-05 -4.897727e-05
#> x3~~x3          -2.238859e-04  9.471622e-05  1.255854e-04 -2.349523e-04
#> x4~~x4           1.667134e-02 -3.171764e-03 -3.495430e-03  8.402710e-05
#> x5~~x5          -3.171764e-03  1.685817e-02 -2.470971e-03 -1.037440e-04
#> x6~~x6          -3.495430e-03 -2.470971e-03  1.486517e-02  1.145233e-05
#> x7~~x7           8.402710e-05 -1.037440e-04  1.145233e-05  1.186519e-02
#> x8~~x8          -1.753871e-05 -1.410673e-04  1.378402e-04  5.289076e-03
#> x9~~x9          -6.330128e-05  1.768555e-04 -9.404482e-05 -5.706960e-03
#> visual~~textual -2.520747e-04  1.023875e-03 -3.228116e-04 -5.013071e-05
#> visual~~speed    7.152447e-05 -9.631759e-05  1.425938e-05  3.689237e-03
#> textual~~speed   1.288246e-04  8.172103e-05 -1.575283e-05  9.222448e-04
#>                        x8~~x8        x9~~x9 visual~~textual visual~~speed
#> visual=~x1      -3.313343e-04  3.489718e-04    9.415274e-04 -4.800760e-04
#> visual=~x2      -4.319711e-05 -1.892216e-05    2.075824e-05  3.732852e-04
#> visual=~x3       3.765540e-04 -3.588609e-04   -3.777095e-04  9.826828e-04
#> textual=~x4     -8.262366e-06  2.009704e-05    7.678324e-04  2.245913e-04
#> textual=~x5      1.570784e-05 -2.667457e-05    5.947602e-04  2.844870e-04
#> textual=~x6     -3.750155e-05  2.589312e-05    7.300952e-04  2.176762e-04
#> speed=~x7       -3.413783e-03  3.678129e-03    1.882035e-04 -1.916930e-03
#> speed=~x8       -9.277591e-03  5.834693e-03    2.270434e-04 -2.309643e-03
#> speed=~x9        7.476746e-03 -7.583367e-03    1.174709e-04  3.673533e-03
#> x1~~x1           9.976715e-04 -1.066922e-03   -9.425100e-04  3.677837e-03
#> x2~~x2           2.597746e-05  2.701198e-05    3.011801e-04 -1.046462e-05
#> x3~~x3          -6.159681e-04  5.841872e-04    1.334329e-03 -7.908254e-04
#> x4~~x4          -1.753871e-05 -6.330128e-05   -2.520747e-04  7.152447e-05
#> x5~~x5          -1.410673e-04  1.768555e-04    1.023875e-03 -9.631759e-05
#> x6~~x6           1.378402e-04 -9.404482e-05   -3.228116e-04  1.425938e-05
#> x7~~x7           5.289076e-03 -5.706960e-03   -5.013071e-05  3.689237e-03
#> x8~~x8           3.461087e-02 -1.756801e-02   -1.298722e-04  8.581362e-03
#> x9~~x9          -1.756801e-02  2.446224e-02    1.174843e-04 -7.473895e-03
#> visual~~textual -1.298722e-04  1.174843e-04    6.402515e-03  7.171685e-04
#> visual~~speed    8.581362e-03 -7.473895e-03    7.171685e-04  1.173705e-02
#> textual~~speed   3.640917e-03 -2.521933e-03    1.488903e-03  3.057236e-03
#>                 textual~~speed
#> visual=~x1        4.293306e-04
#> visual=~x2        2.641969e-04
#> visual=~x3        4.007179e-04
#> textual=~x4       4.074328e-04
#> textual=~x5       4.639551e-04
#> textual=~x6       4.028189e-04
#> speed=~x7        -3.232058e-04
#> speed=~x8        -8.884590e-04
#> speed=~x9         1.364862e-03
#> x1~~x1            1.318368e-04
#> x2~~x2           -2.688307e-06
#> x3~~x3           -8.861014e-05
#> x4~~x4            1.288246e-04
#> x5~~x5            8.172103e-05
#> x6~~x6           -1.575283e-05
#> x7~~x7            9.222448e-04
#> x8~~x8            3.640917e-03
#> x9~~x9           -2.521933e-03
#> visual~~textual   1.488903e-03
#> visual~~speed     3.057236e-03
#> textual~~speed    5.925664e-03
# }