Variance-Covariance Matrix for INLAvaan Models

Extract the posterior variance-covariance matrix of model parameters from a fitted INLAvaan model.

Usage

# S4 method for class 'INLAvaan'
vcov(object, type = c("lavaan", "theta"), ...)

Arguments

object

An object of class INLAvaan.

type

Character. "lavaan" (default) returns the posterior covariance matrix of the model parameters computed from posterior samples (matching lavaan output). "theta" returns the Laplace approximation covariance in the internal parameterisation.

...

Currently unused.

Value

A square numeric matrix.

See also

Examples

# \donttest{
HS.model <- "
  visual  =~ x1 + x2 + x3
  textual =~ x4 + x5 + x6
  speed   =~ x7 + x8 + x9
"
utils::data("HolzingerSwineford1939", package = "lavaan")
fit <- acfa(HS.model, HolzingerSwineford1939, std.lv = TRUE, nsamp = 100,
            test = "none", verbose = FALSE)

# Default: posterior covariance of lavaan parameters
vcov(fit)
#>                    visual=~x1    visual=~x2    visual=~x3   textual=~x4
#> visual=~x1       5.768236e-03 -2.180747e-04 -9.682749e-05  3.862199e-04
#> visual=~x2      -2.180747e-04  6.242256e-03  1.724179e-03 -4.592140e-05
#> visual=~x3      -9.682749e-05  1.724179e-03  6.375007e-03 -3.338855e-04
#> textual=~x4      3.862199e-04 -4.592140e-05 -3.338855e-04  3.477944e-03
#> textual=~x5      6.018279e-04 -2.176292e-04 -2.770488e-04  1.517196e-03
#> textual=~x6      7.605568e-04 -5.382841e-04  1.954332e-04  2.033928e-03
#> speed=~x7        6.128507e-04  1.701263e-03  6.126915e-04 -7.641123e-06
#> speed=~x8       -2.076148e-04  7.669074e-04  1.374674e-03  9.176464e-05
#> speed=~x9        6.767962e-04  5.338867e-05  3.851823e-04 -5.010991e-04
#> x1~~x1          -5.868375e-03  2.624702e-03  2.975937e-03 -3.578932e-04
#> x2~~x2           1.468701e-03 -3.077622e-03 -1.683170e-03 -4.363985e-04
#> x3~~x3           1.291399e-03 -1.738057e-03 -3.137793e-03 -6.303240e-05
#> x4~~x4           4.242329e-06 -1.983314e-05  5.631563e-04 -7.200984e-04
#> x5~~x5           1.757457e-05 -5.840046e-04 -2.001436e-04  2.858206e-04
#> x6~~x6          -2.013611e-06  2.471410e-04 -2.498341e-04  2.438519e-04
#> x7~~x7          -6.200972e-05 -1.199210e-03  8.821869e-05 -1.114422e-03
#> x8~~x8           3.412162e-04 -9.899260e-04 -4.981966e-04 -3.496915e-04
#> x9~~x9          -1.169325e-04 -1.756443e-04 -7.484516e-05  2.374742e-04
#> visual~~textual  5.838731e-04  2.527877e-04 -1.637993e-04  1.907565e-04
#> visual~~speed    2.175891e-04 -6.447806e-04 -1.611589e-04 -7.618614e-04
#> textual~~speed   4.140355e-04 -2.448115e-04 -8.726660e-05  6.013306e-05
#>                   textual=~x5   textual=~x6     speed=~x7     speed=~x8
#> visual=~x1       6.018279e-04  7.605568e-04  6.128507e-04 -2.076148e-04
#> visual=~x2      -2.176292e-04 -5.382841e-04  1.701263e-03  7.669074e-04
#> visual=~x3      -2.770488e-04  1.954332e-04  6.126915e-04  1.374674e-03
#> textual=~x4      1.517196e-03  2.033928e-03 -7.641123e-06  9.176464e-05
#> textual=~x5      4.044699e-03  2.005954e-03  2.100590e-04 -2.847953e-05
#> textual=~x6      2.005954e-03  3.430091e-03  2.490615e-04 -1.223002e-04
#> speed=~x7        2.100590e-04  2.490615e-04  5.197611e-03  7.832762e-04
#> speed=~x8       -2.847953e-05 -1.223002e-04  7.832762e-04  5.715435e-03
#> speed=~x9        5.328572e-04 -2.767663e-04 -4.331903e-04 -1.014387e-03
#> x1~~x1          -6.988515e-04 -3.617586e-04 -5.097893e-04  8.344561e-05
#> x2~~x2           2.446325e-04  2.435424e-04 -7.134969e-04  6.936151e-04
#> x3~~x3           1.776364e-04 -1.798778e-04  1.401799e-05 -1.225701e-03
#> x4~~x4           7.638350e-04  1.959890e-04  2.689462e-04  5.219829e-05
#> x5~~x5          -8.365361e-04  5.557985e-04  2.175529e-04 -2.639875e-04
#> x6~~x6          -3.388541e-04 -5.325474e-04 -3.853165e-04 -7.193402e-04
#> x7~~x7          -2.140826e-04 -3.287862e-04 -3.615443e-03 -1.847806e-03
#> x8~~x8          -6.833667e-04 -8.104702e-05 -4.649879e-04 -3.908064e-03
#> x9~~x9          -7.464858e-05  3.043284e-04  5.481364e-04  2.104002e-03
#> visual~~textual  4.136994e-04  1.876363e-04  2.160835e-04 -9.493487e-05
#> visual~~speed   -3.092461e-04 -4.145462e-04 -2.691227e-04 -1.401959e-03
#> textual~~speed   7.714795e-05 -5.904730e-04  4.188721e-04 -9.036676e-06
#>                     speed=~x9        x1~~x1        x2~~x2        x3~~x3
#> visual=~x1       6.767962e-04 -5.868375e-03  0.0014687010  1.291399e-03
#> visual=~x2       5.338867e-05  2.624702e-03 -0.0030776221 -1.738057e-03
#> visual=~x3       3.851823e-04  2.975937e-03 -0.0016831697 -3.137793e-03
#> textual=~x4     -5.010991e-04 -3.578932e-04 -0.0004363985 -6.303240e-05
#> textual=~x5      5.328572e-04 -6.988515e-04  0.0002446325  1.776364e-04
#> textual=~x6     -2.767663e-04 -3.617586e-04  0.0002435424 -1.798778e-04
#> speed=~x7       -4.331903e-04 -5.097893e-04 -0.0007134969  1.401799e-05
#> speed=~x8       -1.014387e-03  8.344561e-05  0.0006936151 -1.225701e-03
#> speed=~x9        5.441419e-03 -1.123255e-04 -0.0006032298 -2.793881e-04
#> x1~~x1          -1.123255e-04  1.290715e-02 -0.0027019129 -4.040227e-03
#> x2~~x2          -6.032298e-04 -2.701913e-03  0.0117433876  2.013394e-03
#> x3~~x3          -2.793881e-04 -4.040227e-03  0.0020133937  9.207058e-03
#> x4~~x4          -8.031481e-06  2.112193e-04 -0.0009034051 -2.045253e-04
#> x5~~x5          -7.478416e-05 -3.029082e-04  0.0003681102  5.733665e-04
#> x6~~x6           3.560314e-04  2.321858e-04  0.0006491090  4.200302e-04
#> x7~~x7           1.523557e-03  4.493494e-04  0.0009648970  9.188894e-04
#> x8~~x8           2.862057e-03  4.071767e-04 -0.0012805886  5.484940e-04
#> x9~~x9          -4.012708e-03 -1.081404e-03  0.0010987051  6.011874e-04
#> visual~~textual  2.058843e-04 -6.676085e-05  0.0006765625  1.853486e-04
#> visual~~speed    2.536981e-03  2.590018e-04  0.0004323332  5.133066e-04
#> textual~~speed   1.588119e-03 -5.526231e-04  0.0002582848  3.344160e-04
#>                        x4~~x4        x5~~x5        x6~~x6        x7~~x7
#> visual=~x1       4.242329e-06  1.757457e-05 -2.013611e-06 -6.200972e-05
#> visual=~x2      -1.983314e-05 -5.840046e-04  2.471410e-04 -1.199210e-03
#> visual=~x3       5.631563e-04 -2.001436e-04 -2.498341e-04  8.821869e-05
#> textual=~x4     -7.200984e-04  2.858206e-04  2.438519e-04 -1.114422e-03
#> textual=~x5      7.638350e-04 -8.365361e-04 -3.388541e-04 -2.140826e-04
#> textual=~x6      1.959890e-04  5.557985e-04 -5.325474e-04 -3.287862e-04
#> speed=~x7        2.689462e-04  2.175529e-04 -3.853165e-04 -3.615443e-03
#> speed=~x8        5.219829e-05 -2.639875e-04 -7.193402e-04 -1.847806e-03
#> speed=~x9       -8.031481e-06 -7.478416e-05  3.560314e-04  1.523557e-03
#> x1~~x1           2.112193e-04 -3.029082e-04  2.321858e-04  4.493494e-04
#> x2~~x2          -9.034051e-04  3.681102e-04  6.491090e-04  9.648970e-04
#> x3~~x3          -2.045253e-04  5.733665e-04  4.200302e-04  9.188894e-04
#> x4~~x4           2.369284e-03 -9.343907e-04 -4.792527e-04  3.630016e-04
#> x5~~x5          -9.343907e-04  3.416810e-03  8.941463e-06 -9.070329e-05
#> x6~~x6          -4.792527e-04  8.941463e-06  2.014369e-03  1.419110e-03
#> x7~~x7           3.630016e-04 -9.070329e-05  1.419110e-03  9.489695e-03
#> x8~~x8          -2.016503e-04  3.930651e-04  7.146565e-04  1.681090e-03
#> x9~~x9           3.220988e-04  7.447255e-05 -3.950168e-04 -1.474141e-03
#> visual~~textual -5.569200e-05  3.415000e-04 -1.282415e-04 -7.142344e-04
#> visual~~speed   -2.838457e-04  4.112181e-04  4.945674e-04  1.847850e-03
#> textual~~speed  -5.044799e-04 -4.062312e-04  6.238473e-04 -5.473036e-04
#>                        x8~~x8        x9~~x9 visual~~textual visual~~speed
#> visual=~x1       3.412162e-04 -1.169325e-04    5.838731e-04  0.0002175891
#> visual=~x2      -9.899260e-04 -1.756443e-04    2.527877e-04 -0.0006447806
#> visual=~x3      -4.981966e-04 -7.484516e-05   -1.637993e-04 -0.0001611589
#> textual=~x4     -3.496915e-04  2.374742e-04    1.907565e-04 -0.0007618614
#> textual=~x5     -6.833667e-04 -7.464858e-05    4.136994e-04 -0.0003092461
#> textual=~x6     -8.104702e-05  3.043284e-04    1.876363e-04 -0.0004145462
#> speed=~x7       -4.649879e-04  5.481364e-04    2.160835e-04 -0.0002691227
#> speed=~x8       -3.908064e-03  2.104002e-03   -9.493487e-05 -0.0014019586
#> speed=~x9        2.862057e-03 -4.012708e-03    2.058843e-04  0.0025369811
#> x1~~x1           4.071767e-04 -1.081404e-03   -6.676085e-05  0.0002590018
#> x2~~x2          -1.280589e-03  1.098705e-03    6.765625e-04  0.0004323332
#> x3~~x3           5.484940e-04  6.011874e-04    1.853486e-04  0.0005133066
#> x4~~x4          -2.016503e-04  3.220988e-04   -5.569200e-05 -0.0002838457
#> x5~~x5           3.930651e-04  7.447255e-05    3.415000e-04  0.0004112181
#> x6~~x6           7.146565e-04 -3.950168e-04   -1.282415e-04  0.0004945674
#> x7~~x7           1.681090e-03 -1.474141e-03   -7.142344e-04  0.0018478500
#> x8~~x8           8.144486e-03 -4.022520e-03    2.224217e-04  0.0025829541
#> x9~~x9          -4.022520e-03  6.621775e-03   -1.112289e-04 -0.0025470436
#> visual~~textual  2.224217e-04 -1.112289e-04    3.113457e-03 -0.0001154302
#> visual~~speed    2.582954e-03 -2.547044e-03   -1.154302e-04  0.0049119015
#> textual~~speed   1.643378e-03 -1.492357e-03    9.436573e-04  0.0022381628
#>                 textual~~speed
#> visual=~x1        4.140355e-04
#> visual=~x2       -2.448115e-04
#> visual=~x3       -8.726660e-05
#> textual=~x4       6.013306e-05
#> textual=~x5       7.714795e-05
#> textual=~x6      -5.904730e-04
#> speed=~x7         4.188721e-04
#> speed=~x8        -9.036676e-06
#> speed=~x9         1.588119e-03
#> x1~~x1           -5.526231e-04
#> x2~~x2            2.582848e-04
#> x3~~x3            3.344160e-04
#> x4~~x4           -5.044799e-04
#> x5~~x5           -4.062312e-04
#> x6~~x6            6.238473e-04
#> x7~~x7           -5.473036e-04
#> x8~~x8            1.643378e-03
#> x9~~x9           -1.492357e-03
#> visual~~textual   9.436573e-04
#> visual~~speed     2.238163e-03
#> textual~~speed    5.452039e-03

# Internal parameterisation (Laplace approximation)
vcov(fit, type = "theta")
#>                    visual=~x1    visual=~x2    visual=~x3   textual=~x4
#> visual=~x1       6.898587e-03 -9.165857e-04 -9.867557e-04  3.367678e-04
#> visual=~x2      -9.165857e-04  6.525261e-03  1.566651e-03  1.401828e-04
#> visual=~x3      -9.867557e-04  1.566651e-03  6.021448e-03  1.929055e-04
#> textual=~x4      3.367678e-04  1.401828e-04  1.929055e-04  3.202281e-03
#> textual=~x5      3.121193e-04  1.987118e-04  2.574554e-04  1.677661e-03
#> textual=~x6      2.529953e-04  1.681537e-04  2.191811e-04  1.367829e-03
#> speed=~x7        3.029206e-04  7.609836e-05  4.973399e-05  8.864062e-05
#> speed=~x8        3.412604e-04  1.418067e-04  4.229667e-05  9.296407e-05
#> speed=~x9        6.195091e-05  1.241973e-04  3.052028e-04  7.359271e-05
#> x1~~x1          -1.216302e-02  5.382504e-03  6.370959e-03 -1.273305e-04
#> x2~~x2           1.458995e-03 -2.080412e-03 -9.054328e-04  2.226225e-05
#> x3~~x3           3.057214e-03 -1.602436e-03 -3.925080e-03  3.897942e-05
#> x4~~x4          -2.104044e-04  1.393482e-04  1.437300e-04 -1.879696e-03
#> x5~~x5           9.295670e-05 -6.679914e-05 -6.251812e-05  5.925565e-04
#> x6~~x6           1.148743e-04 -7.247747e-05 -7.985547e-05  6.570178e-04
#> x7~~x7          -1.661245e-04  4.988021e-05  1.430742e-04 -1.892993e-05
#> x8~~x8          -3.313343e-04 -4.319711e-05  3.765540e-04 -8.262366e-06
#> x9~~x9           3.489718e-04 -1.892216e-05 -3.588609e-04  2.009704e-05
#> visual~~textual  9.415274e-04  2.075824e-05 -3.777095e-04  7.678324e-04
#> visual~~speed   -4.800760e-04  3.732852e-04  9.826828e-04  2.245913e-04
#> textual~~speed   4.293306e-04  2.641969e-04  4.007179e-04  4.074328e-04
#>                   textual=~x5   textual=~x6     speed=~x7     speed=~x8
#> visual=~x1       3.121193e-04  2.529953e-04  3.029206e-04  3.412604e-04
#> visual=~x2       1.987118e-04  1.681537e-04  7.609836e-05  1.418067e-04
#> visual=~x3       2.574554e-04  2.191811e-04  4.973399e-05  4.229667e-05
#> textual=~x4      1.677661e-03  1.367829e-03  8.864062e-05  9.296407e-05
#> textual=~x5      3.904073e-03  1.568076e-03  7.389059e-05  9.519861e-05
#> textual=~x6      1.568076e-03  2.879717e-03  7.409026e-05  9.590733e-05
#> speed=~x7        7.389059e-05  7.409026e-05  5.439411e-03  1.886296e-03
#> speed=~x8        9.519861e-05  9.590733e-05  1.886296e-03  5.575902e-03
#> speed=~x9        1.028171e-04  6.498526e-05 -8.811822e-04 -1.676962e-03
#> x1~~x1           6.449054e-05  7.346268e-05 -3.316229e-04 -3.374218e-04
#> x2~~x2          -1.253501e-05 -1.321498e-05  3.141020e-05 -8.613488e-06
#> x3~~x3          -2.257528e-05 -2.723894e-05  1.520326e-04  2.054104e-04
#> x4~~x4           6.427250e-04  6.798769e-04 -5.422684e-05  5.431733e-06
#> x5~~x5          -2.069644e-03  4.778096e-04  6.645149e-05  4.656548e-05
#> x6~~x6           5.014459e-04 -1.595198e-03 -7.304155e-06 -4.557413e-05
#> x7~~x7           1.749729e-05 -5.150650e-06 -3.352527e-03 -1.759687e-03
#> x8~~x8           1.570784e-05 -3.750155e-05 -3.413783e-03 -9.277591e-03
#> x9~~x9          -2.667457e-05  2.589312e-05  3.678129e-03  5.834693e-03
#> visual~~textual  5.947602e-04  7.300952e-04  1.882035e-04  2.270434e-04
#> visual~~speed    2.844870e-04  2.176762e-04 -1.916930e-03 -2.309643e-03
#> textual~~speed   4.639551e-04  4.028189e-04 -3.232058e-04 -8.884590e-04
#>                     speed=~x9        x1~~x1        x2~~x2        x3~~x3
#> visual=~x1       6.195091e-05 -1.216302e-02  1.458995e-03  3.057214e-03
#> visual=~x2       1.241973e-04  5.382504e-03 -2.080412e-03 -1.602436e-03
#> visual=~x3       3.052028e-04  6.370959e-03 -9.054328e-04 -3.925080e-03
#> textual=~x4      7.359271e-05 -1.273305e-04  2.226225e-05  3.897942e-05
#> textual=~x5      1.028171e-04  6.449054e-05 -1.253501e-05 -2.257528e-05
#> textual=~x6      6.498526e-05  7.346268e-05 -1.321498e-05 -2.723894e-05
#> speed=~x7       -8.811822e-04 -3.316229e-04  3.141020e-05  1.520326e-04
#> speed=~x8       -1.676962e-03 -3.374218e-04 -8.613488e-06  2.054104e-04
#> speed=~x9        5.855960e-03  4.488534e-04 -1.154329e-05 -2.479650e-04
#> x1~~x1           4.488534e-04  4.620665e-02 -4.749730e-03 -9.951987e-03
#> x2~~x2          -1.154329e-05 -4.749730e-03  8.482971e-03  1.412339e-03
#> x3~~x3          -2.479650e-04 -9.951987e-03  1.412339e-03  1.277592e-02
#> x4~~x4           2.667909e-05  6.858895e-04 -1.226832e-04 -2.238859e-04
#> x5~~x5          -7.560511e-05 -3.106132e-04  5.767489e-05  9.471622e-05
#> x6~~x6           4.018847e-05 -3.710817e-04  6.431582e-05  1.255854e-04
#> x7~~x7           2.427752e-03  5.073830e-04 -4.897727e-05 -2.349523e-04
#> x8~~x8           7.476746e-03  9.976715e-04  2.597746e-05 -6.159681e-04
#> x9~~x9          -7.583367e-03 -1.066922e-03  2.701198e-05  5.841872e-04
#> visual~~textual  1.174709e-04 -9.425100e-04  3.011801e-04  1.334329e-03
#> visual~~speed    3.673533e-03  3.677837e-03 -1.046462e-05 -7.908254e-04
#> textual~~speed   1.364862e-03  1.318368e-04 -2.688307e-06 -8.861014e-05
#>                        x4~~x4        x5~~x5        x6~~x6        x7~~x7
#> visual=~x1      -2.104044e-04  9.295670e-05  1.148743e-04 -1.661245e-04
#> visual=~x2       1.393482e-04 -6.679914e-05 -7.247747e-05  4.988021e-05
#> visual=~x3       1.437300e-04 -6.251812e-05 -7.985547e-05  1.430742e-04
#> textual=~x4     -1.879696e-03  5.925565e-04  6.570178e-04 -1.892993e-05
#> textual=~x5      6.427250e-04 -2.069644e-03  5.014459e-04  1.749729e-05
#> textual=~x6      6.798769e-04  4.778096e-04 -1.595198e-03 -5.150650e-06
#> speed=~x7       -5.422684e-05  6.645149e-05 -7.304155e-06 -3.352527e-03
#> speed=~x8        5.431733e-06  4.656548e-05 -4.557413e-05 -1.759687e-03
#> speed=~x9        2.667909e-05 -7.560511e-05  4.018847e-05  2.427752e-03
#> x1~~x1           6.858895e-04 -3.106132e-04 -3.710817e-04  5.073830e-04
#> x2~~x2          -1.226832e-04  5.767489e-05  6.431582e-05 -4.897727e-05
#> x3~~x3          -2.238859e-04  9.471622e-05  1.255854e-04 -2.349523e-04
#> x4~~x4           1.667134e-02 -3.171764e-03 -3.495430e-03  8.402710e-05
#> x5~~x5          -3.171764e-03  1.685817e-02 -2.470971e-03 -1.037440e-04
#> x6~~x6          -3.495430e-03 -2.470971e-03  1.486517e-02  1.145233e-05
#> x7~~x7           8.402710e-05 -1.037440e-04  1.145233e-05  1.186519e-02
#> x8~~x8          -1.753871e-05 -1.410673e-04  1.378402e-04  5.289076e-03
#> x9~~x9          -6.330128e-05  1.768555e-04 -9.404482e-05 -5.706960e-03
#> visual~~textual -2.520747e-04  1.023875e-03 -3.228116e-04 -5.013071e-05
#> visual~~speed    7.152447e-05 -9.631759e-05  1.425938e-05  3.689237e-03
#> textual~~speed   1.288246e-04  8.172103e-05 -1.575283e-05  9.222448e-04
#>                        x8~~x8        x9~~x9 visual~~textual visual~~speed
#> visual=~x1      -3.313343e-04  3.489718e-04    9.415274e-04 -4.800760e-04
#> visual=~x2      -4.319711e-05 -1.892216e-05    2.075824e-05  3.732852e-04
#> visual=~x3       3.765540e-04 -3.588609e-04   -3.777095e-04  9.826828e-04
#> textual=~x4     -8.262366e-06  2.009704e-05    7.678324e-04  2.245913e-04
#> textual=~x5      1.570784e-05 -2.667457e-05    5.947602e-04  2.844870e-04
#> textual=~x6     -3.750155e-05  2.589312e-05    7.300952e-04  2.176762e-04
#> speed=~x7       -3.413783e-03  3.678129e-03    1.882035e-04 -1.916930e-03
#> speed=~x8       -9.277591e-03  5.834693e-03    2.270434e-04 -2.309643e-03
#> speed=~x9        7.476746e-03 -7.583367e-03    1.174709e-04  3.673533e-03
#> x1~~x1           9.976715e-04 -1.066922e-03   -9.425100e-04  3.677837e-03
#> x2~~x2           2.597746e-05  2.701198e-05    3.011801e-04 -1.046462e-05
#> x3~~x3          -6.159681e-04  5.841872e-04    1.334329e-03 -7.908254e-04
#> x4~~x4          -1.753871e-05 -6.330128e-05   -2.520747e-04  7.152447e-05
#> x5~~x5          -1.410673e-04  1.768555e-04    1.023875e-03 -9.631759e-05
#> x6~~x6           1.378402e-04 -9.404482e-05   -3.228116e-04  1.425938e-05
#> x7~~x7           5.289076e-03 -5.706960e-03   -5.013071e-05  3.689237e-03
#> x8~~x8           3.461087e-02 -1.756801e-02   -1.298722e-04  8.581362e-03
#> x9~~x9          -1.756801e-02  2.446224e-02    1.174843e-04 -7.473895e-03
#> visual~~textual -1.298722e-04  1.174843e-04    6.402515e-03  7.171685e-04
#> visual~~speed    8.581362e-03 -7.473895e-03    7.171685e-04  1.173705e-02
#> textual~~speed   3.640917e-03 -2.521933e-03    1.488903e-03  3.057236e-03
#>                 textual~~speed
#> visual=~x1        4.293306e-04
#> visual=~x2        2.641969e-04
#> visual=~x3        4.007179e-04
#> textual=~x4       4.074328e-04
#> textual=~x5       4.639551e-04
#> textual=~x6       4.028189e-04
#> speed=~x7        -3.232058e-04
#> speed=~x8        -8.884590e-04
#> speed=~x9         1.364862e-03
#> x1~~x1            1.318368e-04
#> x2~~x2           -2.688307e-06
#> x3~~x3           -8.861014e-05
#> x4~~x4            1.288246e-04
#> x5~~x5            8.172103e-05
#> x6~~x6           -1.575283e-05
#> x7~~x7            9.222448e-04
#> x8~~x8            3.640917e-03
#> x9~~x9           -2.521933e-03
#> visual~~textual   1.488903e-03
#> visual~~speed     3.057236e-03
#> textual~~speed    5.925664e-03
# }